Cardano.Ledger.Address

Documentation

mkRwdAcnt ∷ Network → Credential 'Staking c → RewardAcnt c

Deprecated: In favor of RewardAcnt

serialiseAddr ∷ Addr c → ByteString

Serialise an address to the external format.

deserialiseAddr ∷ Crypto c ⇒ ByteString → Maybe (Addr c)

Deprecated: In favor of decodeAddr or decodeAddrLenient. Please choose the appropriate variant carefully depending on your use case

Deserialise an address from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

data Addr c

An address for UTxO.

Contents of Addr data type are intentionally left as lazy, otherwise operating on compact form of an address will result in redundant work.

Constructors

Addr Network (PaymentCredential c) (StakeReference c)

AddrBootstrap (BootstrapAddress c)

addrPtrNormalize ∷ Addr c → Addr c

This function is implemented solely for the purpose of translating garbage pointers into knowingly invalid ones. Any pointer that contains a SlotNo, TxIx or CertIx that is too large to fit into Word32, Word16 and Word16 respectively, will have all of its values set to 0 using normalizePtr.

There are two reasons why we can safely do that at the Babbage/Conway era boundary:

Invalid pointers are no longer allowed in transactions starting with Babbage era

There are only a handful of Ptrs on mainnet that are invalid.

Once the transition is complete and we are officially in Conway era, this translation logic can be removed in favor of a fixed deserializer that does the same thing for all eras prior to Babbage.

newtype BootstrapAddress c

Constructors

BootstrapAddress

Fields

unBootstrapAddress ∷ Address

bootstrapAddressAttrsSize ∷ BootstrapAddress c → Int

The size of the extra attributes in a bootstrap (ie Byron) address. Used to help enforce that people do not post huge ones on the chain.

isBootstrapRedeemer ∷ BootstrapAddress c → Bool

Return True if a given address is a redeemer address from the Byron Era

getNetwork ∷ Addr c → Network

Lookup a Network Id for an Address

data RewardAccount c

An account based address for rewards

Constructors

RewardAccount

Fields

raNetwork ∷ !Network

raCredential ∷ !(Credential 'Staking c)

Bundled Patterns

pattern RewardAcnt ∷ Network → Credential 'Staking c → RewardAccount c

Deprecated: Use RewardAccount instead

serialiseRewardAccount ∷ RewardAcnt c → ByteString

Serialise a reward account to the external format.

deserialiseRewardAccount ∷ Crypto c ⇒ ByteString → Maybe (RewardAcnt c)

Deserialise a reward account from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

bootstrapKeyHash ∷ ∀ c. Crypto c ⇒ BootstrapAddress c → KeyHash 'Payment c

putAddr ∷ Addr c → Put

putCredential ∷ Credential kr c → Put

putPtr ∷ Ptr → Put

putRewardAccount ∷ RewardAcnt c → Put

putVariableLengthWord64 ∷ Word64 → Put

newtype Word7

Constructors

Word7 Word8

toWord7 ∷ Word8 → Word7

Compact Address

fromBoostrapCompactAddress ∷ CompactAddress → CompactAddr c

Convert Byron's comapct address into CompactAddr. This is just an efficient type cast.

compactAddr ∷ Addr c → CompactAddr c

decompactAddr ∷ ∀ c. (HasCallStack, Crypto c) ⇒ CompactAddr c → Addr c

data CompactAddr c

unCompactAddr ∷ CompactAddr c → ShortByteString

Unwrap the compact address and get to the address' binary representation.

isPayCredScriptCompactAddr ∷ CompactAddr c → Bool

Efficiently check whether compacted adddress is an address with a credential that is a payment script.

isBootstrapCompactAddr ∷ CompactAddr c → Bool

Efficiently check whether compated adddress is a Byron address.

decodeAddr ∷ ∀ c m. (Crypto c, MonadFail m) ⇒ ByteString → m (Addr c)

Strict decoder for an address from a ByteString. This will not let you decode some of the buggy addresses that have been placed on chain. This decoder is intended for addresses that are to be placed on chian today.

decodeAddrEither ∷ ∀ c. Crypto c ⇒ ByteString → Either String (Addr c)

Same as decodeAddr, but produces an Either result

decodeAddrStateT ∷ (Crypto c, MonadFail m, AddressBuffer b) ⇒ b → StateT Int m (Addr c)

Just like decodeAddrStateLenientT, but enforces the address to be well-formed.

decodeAddrStateLenientT

Arguments

∷ (Crypto c, MonadFail m, AddressBuffer b)

⇒ Bool
Enable lenient decoding for Ptrs, i.e. indicate whether junk can follow a Ptr. This is necessary for backwards compatibility only. Setting this argument to True is only needed for backwards compatibility.

→ Bool
Indicate whether decoder should not enforce the full input to be consumed or not, i.e. allow garbage at the end or not. Setting this argument to True is only needed for backwards compatibility.

→ b

→ StateT Int m (Addr c)

This is the most general decoder for a Cardano address. This function is not meant to be used directly, but it is exported for convenice. decodeAddr and other should be used instead.

While decoding an Addr the header (the first byte in the buffer) is expected to be in a certain format. Here are the meaning of all the bits:

@@@

┏━━━━━━━━━━━━━━━━┳━┯━┯━┯━┯━┯━┯━┯━┓ ┃ Byron Address ┃1┊0┊0┊0┊0┊0┊1┊0┃ ┣━━━━━━━━━━━━━━━━╋━┿━┿━┿━┿━┿━┿━┿━┫ ┃Shelley Address ┃0┊x┊x┊x┊0┊0┊0┊x┃ ┗━━━━━━━━━━━━━━━━╋━┿━┿━┿━┿━┿━┿━┿━╋━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓ ┃0┊0┊0┊0┊0┊0┊0┊0┃ Testnet PaymentKey StakingKey ┃ ┃0┊0┊0┊0┊0┊0┊0┊1┃ Mainnet PaymentKey StakingKey ┃ ┃0┊0┊0┊1┊0┊0┊0┊0┃ Testnet PaymentScript StakingKey ┃ ┃0┊0┊0┊1┊0┊0┊0┊1┃ Mainnet PaymentScript StakingKey ┃ ┃0┊0┊1┊0┊0┊0┊0┊0┃ Testnet PaymentKey StakingScript ┃ ┃0┊0┊1┊0┊0┊0┊0┊1┃ Mainnet PaymentKey StakingScript ┃ ┃0┊0┊1┊1┊0┊0┊0┊0┃ Testnet PaymentScript StakingScript ┃ ┃0┊0┊1┊1┊0┊0┊0┊1┃ Mainnet PaymentScript StakingScript ┃ ┃0┊1┊0┊0┊0┊0┊0┊0┃ Testnet PaymentKey StakingPtr ┃ ┃0┊1┊0┊0┊0┊0┊0┊1┃ Mainnet PaymentKey StakingPtr ┃ ┃0┊1┊0┊1┊0┊0┊0┊0┃ Testnet PaymentScript StakingPtr ┃ ┃0┊1┊0┊1┊0┊0┊0┊1┃ Mainnet PaymentScript StakingPtr ┃ ┃0┊1┊1┊0┊0┊0┊0┊0┃ Testnet PaymentKey StakingNull ┃ ┃0┊1┊1┊0┊0┊0┊0┊1┃ Mainnet PaymentKey StakingNull ┃ ┃0┊1┊1┊1┊0┊0┊0┊0┃ Testnet PaymentScript StakingNull ┃ ┃0┊1┊1┊1┊0┊0┊0┊1┃ Mainnet PaymentScript StakingNull ┃ ┗━┷━┷━┷━┷━┷━┷━┷━┻━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛ `Is Mainnet Address `Payment Credential is a Script `Staking Credential is a Script / No Staking Credential `Not a Base Address @@@

fromCborAddr ∷ ∀ c s. Crypto c ⇒ Decoder s (Addr c)

Decoder for an Addr. Works in all eras

fromCborBothAddr ∷ ∀ c s. Crypto c ⇒ Decoder s (Addr c, CompactAddr c)

This is the decoder for an address that returns both the actual Addr and the bytes, that it was encoded as.

fromCborCompactAddr ∷ ∀ c s. Crypto c ⇒ Decoder s (CompactAddr c)

Returns the actual bytes that represent an addres, while ensuring that they can be decoded in any era as an Addr when need be.

fromCborBackwardsBothAddr ∷ ∀ c s. Crypto c ⇒ Decoder s (Addr c, CompactAddr c)

Prior to Babbage era we did not check if a binary blob representing an address was fully consumed, so unfortunately we must preserve this behavior. However, we do not need to preserve the unconsumed bytes in memory, therefore we can to drop the garbage after we successfully decoded the malformed address. We also need to allow bogus pointer address to be deserializeable prior to Babbage era.

decodeRewardAccount ∷ ∀ c b m. (Crypto c, AddressBuffer b, MonadFail m) ⇒ b → m (RewardAcnt c)

fromCborRewardAccount ∷ ∀ c s. Crypto c ⇒ Decoder s (RewardAcnt c)

newtype Withdrawals c

This is called wdrl in the spec.

Constructors

Withdrawals

Fields

unWithdrawals ∷ Map (RewardAcnt c) Coin

Deprecations

type RewardAcnt = RewardAccount

Deprecated: Use RewardAccount instead

serialiseRewardAcnt ∷ RewardAcnt c → ByteString

Deprecated: Use serialiseRewardAccount instead

deserialiseRewardAcnt ∷ Crypto c ⇒ ByteString → Maybe (RewardAcnt c)

Deprecated: Use deserialiseRewardAccount instead

putRewardAcnt ∷ RewardAcnt c → Put

Deprecated: Use putRewardAccount instead

decodeRewardAcnt ∷ ∀ c b m. (Crypto c, AddressBuffer b, MonadFail m) ⇒ b → m (RewardAcnt c)

Deprecated: Use decodeRewardAccount instead

fromCborRewardAcnt ∷ ∀ c s. Crypto c ⇒ Decoder s (RewardAcnt c)

Deprecated: Use fromCborRewardAccount instead

rewardAccountCredentialL ∷ Lens' (RewardAccount c) (Credential 'Staking c)

rewardAccountNetworkL ∷ Lens' (RewardAccount c) Network