Cardano.Ledger.Ap

Description

Deprecated: Because it is no longer used in Ledger

This is taken from Edward Kmett's free library See: https://hackage.haskell.org/package/free-5.2

Documentation

newtype Ap f a

The faster free Applicative.

Constructors

Ap

Fields

unAp ∷ ∀ u y z. (∀ x. (x → y) → ASeq f x → z) → (u → a → y) → ASeq f u → z

hoistAp ∷ (∀ x. f x → g x) → Ap f a → Ap g a

Given a natural transformation from f to g this gives a monoidal natural transformation from Ap f to Ap g.

runAp ∷ Applicative g ⇒ (∀ x. f x → g x) → Ap f a → g a

Given a natural transformation from f to g, this gives a canonical monoidal natural transformation from Ap f to g.

runAp t == retractApp . hoistApp t

runAp_ ∷ Monoid m ⇒ (∀ a. f a → m) → Ap f b → m

Perform a monoidal analysis over free applicative value.

Example:

count :: Ap f a -> Int
count = getSum . runAp_ (\_ -> Sum 1)