Cardano.Ledger.BHeaderView

Documentation

data BHeaderView crypto

BHeaderView provides an interface between block headers from different Cardano protocols and packages that should be agnostic of Cardano protocol specific details, such as those in TPraos, Praos, Genesis, etc.

In particular, the BBODY rule comprises most of the ledger logic and should work independently of the protocol. The values in BHeaderView provide BBODY all the data that it needs from the block headers.

Constructors

BHeaderView

Fields

bhviewID :: KeyHash 'BlockIssuer crypto

The block issuer. In the TPraos protocol, this can be a Genesis delegate, everywhere else it is the stake pool ID.

bhviewBSize :: Natural

The purported size (in bytes) of the block body.

bhviewHSize :: Int

The purported size (in bytes) of the block header.

bhviewBHash :: Hash crypto EraIndependentBlockBody

The purported hash of the block body.

bhviewSlot :: SlotNo

The slot for which this block was submitted to the chain.

isOverlaySlot

Arguments

:: SlotNo
The first slot of the given epoch.

-> UnitInterval
The decentralization parameter.

-> SlotNo
The slot to check.

-> Bool

Determine if the given slot is reserved for the overlay schedule.