Cardano.Ledger.CompactAddress

Documentation

compactAddr :: Addr crypto -> CompactAddr crypto

decompactAddr :: forall crypto. Crypto crypto => CompactAddr crypto -> Addr crypto

newtype CompactAddr crypto

Constructors

UnsafeCompactAddr ShortByteString

substring :: ShortByteString -> Int -> Int -> ShortByteString

isPayCredScriptCompactAddr :: CompactAddr crypto -> Bool

Efficiently check whether compated adddress is an address with a credential that is a payment script.

isBootstrapCompactAddr :: CompactAddr crypto -> Bool

Efficiently check whether compated adddress is a Byron address.

decodeAddr :: forall crypto m. (Crypto crypto, MonadFail m) => ByteString -> m (Addr crypto)

decodeAddrShort :: forall crypto m. (Crypto crypto, MonadFail m) => ShortByteString -> m (Addr crypto)

decodeAddrEither :: forall crypto. Crypto crypto => ByteString -> Either String (Addr crypto)

decodeAddrShortEither :: forall crypto. Crypto crypto => ShortByteString -> Either String (Addr crypto)

fromCborAddr :: forall crypto s. Crypto crypto => Decoder s (Addr crypto)

fromCborBothAddr :: forall crypto s. Crypto crypto => Decoder s (Addr crypto, CompactAddr crypto)

fromCborCompactAddr :: forall crypto s. Crypto crypto => Decoder s (CompactAddr crypto)

fromCborBackwardsBothAddr :: forall crypto s. Crypto crypto => Decoder s (Addr crypto, CompactAddr crypto)

decodeRewardAcnt :: forall crypto b m. (Crypto crypto, AddressBuffer b, MonadFail m) => b -> m (RewardAcnt crypto)

fromCborRewardAcnt :: forall crypto s. Crypto crypto => Decoder s (RewardAcnt crypto)

Exported for benchmarking only

fromCborCompactAddrOld :: forall s crypto. Crypto crypto => Decoder s (CompactAddr crypto)

decompactAddrLazy :: forall crypto. Crypto crypto => CompactAddr crypto -> Addr crypto

This lazy deserializer is kept around purely for benchmarking, so we can verify that new deserializer decodeAddrStateT is doing the work lazily.