Cardano.Ledger.Core

Description

This module defines core type families which we know to vary from era to era.

Families in this module should be indexed on era.

It is intended for qualified import: > import qualified Cardano.Ledger.Core as Core

Era-changing types

class (EraTxBody era, EraTxWits era, EraTxAuxData era, EraPParams era, NoThunks (Tx era), DecCBOR (Annotator (Tx era)), EncCBOR (Tx era), ToCBOR (Tx era), Show (Tx era), Eq (Tx era), EqRaw (Tx era)) ⇒ EraTx era where

A transaction.

Associated Types

type Tx era = (r ∷ Type) | r → era

type TxUpgradeError era ∷ Type

type TxUpgradeError era = Void

Methods

mkBasicTx ∷ TxBody era → Tx era

bodyTxL ∷ Lens' (Tx era) (TxBody era)

witsTxL ∷ Lens' (Tx era) (TxWits era)

auxDataTxL ∷ Lens' (Tx era) (StrictMaybe (AuxiliaryData era))

sizeTxF ∷ SimpleGetter (Tx era) Integer

validateNativeScript ∷ Tx era → NativeScript era → Bool

Using information from the transaction validate the supplied native script.

getMinFeeTx

Arguments

∷ PParams era

→ Tx era

→ Int
Size in bytes of reference scripts present in this transaction

→ Coin

Minimum fee calculation excluding witnesses

upgradeTx ∷ EraTx (PreviousEra era) ⇒ Tx (PreviousEra era) → Either (TxUpgradeError era) (Tx era)

txIdTx ∷ EraTx era ⇒ Tx era → TxId (EraCrypto era)

class (Val (Value era), ToJSON (TxOut era), DecCBOR (Value era), DecCBOR (CompactForm (Value era)), EncCBOR (Value era), ToCBOR (TxOut era), FromCBOR (TxOut era), EncCBOR (TxOut era), DecCBOR (TxOut era), DecShareCBOR (TxOut era), Share (TxOut era) ~ Interns (Credential 'Staking (EraCrypto era)), NoThunks (TxOut era), NFData (TxOut era), Show (TxOut era), Eq (TxOut era), EraPParams era) ⇒ EraTxOut era where

Abstract interface into specific fields of a TxOut

Minimal complete definition

mkBasicTxOut, upgradeTxOut, valueEitherTxOutL, addrEitherTxOutL, (getMinCoinSizedTxOut | getMinCoinTxOut)

Associated Types

type TxOut era = (r ∷ Type) | r → era

The output of a UTxO for a particular era

Methods

mkBasicTxOut ∷ HasCallStack ⇒ Addr (EraCrypto era) → Value era → TxOut era

upgradeTxOut ∷ EraTxOut (PreviousEra era) ⇒ TxOut (PreviousEra era) → TxOut era

Every era, except Shelley, must be able to upgrade a TxOut from a previous era.

valueTxOutL ∷ Lens' (TxOut era) (Value era)

compactValueTxOutL ∷ HasCallStack ⇒ Lens' (TxOut era) (CompactForm (Value era))

valueEitherTxOutL ∷ Lens' (TxOut era) (Either (Value era) (CompactForm (Value era)))

Lens for getting and setting in TxOut either an address or its compact version by doing the least amount of work.

addrTxOutL ∷ Lens' (TxOut era) (Addr (EraCrypto era))

compactAddrTxOutL ∷ Lens' (TxOut era) (CompactAddr (EraCrypto era))

addrEitherTxOutL ∷ Lens' (TxOut era) (Either (Addr (EraCrypto era)) (CompactAddr (EraCrypto era)))

Lens for getting and setting in TxOut either an address or its compact version by doing the least amount of work.

The utility of this function comes from the fact that TxOut usually stores the address in either one of two forms: compacted or unpacked. In order to avoid extroneous conversions in getTxOutAddr and getTxOutCompactAddr we can define just this functionality. Also sometimes it is crucial to know at the callsite which form of address we have readily available without any conversions (eg. searching millions of TxOuts for a particular address)

getMinCoinSizedTxOut ∷ PParams era → Sized (TxOut era) → Coin

Produce the minimum lovelace that a given transaction output must contain. Information about the size of the TxOut is required in some eras. Use getMinCoinTxOut if you don't have the size readily available to you.

getMinCoinTxOut ∷ PParams era → TxOut era → Coin

Same as getMinCoinSizedTxOut, except information about the size of TxOut will be computed by serializing the TxOut. If the size turns out to be not needed, then serialization will have no overhead, since it is computed lazily.

bootAddrTxOutF ∷ EraTxOut era ⇒ SimpleGetter (TxOut era) (Maybe (BootstrapAddress (EraCrypto era)))

coinTxOutL ∷ (HasCallStack, EraTxOut era) ⇒ Lens' (TxOut era) Coin

compactCoinTxOutL ∷ (HasCallStack, EraTxOut era) ⇒ Lens' (TxOut era) (CompactForm Coin)

isAdaOnlyTxOutF ∷ EraTxOut era ⇒ SimpleGetter (TxOut era) Bool

This is a getter that implements an efficient way to check whether TxOut contains ADA only.

class (EraTxOut era, EraTxCert era, EraPParams era, HashAnnotated (TxBody era) EraIndependentTxBody (EraCrypto era), DecCBOR (Annotator (TxBody era)), EncCBOR (TxBody era), ToCBOR (TxBody era), NoThunks (TxBody era), NFData (TxBody era), Show (TxBody era), Eq (TxBody era), EqRaw (TxBody era)) ⇒ EraTxBody era where

Minimal complete definition

mkBasicTxBody, inputsTxBodyL, outputsTxBodyL, feeTxBodyL, withdrawalsTxBodyL, auxDataHashTxBodyL, spendableInputsTxBodyF, allInputsTxBodyF, certsTxBodyL, upgradeTxBody

Associated Types

type TxBody era = (r ∷ Type) | r → era

The body of a transaction.

type TxBodyUpgradeError era ∷ Type

type TxBodyUpgradeError era = Void

Methods

mkBasicTxBody ∷ TxBody era

inputsTxBodyL ∷ Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

outputsTxBodyL ∷ Lens' (TxBody era) (StrictSeq (TxOut era))

feeTxBodyL ∷ Lens' (TxBody era) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody era) (Withdrawals (EraCrypto era))

auxDataHashTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

spendableInputsTxBodyF ∷ SimpleGetter (TxBody era) (Set (TxIn (EraCrypto era)))

This getter will produce all inputs from the UTxO map that this transaction might spend, which ones will depend on the validity of the transaction itself. Starting in Alonzo this will include collateral inputs.

allInputsTxBodyF ∷ SimpleGetter (TxBody era) (Set (TxIn (EraCrypto era)))

This getter will produce all inputs from the UTxO map that this transaction is referencing, even if some of them cannot be spent by the transaction. For example starting with Babbage era it will also include reference inputs.

certsTxBodyL ∷ Lens' (TxBody era) (StrictSeq (TxCert era))

getTotalDepositsTxBody

Arguments

∷ PParams era

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether stake pool is registered or not

→ TxBody era

→ Coin

Compute the total deposits from the certificates in a TxBody.

This is the contribution of a TxBody towards the consumed amount by the transaction

getTotalRefundsTxBody

Arguments

∷ PParams era

→ (Credential 'Staking (EraCrypto era) → Maybe Coin)
Lookup current deposit for Staking credential if one is registered

→ (Credential 'DRepRole (EraCrypto era) → Maybe Coin)
Lookup current deposit for DRep credential if one is registered

→ TxBody era

→ Coin

Compute the total refunds from the Certs of a TxBody.

This is the contribution of a TxBody towards produced amount by the transaction

getGenesisKeyHashCountTxBody ∷ TxBody era → Int

This function is not used in the ledger rules. It is only used by the downstream tooling to figure out how many witnesses should be supplied for Genesis keys.

upgradeTxBody ∷ EraTxBody (PreviousEra era) ⇒ TxBody (PreviousEra era) → Either (TxBodyUpgradeError era) (TxBody era)

Upgrade the transaction body from the previous era.

This can fail where elements of the transaction body are deprecated. Compare this to translateEraThroughCBOR: - upgradeTxBody will use the Haskell representation, but will not preserve the serialised form. However, it will be suitable for iterated translation through eras. - translateEraThroughCBOR will preserve the binary representation, but is not guaranteed to work through multiple eras - that is, the serialised representation from era n is guaranteed valid in era n + 1, but not necessarily in era n + 2.

txIdTxBody ∷ EraTxBody era ⇒ TxBody era → TxId (EraCrypto era)

class (Era era, Eq (TxAuxData era), EqRaw (TxAuxData era), Show (TxAuxData era), NoThunks (TxAuxData era), ToCBOR (TxAuxData era), EncCBOR (TxAuxData era), DecCBOR (Annotator (TxAuxData era)), HashAnnotated (TxAuxData era) EraIndependentTxAuxData (EraCrypto era)) ⇒ EraTxAuxData era where

TxAuxData which may be attached to a transaction

Associated Types

type TxAuxData era = (r ∷ Type) | r → era

Methods

upgradeTxAuxData ∷ EraTxAuxData (PreviousEra era) ⇒ TxAuxData (PreviousEra era) → TxAuxData era

Every era, except Shelley, must be able to upgrade a TxAuxData from a previous era.

Warning - Important to note that any memoized binary representation will not be preserved. If you need to retain underlying bytes you can use translateEraThroughCBOR

hashTxAuxData ∷ TxAuxData era → AuxiliaryDataHash (EraCrypto era)

validateTxAuxData ∷ ProtVer → TxAuxData era → Bool

class (EraScript era, Eq (TxWits era), EqRaw (TxWits era), Show (TxWits era), Monoid (TxWits era), NoThunks (TxWits era), ToCBOR (TxWits era), EncCBOR (TxWits era), DecCBOR (Annotator (TxWits era))) ⇒ EraTxWits era where

A collection of witnesses in a Tx

Minimal complete definition

addrTxWitsL, bootAddrTxWitsL, scriptTxWitsL, upgradeTxWits

Associated Types

type TxWits era = (r ∷ Type) | r → era

Methods

mkBasicTxWits ∷ TxWits era

addrTxWitsL ∷ Lens' (TxWits era) (Set (WitVKey 'Witness (EraCrypto era)))

bootAddrTxWitsL ∷ Lens' (TxWits era) (Set (BootstrapWitness (EraCrypto era)))

scriptTxWitsL ∷ Lens' (TxWits era) (Map (ScriptHash (EraCrypto era)) (Script era))

upgradeTxWits ∷ EraTxWits (PreviousEra era) ⇒ TxWits (PreviousEra era) → TxWits era

class (Era era, Show (Script era), Eq (Script era), EqRaw (Script era), ToCBOR (Script era), EncCBOR (Script era), DecCBOR (Annotator (Script era)), NoThunks (Script era), SafeToHash (Script era), Eq (NativeScript era), Show (NativeScript era), NFData (NativeScript era), NoThunks (NativeScript era), EncCBOR (NativeScript era), DecCBOR (Annotator (NativeScript era))) ⇒ EraScript era where

Typeclass for script data types. Allows for script validation and hashing. You must understand the role of SafeToHash and scriptPrefixTag to make new instances. scriptPrefixTag is a magic number representing the tag of the script language. For each new script language defined, a new tag is chosen and the tag is included in the script hash for a script. The safeToHash constraint ensures that Scripts are never reserialised.

Associated Types

type Script era = (r ∷ Type) | r → era

Scripts which may lock transaction outputs in this era

type NativeScript era = (r ∷ Type) | r → era

Methods

upgradeScript ∷ EraScript (PreviousEra era) ⇒ Script (PreviousEra era) → Script era

Every era, except Shelley, must be able to upgrade a Script from a previous era.

Warning - Important to note that any memoized binary representation will not be preserved, you need to retain underlying bytes you can use translateEraThroughCBOR

scriptPrefixTag ∷ Script era → ByteString

getNativeScript ∷ Script era → Maybe (NativeScript era)

fromNativeScript ∷ NativeScript era → Script era

hashScript ∷ ∀ era. EraScript era ⇒ Script era → ScriptHash (EraCrypto era)

Compute ScriptHash of a Script for a particular era.

isNativeScript ∷ EraScript era ⇒ Script era → Bool

hashScriptTxWitsL ∷ EraTxWits era ⇒ Lens (TxWits era) (TxWits era) (Map (ScriptHash (EraCrypto era)) (Script era)) [Script era]

This is a helper lens that will hash the scripts when adding as witnesses.

type family Value era ∷ Type

A value is something which quantifies a transaction output.

class (Era era, Eq (PParamsHKD Identity era), Ord (PParamsHKD Identity era), Show (PParamsHKD Identity era), NFData (PParamsHKD Identity era), EncCBOR (PParamsHKD Identity era), DecCBOR (PParamsHKD Identity era), ToCBOR (PParamsHKD Identity era), FromCBOR (PParamsHKD Identity era), NoThunks (PParamsHKD Identity era), ToJSON (PParamsHKD Identity era), FromJSON (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Ord (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), NFData (PParamsHKD StrictMaybe era), EncCBOR (PParamsHKD StrictMaybe era), DecCBOR (PParamsHKD StrictMaybe era), ToCBOR (PParamsHKD StrictMaybe era), FromCBOR (PParamsHKD StrictMaybe era), NoThunks (PParamsHKD StrictMaybe era), ToJSON (PParamsHKD StrictMaybe era)) ⇒ EraPParams era where

Minimal complete definition

emptyPParamsIdentity, emptyPParamsStrictMaybe, upgradePParamsHKD, downgradePParamsHKD, hkdMinFeeAL, hkdMinFeeBL, hkdMaxBBSizeL, hkdMaxTxSizeL, hkdMaxBHSizeL, hkdKeyDepositL, hkdPoolDepositL, hkdEMaxL, hkdNOptL, hkdA0L, hkdRhoL, hkdTauL, hkdDL, hkdExtraEntropyL, hkdProtocolVersionL, hkdMinUTxOValueL, hkdMinPoolCostL

Associated Types

type PParamsHKD (f ∷ Type → Type) era = (r ∷ Type) | r → era

Protocol parameters where the fields are represented with a HKD

type UpgradePParams (f ∷ Type → Type) era ∷ Type

type DowngradePParams (f ∷ Type → Type) era ∷ Type

Methods

applyPPUpdates ∷ PParams era → PParamsUpdate era → PParams era

Applies a protocol parameters update

default applyPPUpdates ∷ ∀ a u. (Generic (PParamsHKD Identity era), Generic (PParamsHKD StrictMaybe era), Updatable (Rep (PParamsHKD Identity era) a) (Rep (PParamsHKD StrictMaybe era) u)) ⇒ PParams era → PParamsUpdate era → PParams era

emptyPParamsIdentity ∷ PParamsHKD Identity era

emptyPParamsStrictMaybe ∷ PParamsHKD StrictMaybe era

upgradePParamsHKD ∷ (HKDFunctor f, EraPParams (PreviousEra era)) ⇒ UpgradePParams f era → PParamsHKD f (PreviousEra era) → PParamsHKD f era

Upgrade PParams from previous era to the current one

downgradePParamsHKD ∷ (HKDFunctor f, EraPParams (PreviousEra era)) ⇒ DowngradePParams f era → PParamsHKD f era → PParamsHKD f (PreviousEra era)

Downgrade PParams from the current era to the previous one

hkdMinFeeAL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The linear factor for the minimum fee calculation

hkdMinFeeBL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The constant factor for the minimum fee calculation

hkdMaxBBSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Word32)

Maximal block body size

hkdMaxTxSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Word32)

Maximal transaction size

hkdMaxBHSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Word16)

Maximal block header size

hkdKeyDepositL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a key registration deposit

hkdPoolDepositL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a pool registration deposit

hkdEMaxL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f EpochInterval)

epoch bound on pool retirement

hkdNOptL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Natural)

Desired number of pools

hkdA0L ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f NonNegativeInterval)

Pool influence

hkdRhoL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f UnitInterval)

Monetary expansion

hkdTauL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f UnitInterval)

Treasury expansion

hkdDL ∷ (HKDFunctor f, ProtVerAtMost era 6) ⇒ Lens' (PParamsHKD f era) (HKD f UnitInterval)

Decentralization parameter

ppDG ∷ SimpleGetter (PParams era) UnitInterval

Decentralization parameter getter

default ppDG ∷ ProtVerAtMost era 6 ⇒ SimpleGetter (PParams era) UnitInterval

hkdExtraEntropyL ∷ (HKDFunctor f, ProtVerAtMost era 6) ⇒ Lens' (PParamsHKD f era) (HKD f Nonce)

Extra entropy

hkdProtocolVersionL ∷ (HKDFunctor f, ProtVerAtMost era 8) ⇒ Lens' (PParamsHKD f era) (HKD f ProtVer)

Protocol version

ppProtocolVersionL ∷ Lens' (PParams era) ProtVer

default ppProtocolVersionL ∷ ProtVerAtMost era 8 ⇒ Lens' (PParams era) ProtVer

ppuProtocolVersionL ∷ ProtVerAtMost era 8 ⇒ Lens' (PParamsUpdate era) (StrictMaybe ProtVer)

PParamsUpdate Protocol version

hkdMinUTxOValueL ∷ HKDFunctor f ⇒ ProtVerAtMost era 4 ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

Minimum UTxO value

hkdMinPoolCostL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

Minimum Stake Pool Cost

Era

module Cardano.Ledger.Core.Era

Segregated Witness

The idea of segregated witnessing is to alter the encoding of transactions in a block such that the witnesses (the information needed to verify the validity of the transactions) can be stored separately from the body (the information needed to update the ledger state). In this way, a node which only cares about replaying transactions need not even decode the witness information.

In order to do this, we introduce two concepts: - A TxSeq, which represents the decoded structure of a sequence of transactions as represented in the encoded block; that is, with witnessing, metadata and other non-body parts split separately.

class (EraTx era, Eq (TxSeq era), Show (TxSeq era), EncCBORGroup (TxSeq era), DecCBOR (Annotator (TxSeq era))) ⇒ EraSegWits era where

Indicates that an era supports segregated witnessing.

This class embodies an isomorphism between 'TxSeq era' and 'StrictSeq (Tx era)', witnessed by fromTxSeq and toTxSeq.

Associated Types

type TxSeq era = (r ∷ Type) | r → era

Methods

fromTxSeq ∷ TxSeq era → StrictSeq (Tx era)

toTxSeq ∷ StrictSeq (Tx era) → TxSeq era

hashTxSeq ∷ TxSeq era → Hash (HASH (EraCrypto era)) EraIndependentBlockBody

Get the block body hash from the TxSeq. Note that this is not a regular "hash the stored bytes" function since the block body hash forms a small Merkle tree.

numSegComponents ∷ Word64

The number of segregated components

bBodySize ∷ ∀ era. EraSegWits era ⇒ ProtVer → TxSeq era → Int

Rewards

data RewardType

The staking rewards in Cardano are all either:

member rewards - rewards given to a registered stake credential which has delegated to a stake pool, or

leader rewards - rewards given to a registered stake pool (in particular, given to the stake credential in the stake pool registration certificate).

See Figure 47, "Functions used in the Reward Splitting", of the formal specification for more details.

Constructors

MemberReward

LeaderReward

data Reward c

The Reward type captures:

if the reward is a member or leader reward

the stake pool ID associated with the reward

the number of Lovelace in the reward

Constructors

Reward

Fields

rewardType ∷ !RewardType

rewardPool ∷ !(KeyHash 'StakePool c)

rewardAmount ∷ !Coin

Re-exports

module Cardano.Ledger.Hashes

module Cardano.Ledger.Core.TxCert

module Cardano.Ledger.Core.PParams

module Cardano.Ledger.Core.Translation

Deprecations

hashAuxiliaryData ∷ EraTxAuxData era ⇒ TxAuxData era → AuxiliaryDataHash (EraCrypto era)

Deprecated: Use hashTxAuxData instead

validateAuxiliaryData ∷ EraTxAuxData era ⇒ ProtVer → TxAuxData era → Bool

Deprecated: Use validateTxAuxData instead