Cardano.Ledger.Core.Era

Era

class (Crypto (EraCrypto era), Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT, ProtVerLow era <= MaxVersion, ProtVerHigh era <= MaxVersion) ⇒ Era era where

Associated Types

type EraCrypto era ∷ Type

type PreviousEra era = (r ∷ Type) | r → era

Map an era to its predecessor.

For example:

type instance PreviousEra (AllegraEra c) = ShelleyEra c

type ProtVerLow era ∷ Nat

Lowest major protocol version for this era

type ProtVerHigh era ∷ Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

Methods

eraName ∷ String

Textual name of the current era.

Designed to be used with TypeApplications:

>>> eraName @(ByronEra StandardCrypto)
Byron

data ByronEra c

This is the era that preceded Shelley era. It cannot have any other class instances, except for Era type class.

Rules

type family EraRule (rule ∷ Symbol) era = (r ∷ Type) | r → rule

Era STS map

type family EraRuleFailure (rule ∷ Symbol) era = (r ∷ Type) | r → rule era

EraRuleFailure type family is needed for injectivity, which STS' PredicateFailure does not provide for us unfortunately.

type family EraRuleEvent (rule ∷ Symbol) era = (r ∷ Type) | r → rule era

data VoidEraRule (rule ∷ Symbol) era

This is a type with no inhabitans for the rules. It is used to indicate that a rule does not have a predicate failure as well as marking rules that have been disabled when comparing to prior eras.

absurdEraRule ∷ VoidEraRule rule era → a

class EraRuleFailure rule era ~ PredicateFailure (EraRule rule era) ⇒ InjectRuleFailure (rule ∷ Symbol) t era where

Minimal complete definition

Nothing

Methods

injectFailure ∷ t era → EraRuleFailure rule era

default injectFailure ∷ t era ~ EraRuleFailure rule era ⇒ t era → EraRuleFailure rule era

class EraRuleEvent rule era ~ Event (EraRule rule era) ⇒ InjectRuleEvent (rule ∷ Symbol) t era where

Minimal complete definition

Nothing

Methods

injectEvent ∷ t era → EraRuleEvent rule era

default injectEvent ∷ t era ~ EraRuleEvent rule era ⇒ t era → EraRuleEvent rule era

Protocol Version

type AtMostEra (eraName ∷ Type → Type) era = ProtVerAtMost era (ProtVerHigh (eraName (EraCrypto era)))

Restrict the era to equal to eraName or come before it.

type AtLeastEra (eraName ∷ Type → Type) era = ProtVerAtLeast era (ProtVerLow (eraName (EraCrypto era)))

Restrict the era to equal to eraName or come after it

type ExactEra (inEra ∷ Type → Type) era = ProtVerInBounds era (ProtVerLow (inEra (EraCrypto era))) (ProtVerHigh (inEra (EraCrypto era)))

Restrict an era to the specific era through the protocol version. This is equivalent to (inEra (Crypto era) ~ era)

type family ProtVerAtMost era (h ∷ Nat) ∷ Constraint where ...

Requirement for the era's lowest protocol version to be lower or equal to the supplied value

Equations

ProtVerAtMost era h = ProtVerIsInBounds "at most" era h (ProtVerLow era <=? h)

type family ProtVerAtLeast era (l ∷ Nat) ∷ Constraint where ...

Requirement for the era's highest protocol version to be higher or equal to the supplied value

Equations

ProtVerAtLeast era l = ProtVerIsInBounds "at least" era l (l <=? ProtVerHigh era)

type ProtVerInBounds era l h = (ProtVerAtLeast era l, ProtVerAtMost era h)

Restrict a lower and upper bounds of the protocol version for the particular era

atLeastEra ∷ AtLeastEra eraName era ⇒ ()

Enforce era to be at least the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

>>> atLeastEra @BabbageEra @(ConwayEra StandardCrypto)
>>> atLeastEra @BabbageEra @(BabbageEra StandardCrypto)

However this will result in a type error

>>> atLeastEra @BabbageEra @(AlonzoEra StandardCrypto)

atMostEra ∷ AtMostEra eraName era ⇒ ()

Enforce era to be at most the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

>>> atMostEra @BabbageEra @(ShelleyEra StandardCrypto)
>>> atMostEra @AlonzoEra @(MaryEra StandardCrypto)

However this will result in a type error

>>> atMostEra @BabbageEra @(ConwayEra StandardCrypto)

notSupportedInThisEra ∷ HasCallStack ⇒ a

notSupportedInThisEraL ∷ HasCallStack ⇒ Lens' a b

eraProtVerLow ∷ ∀ era. Era era ⇒ Version

Get the value level Version of the lowest major protocol version for the supplied era.

eraProtVerHigh ∷ ∀ era. Era era ⇒ Version

Get the value level Version of the highest major protocol version for the supplied era.

toEraCBOR ∷ ∀ era t. (Era era, EncCBOR t) ⇒ t → Encoding

Convert a type that implements EncCBOR to plain Encoding using the lowest protocol version for the supplied era

fromEraCBOR ∷ ∀ era t s. (Era era, DecCBOR t) ⇒ Decoder s t

Convert a type that implements DecCBOR to plain Decoder using the lowest protocol version for the supplied era

fromEraShareCBOR ∷ ∀ era t s. (Era era, DecShareCBOR t) ⇒ Decoder s t

Convert a type that implements DecShareCBOR to plain Decoder using the lowest protocol version for the supplied era

eraDecoder ∷ ∀ era t s. Era era ⇒ Decoder s t → Decoder s t

Convert a versioned Decoder to plain a Decoder using the lowest protocol version for the supplied era