Cardano.Ledger.Core.PParams

Documentation

class (Era era, Eq (PParamsHKD Identity era), Ord (PParamsHKD Identity era), Show (PParamsHKD Identity era), NFData (PParamsHKD Identity era), EncCBOR (PParamsHKD Identity era), DecCBOR (PParamsHKD Identity era), ToCBOR (PParamsHKD Identity era), FromCBOR (PParamsHKD Identity era), NoThunks (PParamsHKD Identity era), ToJSON (PParamsHKD Identity era), FromJSON (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Ord (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), NFData (PParamsHKD StrictMaybe era), EncCBOR (PParamsHKD StrictMaybe era), DecCBOR (PParamsHKD StrictMaybe era), ToCBOR (PParamsHKD StrictMaybe era), FromCBOR (PParamsHKD StrictMaybe era), NoThunks (PParamsHKD StrictMaybe era), ToJSON (PParamsHKD StrictMaybe era)) ⇒ EraPParams era where

Minimal complete definition

emptyPParamsIdentity, emptyPParamsStrictMaybe, upgradePParamsHKD, downgradePParamsHKD, hkdMinFeeAL, hkdMinFeeBL, hkdMaxBBSizeL, hkdMaxTxSizeL, hkdMaxBHSizeL, hkdKeyDepositL, hkdPoolDepositL, hkdEMaxL, hkdNOptL, hkdA0L, hkdRhoL, hkdTauL, hkdDL, hkdExtraEntropyL, hkdProtocolVersionL, hkdMinUTxOValueL, hkdMinPoolCostL

Associated Types

type PParamsHKD (f ∷ Type → Type) era = (r ∷ Type) | r → era

Protocol parameters where the fields are represented with a HKD

type UpgradePParams (f ∷ Type → Type) era ∷ Type

type DowngradePParams (f ∷ Type → Type) era ∷ Type

Methods

applyPPUpdates ∷ PParams era → PParamsUpdate era → PParams era

Applies a protocol parameters update

default applyPPUpdates ∷ ∀ a u. (Generic (PParamsHKD Identity era), Generic (PParamsHKD StrictMaybe era), Updatable (Rep (PParamsHKD Identity era) a) (Rep (PParamsHKD StrictMaybe era) u)) ⇒ PParams era → PParamsUpdate era → PParams era

emptyPParamsIdentity ∷ PParamsHKD Identity era

emptyPParamsStrictMaybe ∷ PParamsHKD StrictMaybe era

upgradePParamsHKD ∷ (HKDApplicative f, EraPParams (PreviousEra era)) ⇒ UpgradePParams f era → PParamsHKD f (PreviousEra era) → PParamsHKD f era

Upgrade PParams from previous era to the current one

downgradePParamsHKD ∷ (HKDFunctor f, EraPParams (PreviousEra era)) ⇒ DowngradePParams f era → PParamsHKD f era → PParamsHKD f (PreviousEra era)

Downgrade PParams from the current era to the previous one

hkdMinFeeAL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The linear factor for the minimum fee calculation

hkdMinFeeBL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The constant factor for the minimum fee calculation

hkdMaxBBSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Word32)

Maximal block body size

hkdMaxTxSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Word32)

Maximal transaction size

hkdMaxBHSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Word16)

Maximal block header size

hkdKeyDepositL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a key registration deposit

hkdPoolDepositL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a pool registration deposit

hkdEMaxL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f EpochInterval)

epoch bound on pool retirement

hkdNOptL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Natural)

Desired number of pools

hkdA0L ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f NonNegativeInterval)

Pool influence

hkdRhoL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f UnitInterval)

Monetary expansion

hkdTauL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f UnitInterval)

Treasury expansion

hkdDL ∷ (HKDFunctor f, ProtVerAtMost era 6) ⇒ Lens' (PParamsHKD f era) (HKD f UnitInterval)

Decentralization parameter

ppDG ∷ SimpleGetter (PParams era) UnitInterval

Decentralization parameter getter

default ppDG ∷ ProtVerAtMost era 6 ⇒ SimpleGetter (PParams era) UnitInterval

hkdExtraEntropyL ∷ (HKDFunctor f, ProtVerAtMost era 6) ⇒ Lens' (PParamsHKD f era) (HKD f Nonce)

Extra entropy

hkdProtocolVersionL ∷ (HKDFunctor f, ProtVerAtMost era 8) ⇒ Lens' (PParamsHKD f era) (HKD f ProtVer)

Protocol version

ppProtocolVersionL ∷ Lens' (PParams era) ProtVer

default ppProtocolVersionL ∷ ProtVerAtMost era 8 ⇒ Lens' (PParams era) ProtVer

ppuProtocolVersionL ∷ ProtVerAtMost era 8 ⇒ Lens' (PParamsUpdate era) (StrictMaybe ProtVer)

PParamsUpdate Protocol version

hkdMinUTxOValueL ∷ HKDFunctor f ⇒ ProtVerAtMost era 4 ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

Minimum UTxO value

hkdMinPoolCostL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

Minimum Stake Pool Cost

newtype PParams era

Protocol parameters

Constructors

PParams (PParamsHKD Identity era)

emptyPParams ∷ EraPParams era ⇒ PParams era

newtype PParamsUpdate era

The type of updates to Protocol parameters

Constructors

PParamsUpdate (PParamsHKD StrictMaybe era)

emptyPParamsUpdate ∷ EraPParams era ⇒ PParamsUpdate era

genericApplyPPUpdates ∷ ∀ era a u. (Generic (PParamsHKD Identity era), Generic (PParamsHKD StrictMaybe era), Updatable (Rep (PParamsHKD Identity era) a) (Rep (PParamsHKD StrictMaybe era) u)) ⇒ PParams era → PParamsUpdate era → PParams era

PParams lens

ppMinFeeAL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) Coin

The linear factor for the minimum fee calculation

ppMinFeeBL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) Coin

The constant factor for the minimum fee calculation

ppMaxBBSizeL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) Word32

Maximal block body size

ppMaxTxSizeL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) Word32

Maximal transaction size

ppMaxBHSizeL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) Word16

Maximal block header size

ppKeyDepositL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) Coin

The amount of a key registration deposit

ppPoolDepositL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) Coin

The amount of a pool registration deposit

ppEMaxL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) EpochInterval

epoch bound on pool retirement

ppNOptL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) Natural

Desired number of pools

ppA0L ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) NonNegativeInterval

Pool influence

ppRhoL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) UnitInterval

Monetary expansion

ppTauL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) UnitInterval

Treasury expansion

ppDL ∷ ∀ era. (EraPParams era, ProtVerAtMost era 6) ⇒ Lens' (PParams era) UnitInterval

Decentralization parameter

ppExtraEntropyL ∷ ∀ era. (EraPParams era, ProtVerAtMost era 6) ⇒ Lens' (PParams era) Nonce

Extra entropy

ppMinUTxOValueL ∷ ∀ era. (EraPParams era, ProtVerAtMost era 4) ⇒ Lens' (PParams era) Coin

Minimum UTxO value

ppMinPoolCostL ∷ ∀ era. EraPParams era ⇒ Lens' (PParams era) Coin

Minimum Stake Pool Cost

PParamsUpdate lens

ppuMinFeeAL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The linear factor for the minimum fee calculation

ppuMinFeeBL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The constant factor for the minimum fee calculation

ppuMaxBBSizeL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word32)

Maximal block body size

ppuMaxTxSizeL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word32)

Maximal transaction size

ppuMaxBHSizeL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word16)

Maximal block header size

ppuKeyDepositL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The amount of a key registration deposit

ppuPoolDepositL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The amount of a pool registration deposit

ppuEMaxL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

epoch bound on pool retirement

ppuNOptL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

Desired number of pools

ppuA0L ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe NonNegativeInterval)

Pool influence

ppuRhoL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Monetary expansion

ppuTauL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Treasury expansion

ppuDL ∷ ∀ era. (EraPParams era, ProtVerAtMost era 6) ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Decentralization parameter

ppuExtraEntropyL ∷ ∀ era. (EraPParams era, ProtVerAtMost era 6) ⇒ Lens' (PParamsUpdate era) (StrictMaybe Nonce)

Extra entropy

ppuMinUTxOValueL ∷ ∀ era. (EraPParams era, ProtVerAtMost era 4) ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

Minimum UTxO value

ppuMinPoolCostL ∷ ∀ era. EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

Minimum Stake Pool Cost

Utility

ppLens ∷ Lens' (PParams era) (PParamsHKD Identity era)

ppuLens ∷ Lens' (PParamsUpdate era) (PParamsHKD StrictMaybe era)

mapPParams ∷ (PParamsHKD Identity era1 → PParamsHKD Identity era2) → PParams era1 → PParams era2

mapPParamsUpdate ∷ (PParamsHKD StrictMaybe era1 → PParamsHKD StrictMaybe era2) → PParamsUpdate era1 → PParamsUpdate era2

upgradePParams ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ UpgradePParams Identity era → PParams (PreviousEra era) → PParams era

downgradePParams ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ DowngradePParams Identity era → PParams era → PParams (PreviousEra era)

upgradePParamsUpdate ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ UpgradePParams StrictMaybe era → PParamsUpdate (PreviousEra era) → PParamsUpdate era

downgradePParamsUpdate ∷ (EraPParams era, EraPParams (PreviousEra era)) ⇒ DowngradePParams StrictMaybe era → PParamsUpdate era → PParamsUpdate (PreviousEra era)

PParamsUpdate to Data

data PParam era where

Pair the tag, and exisitenially hide the type of the lens for the field with that Lens'

Constructors

PParam ∷ ToPlutusData t ⇒ Word → Lens' (PParamsUpdate era) (StrictMaybe t) → PParam era

makePParamMap ∷ [PParam era] → Map Word (PParam era)

Turn a list into a Map, this assures we have no duplicates.