Cardano.Ledger.Core.TxCert

Documentation

class (Era era, ToJSON (TxCert era), DecCBOR (TxCert era), EncCBOR (TxCert era), ToCBOR (TxCert era), FromCBOR (TxCert era), NoThunks (TxCert era), NFData (TxCert era), Show (TxCert era), Eq (TxCert era)) ⇒ EraTxCert era where

Associated Types

type TxCert era = (r ∷ Type) | r → era

type TxCertUpgradeError era ∷ Type

type TxCertUpgradeError era = Void

Methods

upgradeTxCert ∷ EraTxCert (PreviousEra era) ⇒ TxCert (PreviousEra era) → Either (TxCertUpgradeError era) (TxCert era)

Every era, except Shelley, must be able to upgrade a TxCert from a previous era. However, not all certificates can be upgraded, because some eras lose some of the certificates, thus return type is an Either. Eg. from Babbage to Conway: MIR and Genesis certificates were removed.

getVKeyWitnessTxCert ∷ TxCert era → Maybe (KeyHash 'Witness (EraCrypto era))

Return a witness key whenever a certificate requires one

getScriptWitnessTxCert ∷ TxCert era → Maybe (ScriptHash (EraCrypto era))

Return a ScriptHash for certificate types that require a witness

mkRegPoolTxCert ∷ PoolParams (EraCrypto era) → TxCert era

getRegPoolTxCert ∷ TxCert era → Maybe (PoolParams (EraCrypto era))

mkRetirePoolTxCert ∷ KeyHash 'StakePool (EraCrypto era) → EpochNo → TxCert era

getRetirePoolTxCert ∷ TxCert era → Maybe (KeyHash 'StakePool (EraCrypto era), EpochNo)

lookupRegStakeTxCert ∷ TxCert era → Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can register such credential

lookupUnRegStakeTxCert ∷ TxCert era → Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can unregister such credential

getTotalDepositsTxCerts

Arguments

∷ Foldable f

⇒ PParams era

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether stake pool is registered or not

→ f (TxCert era)

→ Coin

Compute the total deposits from a list of certificates.

getTotalRefundsTxCerts

Arguments

∷ Foldable f

⇒ PParams era

→ (Credential 'Staking (EraCrypto era) → Maybe Coin)
Lookup current deposit for Staking credential if one is registered

→ (Credential 'DRepRole (EraCrypto era) → Maybe Coin)
Lookup current deposit for DRep credential if one is registered

→ f (TxCert era)

→ Coin

Compute the total refunds from a list of certificates.

pattern RegPoolTxCert ∷ EraTxCert era ⇒ PoolParams (EraCrypto era) → TxCert era

pattern RetirePoolTxCert ∷ EraTxCert era ⇒ KeyHash 'StakePool (EraCrypto era) → EpochNo → TxCert era

data Delegation c

Deprecated: No longer used

The delegation of one stake key to another.

Constructors

Delegation

Deprecated: No longer used

Fields

dDelegator ∷ !(StakeCredential c)

dDelegatee ∷ !(KeyHash 'StakePool c)

data PoolCert c

Constructors

RegPool !(PoolParams c)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool c) !EpochNo
A stake pool retirement certificate.

getPoolCertTxCert ∷ EraTxCert era ⇒ TxCert era → Maybe (PoolCert (EraCrypto era))

poolCWitness ∷ PoolCert c → Credential 'StakePool c

Deprecated: As useless. PoolIds have nothing to do with credentials

poolCertKeyHashWitness ∷ PoolCert c → KeyHash 'Witness c

isRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Bool

Check if supplied TxCert is a stake registering certificate

isUnRegStakeTxCert ∷ EraTxCert era ⇒ TxCert era → Bool

Check if supplied TxCert is a stake un-registering certificate