Cardano.Ledger.Credential

Documentation

data Credential (kr ∷ KeyRole)

Script hash or key hash for a payment or a staking object.

Note that credentials (unlike raw key hashes) do appear to vary from era to era, since they reference the hash of a script, which can change. This parameter is a phantom, however, so in actuality the instances will remain the same.

Constructors

ScriptHashObj !ScriptHash

KeyHashObj !(KeyHash kr)

type PaymentCredential = Credential 'Payment

credKeyHash ∷ Credential r → Maybe (KeyHash r)

credKeyHashWitness ∷ Credential r → Maybe (KeyHash 'Witness)

Convert a KeyHash into a Witness KeyHash. Does nothing for Script credentials.

credScriptHash ∷ Credential kr → Maybe ScriptHash

Extract ScriptHash from a Credential. Returns Nothing for KeyHashes

credToText ∷ Credential kr → Text

parseCredential ∷ MonadFail m ⇒ Text → m (Credential kr)

data Ptr

Pointer to a slot number, transaction index and an index in certificate list.

Constructors

Ptr !SlotNo32 !TxIx !CertIx

mkPtrNormalized ∷ Word64 → Word64 → Word64 → Ptr

Construct a valid Ptr, while protecting against overflow. Constructs a Ptr with all zeros for its fields, whenever either one of them doesn't fit in without overflowing. Any pointer that contains a SlotNo, TxIx or CertIx that is too large to fit into Word32, Word16 and Word16 respectively is considered to be an invalid Ptr and result in all values to be clamped to zero. In case of all valid arguments the Ptrs will be constructed with values unmodified.

Note - This functionality is in no way related to dangling pointers, with an exception that any invalid Ptr is guarateed to be a dangling Ptr.

ptrSlotNo ∷ Ptr → SlotNo

ptrTxIx ∷ Ptr → TxIx

ptrCertIx ∷ Ptr → CertIx

newtype SlotNo32

Pointers have been deprecated and aren't used anymore. For this reason we can safely use Word32 for slots in pointers

Constructors

SlotNo32 Word32

type StakeCredential = Credential 'Staking

data StakeReference

Constructors

StakeRefBase !StakeCredential

StakeRefPtr !Ptr

StakeRefNull

normalizePtr ∷ Ptr → Ptr

Deprecated: Starting with Conway era all Pointers are now normalized and this logic has been moved into the decoder

Convert any invalid Ptr to a Ptr that contains all zeros for its fields. Any pointer that contains a SlotNo, TxIx or CertIx that is too large to fit into Word32, Word16 and Word16 respectively is considered to be an invalid Ptr. Valid Ptrs will be returned unmodified.

Note - This is in no way related to dangling pointers, with an exception that any invalid Ptr is guarateed to be a dangling Ptr.