Cardano.Ledger.Crypto

Description

Deprecated: Cardano.Ledger.Crypto interface have been completely overhauled. In particular: 1). Crypto type class no longer contains DSIGN, HASH and ADDRHASH type families, instead they have been extracted into type synonyms that point to exact algorithms that previously where specified in StandardCrypto for those type families: a). DSIGN type synonym can be imported from Cardano.Ledger.Keys module b). HASH and ADDRHASH type synonyms can be imported from Cardano.Ledger.Hashes module 2). Crypto type class has retained its KES and VRF type families, which are not used in Ledger, therefore this type class as well as StandardCrypto definition will be migrated to Cardano.Protocol.Crypto module and from now on should be imported from 'cardano-protocol-tpraos' package instead.

Documentation

class (UnsoundPureKESAlgorithm (KES c), VRFAlgorithm (VRF c), ContextKES (KES c) ~ (), ContextVRF (VRF c) ~ (), Typeable c) ⇒ Crypto c

Associated Types

type KES c ∷ Type

type VRF c ∷ Type

data StandardCrypto

The same crypto used on the net

type HASH = Blake2b_256

Hashing algorithm used for hashing everything, except addresses, for which ADDRHASH is used.

type ADDRHASH = Blake2b_224

Hashing algorithm used for hashing cryptographic keys and scripts. As the type synonym name alludes, this is the hashing algorithm used for addresses.

type DSIGN = Ed25519DSIGN

Cryptographic signing algorithm used on Cardano blockchain.