Cardano.Ledger.Genesis

Documentation

class Era era ⇒ EraGenesis era

Associated Types

type Genesis era ∷ Type

type Genesis era = NoGenesis era

data NoGenesis era

Constructors

NoGenesis

mkGenesisWith

Arguments

∷ ∀ era g m. (Typeable era, Typeable g, Monad m)

⇒ m g
Action that will produce the genesis type, but it will be invoked only when it is statically known for g to not be NoGenesis

→ m g

Helper function that will produce NoGenesis for types that match it, otherwise action will be called that will produce the desired genesis type.