Cardano.Ledger.Internal.Definition.Era

Documentation

class (Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT) ⇒ Era era where

Minimal complete definition

Nothing

Associated Types

type EraName era ∷ Symbol

type PreviousEra era = (r ∷ Type) | r → era

Map an era to its predecessor.

For example:

type instance PreviousEra AllegraEra = ShelleyEra

type ProtVerLow era ∷ Nat

Lowest major protocol version for this era

type ProtVerHigh era ∷ Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

Methods

eraName ∷ String

Textual name of the current era.

Designed to be used with TypeApplications:

>>> eraName @ByronEra
"Byron"

default eraName ∷ KnownSymbol (EraName era) ⇒ String

class (KnownSymbol eraName, Era (EraFromName eraName), EraName (EraFromName eraName) ~ eraName) ⇒ EraHasName eraName

This class exists in order to be able to derive the protocol version range for an era from its name. It is achieved by the means of injective definition of an era type from its name. This effectively closes the world for eras to only the ones that are deined in this module, however it does have to be that way. In other words, if there is ever a need to define custom eras that build on top of existing ledger eras, then we will need to remove injectivity from this type class as well as from the PreviousEra type family.

Associated Types

type EraFromName eraName = (era ∷ Type) | era → eraName

data ByronEra

This is the era that preceded Shelley era. It cannot have any other class instances, except for Era type class.

data ShelleyEra

data AllegraEra

data MaryEra

data AlonzoEra

data BabbageEra

data ConwayEra

data DijkstraEra