Cardano.Ledger.MemoBytes

Description

MemoBytes is an abstraction for a data type that encodes its own serialization. The idea is to use a newtype around a MemoBytes applied to a non-memoizing type. For example: newtype Foo = Foo (MemoBytes NonMemoizingFoo) This way all the instances for Foo (Eq, Show, EncCBOR, DecCBOR, NoThunks, Generic`) can be derived for free. MemoBytes plays an important role in the SafeToHash class introduced in the module SafeHash

Documentation

data MemoBytes t where

Pair together a type t and its serialization. Used to encode a type that is serialized over the network, and to remember the original bytes that were used to transmit it. Important since hashes are computed from the serialization of a type, and EncCBOR instances do not have unique serializations.

Bundled Patterns

pattern Memo ∷ t → ShortByteString → MemoBytes t

type family MemoHashIndex (t ∷ Type) ∷ Type

type Mem t = Annotator (MemoBytes t)

Useful when deriving DecCBOR(Annotator T) deriving via (Mem T) instance DecCBOR (Annotator T)

mkMemoBytes ∷ ∀ t. t → ByteString → MemoBytes t

Smart constructor

getMemoBytesType ∷ MemoBytes t → t

Extract the inner type of the MemoBytes

getMemoBytesHash ∷ MemoBytes t → SafeHash (MemoHashIndex t)

Extract the hash value of the binary representation of the MemoBytes

memoBytes ∷ Version → Encode w t → MemoBytes t

Create MemoBytes from its CBOR encoding

memoBytesEra ∷ ∀ era w t. Era era ⇒ Encode w t → MemoBytes t

shorten ∷ ByteString → ShortByteString

Turn a lazy bytestring into a short bytestring.

showMemo ∷ Show t ⇒ MemoBytes t → String

Turn a MemoBytes into a string, Showing both its internal structure and its original bytes. Useful since the Show instance of MemoBytes does not display the original bytes.

printMemo ∷ Show t ⇒ MemoBytes t → IO ()

shortToLazy ∷ ShortByteString → ByteString

Helper function. Converts a short bytestring to a lazy bytestring.

contentsEq ∷ Eq t ⇒ MemoBytes t → MemoBytes t → Bool

Returns true if the contents of the MemoBytes are equal

decodeMemoBytes ∷ ∀ t m. (Typeable t, DecCBOR (Annotator t), MonadFail m) ⇒ Version → ByteString → m (MemoBytes t)

Memoized

class Memoized t

Class that relates the actual type with its raw and byte representations

Associated Types

type RawType t = (r ∷ Type) | r → t

mkMemoized ∷ ∀ t. (EncCBOR (RawType t), Memoized t) ⇒ Version → RawType t → t

Construct memoized type from the raw type using its EncCBOR instance

mkMemoizedEra ∷ ∀ era t. (Era era, EncCBOR (RawType t), Memoized t) ⇒ RawType t → t

decodeMemoized ∷ Decoder s t → Decoder s (MemoBytes t)

getMemoSafeHash ∷ Memoized t ⇒ t → SafeHash (MemoHashIndex (RawType t))

Extract memoized SafeHash

getMemoRawType ∷ Memoized t ⇒ t → RawType t

Extract the raw type from the memoized version

zipMemoRawType ∷ (Memoized t1, Memoized t2) ⇒ (RawType t1 → RawType t2 → a) → t1 → t2 → a

This is a helper function that operates on raw types of two memoized types.

eqRawType ∷ ∀ t. (Memoized t, Eq (RawType t)) ⇒ t → t → Bool

getMemoRawBytes ∷ Memoized t ⇒ t → ShortByteString

Extract the raw bytes from the memoized version

lensMemoRawType ∷ ∀ era t a b. (Era era, EncCBOR (RawType t), Memoized t) ⇒ (RawType t → a) → (RawType t → b → RawType t) → Lens t t a b

This is a helper Lens creator for any Memoized type.

getterMemoRawType ∷ Memoized t ⇒ (RawType t → a) → SimpleGetter t a

This is a helper SimpleGetter creator for any Memoized type

MemoBytes MemPack definitions

byteCountMemoBytes ∷ MemoBytes t → Int

packMemoBytesM ∷ MemoBytes t → Pack s ()

unpackMemoBytesM ∷ ∀ t b. (Typeable t, DecCBOR (Annotator t), Buffer b) ⇒ Version → Unpack b (MemoBytes t)

Raw equality

class EqRaw a where

Type class that implements equality on the Haskell type, ignoring any of the potentially memoized binary representation of the type.

Minimal complete definition

Nothing

Methods

eqRaw ∷ a → a → Bool

default eqRaw ∷ (a ~ t, Memoized t, Eq (RawType t)) ⇒ a → a → Bool