Cardano.Ledger.Plutus.Data

Documentation

newtype PlutusData era

This is a wrapper with a phantom era for PV1.Data, since we need something with kind (* -> *) for MemoBytes

Constructors

PlutusData Data

data Data era where

Bundled Patterns

pattern Data :: Era era => Data -> Data era

unData :: Data era -> Data

type DataHash c = SafeHash c EraIndependentData

upgradeData :: (Era era1, Era era2) => Data era1 -> Data era2

Upgrade Data from one era to another. While the underlying data will remain the same, the memoised serialisation may change to reflect the versioned serialisation of the new era.

hashData :: Era era => Data era -> DataHash (EraCrypto era)

getPlutusData :: Data era -> Data

dataHashSize :: StrictMaybe (DataHash c) -> Integer

data BinaryData era

Inlined data must be stored in the most compact form because it contributes to the memory overhead of the ledger state. Constructor is intentionally not exported, in order to prevent invalid creation of data from arbitrary binary data. Use makeBinaryData for smart construction.

hashBinaryData :: Era era => BinaryData era -> DataHash (EraCrypto era)

makeBinaryData :: Era era => ShortByteString -> Either String (BinaryData era)

Construct BinaryData from a buffer of bytes, while ensuring that it can be later safely converted to Data with binaryDataToData

binaryDataToData :: Era era => BinaryData era -> Data era

It is safe to convert BinaryData to Data because the only way to construct BinaryData is through the smart constructor makeBinaryData that takes care of validation.

dataToBinaryData :: Era era => Data era -> BinaryData era

data Datum era

Datum can be described by a either a data hash or binary data, but not both. It can also be neither one of them.

Constructors

NoDatum

DatumHash !(DataHash (EraCrypto era))

Datum !(BinaryData era)

datumDataHash :: Era era => Datum era -> StrictMaybe (DataHash (EraCrypto era))

Get the Hash of the datum.

translateDatum :: EraCrypto era1 ~ EraCrypto era2 => Datum era1 -> Datum era2

Orphan instances

Crypto c => HeapWords (StrictMaybe (DataHash c))

Instance details

Methods

heapWords :: StrictMaybe (DataHash c) -> Int