Cardano.Ledger.Plutus.Evaluate

Documentation

data PlutusWithContext c where

This type contains all that is necessary from Ledger to evaluate a plutus script.

Constructors

PlutusWithContext

Fields

∷ PlutusLanguage l

⇒ { pwcProtocolVersion ∷ !Version

Mayjor protocol version that is necessary for [de]serialization

, pwcScript ∷ !(Either (Plutus l) (PlutusRunnable l))

Actual plutus script that will be evaluated. Script is allowed to be in two forms: serialized and deserialized. This is necesary for implementing the opptimization that preserves deserialized PlutusRunnable after verifying wellformedness of plutus scripts during transaction validation (yet to be implemented).

, pwcScriptHash ∷ !(ScriptHash c)

, pwcDatums ∷ !PlutusDatums

All of the arguments to the Plutus scripts, including the redeemer and the Plutus context that was obtained from the transaction translation

, pwcExUnits ∷ !ExUnits

Limit on the execution units

, pwcCostModel ∷ !CostModel

CostModel to be used during script evaluation. It must match the language version in the pwcScript

} → PlutusWithContext c

data ScriptFailure c

Constructors

ScriptFailure

Fields

scriptFailureMessage ∷ Text

scriptFailurePlutus ∷ PlutusWithContext c

data ScriptResult c

Constructors

Passes [PlutusWithContext c]

Fails [PlutusWithContext c] (NonEmpty (ScriptFailure c))

scriptPass ∷ PlutusWithContext c → ScriptResult c

scriptFail ∷ ScriptFailure c → ScriptResult c

newtype PlutusDatums

Constructors

PlutusDatums

Fields

unPlutusDatums ∷ [Data]

data PlutusDebugInfo c

Constructors

DebugBadHex String

DebugCannotDecode String

DebugSuccess [Text] ExBudget

DebugFailure [Text] EvaluationError (PlutusWithContext c)

debugPlutus ∷ Crypto c ⇒ String → PlutusDebugInfo c

runPlutusScript ∷ PlutusWithContext c → ScriptResult c

runPlutusScriptWithLogs ∷ PlutusWithContext c → ([Text], ScriptResult c)

evaluatePlutusWithContext ∷ VerboseMode → PlutusWithContext c → ([Text], Either EvaluationError ExBudget)

explainPlutusEvaluationError ∷ PlutusWithContext c → EvaluationError → ScriptResult c

Explain why a script might fail. Scripts come in two flavors:

with 3 data arguments [data,redeemer,context]

with 2 data arguments [redeemer,context].

It pays to decode the context data into a real context because that provides way more information. But there is no guarantee the context data really can be decoded.