Cardano.Ledger.Plutus.ExUnits

Documentation

txscriptfee ∷ Prices → ExUnits → Coin

Compute the cost of a script based upon prices and the number of execution units.

pointWiseExUnits ∷ (Natural → Natural → Bool) → ExUnits → ExUnits → Bool

It is deliberate that there is no Ord instance for ExUnits. Use this function to compare if one ExUnit is pointwise compareable to another. In case when Ord instance like comparison is necessary you can use zipSemiExUnits compare

zipSemiExUnits ∷ Semigroup a ⇒ (Natural → Natural → a) → ExUnits → ExUnits → a

Pointwise combine units into a semigroup and mappened the results.

newtype ExUnits

This newtype wrapper of ExUnits' is used to hide an implementation detail inside the ExUnits pattern.

Constructors

WrapExUnits

Fields

unWrapExUnits ∷ ExUnits' Natural

Bundled Patterns

pattern ExUnits ∷ Natural → Natural → ExUnits
Arbitrary execution unit in which we measure the cost of scripts in terms of space in memory and execution time.

This pattern hides the fact that ExUnits' is parametric in the underlying type. The ledger itself uses ExUnits Natural' exclusively.

We would have preferred to use a type alias for ExUnits Natural', but this is not possible: https://gitlab.haskell.org/ghc/ghc/-/issues/19507.

data ExUnits' a

Arbitrary execution unit in which we measure the cost of scripts in terms of space in memory and execution time.

The ledger itself uses ExUnits Natural' exclusively, but the flexibility here allows the consensus layer to translate the execution units into something equivalent to 'ExUnits (Inf Natural)'. This is needed in order to provide a BoundedMeasure instance, which itself is needed for the alonzo instance of TxLimits (in consensus).

Constructors

ExUnits'

Fields

exUnitsMem' ∷ !a

exUnitsSteps' ∷ !a

data Prices

Prices per execution unit

Constructors

Prices

Fields

prMem ∷ !NonNegativeInterval

prSteps ∷ !NonNegativeInterval