Cardano.Ledger.PoolParams

Description

Deprecated: Use State instead.

This module is deprecated and will be removed in a future version. Please import these types from Cardano.Ledger.State instead.

The contents of this module have been moved to Cardano.Ledger.State.StakePool to provide better separation between pool registration parameters and pool state management. All types are now re-exported through Cardano.Ledger.State for convenient access.

Documentation

data PoolParams

A stake pool.

Constructors

PoolParams

Fields

ppId ∷ !(KeyHash 'StakePool)

ppVrf ∷ !(VRFVerKeyHash 'StakePoolVRF)

ppPledge ∷ !Coin

ppCost ∷ !Coin

ppMargin ∷ !UnitInterval

ppRewardAccount ∷ !RewardAccount

ppOwners ∷ !(Set (KeyHash 'Staking))

ppRelays ∷ !(StrictSeq StakePoolRelay)

ppMetadata ∷ !(StrictMaybe PoolMetadata)

data PoolMetadata

Constructors

PoolMetadata

Fields

pmUrl ∷ !Url

pmHash ∷ !ByteString

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

data SizeOfPoolRelays

The size of the ppRelays Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolRelays

data SizeOfPoolOwners

The size of the ppOwners Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolOwners