Cardano.Ledger.Serialization

Documentation

class Typeable a => ToCBORGroup a where

Methods

toCBORGroup :: a -> Encoding

encodedGroupSizeExpr :: (forall x. ToCBOR x => Proxy x -> Size) -> Proxy a -> Size

listLen :: a -> Word

listLenBound :: Proxy a -> Word

an upper bound for listLen, used in Size expressions.

class Typeable a => FromCBORGroup a where

Methods

fromCBORGroup :: Decoder s a

newtype CBORGroup a

Constructors

CBORGroup

Fields

unCBORGroup :: a

newtype CborSeq a

Constructors

CborSeq

Fields

unwrapCborSeq :: Seq a

decodeList :: Decoder s a -> Decoder s [a]

decodeSeq :: Decoder s a -> Decoder s (Seq a)

decodeStrictSeq :: Decoder s a -> Decoder s (StrictSeq a)

decodeSet :: Ord a => Decoder s a -> Decoder s (Set a)

decodeMap :: Ord a => Decoder s a -> Decoder s b -> Decoder s (Map a b)

decodeMapContents :: Decoder s a -> Decoder s [a]

decodeMapTraverse :: (Ord a, Applicative t) => Decoder s (t a) -> Decoder s (t b) -> Decoder s (t (Map a b))

decodeMaybe :: Decoder s a -> Decoder s (Maybe a)

decodeRecordNamed :: Text -> (a -> Int) -> Decoder s a -> Decoder s a

decodeRecordNamedT :: (MonadTrans m, Monad (m (Decoder s))) => Text -> (a -> Int) -> m (Decoder s) a -> m (Decoder s) a

decodeRecordSum :: String -> (Word -> Decoder s (Int, a)) -> Decoder s a

decodeNullMaybe :: Decoder s a -> Decoder s (Maybe a)

encodeFoldable :: (ToCBOR a, Foldable f) => f a -> Encoding

encodeFoldableEncoder :: Foldable f => (a -> Encoding) -> f a -> Encoding

encodeFoldableMapEncoder :: Foldable f => (Word -> a -> Maybe Encoding) -> f a -> Encoding

encodeNullMaybe :: (a -> Encoding) -> Maybe a -> Encoding

encodeMap :: (a -> Encoding) -> (b -> Encoding) -> Map a b -> Encoding

groupRecord :: forall a s. (ToCBORGroup a, FromCBORGroup a) => Decoder s a

ratioToCBOR :: ToCBOR a => Ratio a -> Encoding

ratioFromCBOR :: (Bounded a, Integral a, FromCBOR a) => Decoder s (Ratio a)

mapToCBOR :: (ToCBOR a, ToCBOR b) => Map a b -> Encoding

mapFromCBOR :: (Ord a, FromCBOR a, FromCBOR b) => Decoder s (Map a b)

translateViaCBORAnn :: (ToCBOR a, FromCBOR (Annotator b)) => Text -> a -> Except DecoderError b

ipv4ToBytes :: IPv4 -> ByteString

ipv4FromBytes :: ByteString -> Either String IPv4

ipv4ToCBOR :: IPv4 -> Encoding

ipv4FromCBOR :: Decoder s IPv4

ipv6ToBytes :: IPv6 -> ByteString

ipv6FromBytes :: ByteString -> Either String IPv6

ipv6ToCBOR :: IPv6 -> Encoding

ipv6FromCBOR :: Decoder s IPv6

listLenInt :: ToCBORGroup a => a -> Int

runByteBuilder :: Int -> Builder -> ByteString

Run a ByteString Builder using a strategy aimed at making smaller things efficiently.

It takes a size hint and produces a strict ByteString. This will be fast when the size hint is the same or slightly bigger than the true size.

utcTimeToCBOR :: UTCTime -> Encoding

utcTimeFromCBOR :: Decoder s UTCTime

data Sized a

A CBOR deserialized value together with its size. When deserializing use either sizedDecoder or its FromCBOR instance.

Use mkSized to construct such value.

Constructors

Sized

Fields

sizedValue :: !a

sizedSize :: Int64

Overhead in bytes. The field is lazy on purpose, because it might not be needed, but it can be expensive to compute.

mkSized :: ToCBOR a => a -> Sized a

Construct a Sized value by serializing it first and recording the amount of bytes it requires. Note, however, CBOR serialization is not canonical, therefore it is *NOT* a requirement that this property holds:

sizedSize (mkSized a) === sizedSize (unsafeDeserialize (serialize a) :: a)

sizedDecoder :: Decoder s a -> Decoder s (Sized a)