Cardano.Ledger.State

Documentation

class CanGetAccounts t where

Minimal complete definition

Nothing

Methods

accountsG ∷ SimpleGetter (t era) (Accounts era)

default accountsG ∷ CanSetAccounts t ⇒ SimpleGetter (t era) (Accounts era)

class CanGetAccounts t ⇒ CanSetAccounts t where

Methods

accountsL ∷ Lens' (t era) (Accounts era)

class (Era era, Eq (Accounts era), Show (Accounts era), Default (Accounts era), EncCBOR (Accounts era), DecShareCBOR (Accounts era), Share (Accounts era) ~ (Interns (Credential Staking), Interns (KeyHash StakePool), Interns (Credential DRepRole)), ToJSON (Accounts era), NFData (Accounts era), NoThunks (Accounts era), Eq (AccountState era), Show (AccountState era), NFData (AccountState era), NoThunks (AccountState era)) ⇒ EraAccounts era where

Associated Types

type AccountState era = (r ∷ Type) | r → era

type Accounts era = (r ∷ Type) | r → era

Methods

addAccountState ∷ Credential 'Staking → AccountState era → Accounts era → Accounts era

Add AccountState to Accounts. There are no checks whether account is already registered or not.

accountsMapL ∷ Lens' (Accounts era) (Map (Credential 'Staking) (AccountState era))

balanceAccountStateL ∷ Lens' (AccountState era) (CompactForm Coin)

depositAccountStateL ∷ Lens' (AccountState era) (CompactForm Coin)

stakePoolDelegationAccountStateL ∷ Lens' (AccountState era) (Maybe (KeyHash 'StakePool))

unregisterAccount

Arguments

∷ Credential 'Staking
Credential to unregister

→ Accounts era
Accounts to remove the account state from

→ (Maybe (AccountState era), Accounts era)
Returns Just whenever account was registered and Nothing otherwise. Produced Accounts will have the account state removed, if it was present there to begin with.

Remove the account from the state. Note that it is not capable of affecting state for DReps and StakePools, those have to be handled separately.

There is no counterpart for registering an account, because different eras require different information. However for testing purposed there is registerTestAccount that can be used for all eras.

lookupAccountState ∷ EraAccounts era ⇒ Credential 'Staking → Accounts era → Maybe (AccountState era)

Lookup an account state by its credential. Returns Nothing if such account is not registrered

updateLookupAccountState ∷ EraAccounts era ⇒ (AccountState era → AccountState era) → Credential 'Staking → Accounts era → (Maybe (AccountState era), Accounts era)

Update account state. Returns Nothing if the value is not present and modified value otherwise

isAccountRegistered ∷ EraAccounts era ⇒ Credential 'Staking → Accounts era → Bool

Check whether account for this staking credential is registered

adjustAccountState ∷ EraAccounts era ⇒ (AccountState era → AccountState era) → Credential 'Staking → Accounts era → Accounts era

lookupStakePoolDelegation ∷ EraAccounts era ⇒ Credential 'Staking → Accounts era → Maybe (KeyHash 'StakePool)

In case when account state is registered and it is delegated to a stake pool this function will return that delegation.

sumBalancesAccounts ∷ EraAccounts era ⇒ Accounts era → Coin

sumDepositsAccounts ∷ EraAccounts era ⇒ Accounts era → Coin

addToBalanceAccounts

Arguments

∷ EraAccounts era

⇒ Map (Credential 'Staking) (CompactForm Coin)
Map containing amounts that the balance in the account should be increased by. It is important to ensure that all of the credentials in this Map are actually registered.

→ Accounts era
Accounts that will have their balance increased.

→ Accounts era

Top up balance in accounts

Warning - it is an error to try to increase a balance of an account that is not present in Accounts

withdrawalsThatDoNotDrainAccounts

Arguments

∷ EraAccounts era

⇒ Withdrawals

→ Network

→ Accounts era

→ Maybe (Withdrawals, Withdrawals)
invalid withdrawal = that which does not have a reward account or is in the wrong network. incomplete withdrawal = that which does not withdraw the exact account balance.

This function returns Nothing iff all of the accounts that withdrawals are trying to drain are indeed registered and all of the amounts in the withdrawals match the respective balances exactly. It returns a 2-tuple where the fst is withdrawals with missing reward accounts or the wrong network, and snd is incomplete withdrawals.

NOTE: We simply checkBadWithdrawals to avoid allocating new variables for the most likely case.

drainAccounts ∷ EraAccounts era ⇒ Withdrawals → Accounts era → Accounts era

Reset balances to zero for all accounts that are specified in the supplied Withdrawals.

Note - There are no checks that withdrawals mention only registered accounts with correct NetworkId. Nor there are any checks that amounts in withdrawals match up the balance in the corresponding accounts. Use withdrawalsThatDoNotDrainAccounts to verify that calling drainAccounts is actually safe on the supplied arguments

removeStakePoolDelegations ∷ EraAccounts era ⇒ Set (KeyHash 'StakePool) → Accounts era → Accounts era

Remove delegations for the supplied Stake

class CanGetChainAccountState t where

Minimal complete definition

Nothing

Methods

chainAccountStateG ∷ SimpleGetter (t era) ChainAccountState

default chainAccountStateG ∷ CanSetChainAccountState t ⇒ SimpleGetter (t era) ChainAccountState

class CanGetChainAccountState t ⇒ CanSetChainAccountState t where

Methods

chainAccountStateL ∷ Lens' (t era) ChainAccountState

data ChainAccountState

Constructors

ChainAccountState

Fields

casTreasury ∷ !Coin

casReserves ∷ !Coin

Bundled Patterns

pattern AccountState ∷ Coin → Coin → ChainAccountState

Deprecated: In favor of ChainAccountState

casTreasuryL ∷ Lens' ChainAccountState Coin

casReservesL ∷ Lens' ChainAccountState Coin

treasuryL ∷ CanSetChainAccountState t ⇒ Lens' (t era) Coin

reservesL ∷ CanSetChainAccountState t ⇒ Lens' (t era) Coin

class (EraAccounts era, ToJSON (CertState era), EncCBOR (CertState era), DecShareCBOR (CertState era), Share (CertState era) ~ (Interns (Credential 'Staking), Interns (KeyHash 'StakePool), Interns (Credential 'DRepRole), Interns (Credential 'HotCommitteeRole)), Default (CertState era), NoThunks (CertState era), NFData (CertState era), Show (CertState era), Eq (CertState era)) ⇒ EraCertState era where

The state associated with the DELPL rule, which combines the DELEG rule and the POOL rule.

Associated Types

type CertState era = (r ∷ Type) | r → era

Methods

certDStateL ∷ Lens' (CertState era) (DState era)

certPStateL ∷ Lens' (CertState era) (PState era)

obligationCertState ∷ CertState era → Obligations

Calculate total possible refunds in the system that are related to certificates

There is an invariant that the sum of all the fields should be the same as the utxosDeposited field of the UTxOState. Note that this does not depend upon the current values of the Key and Pool deposits of the PParams.

certsTotalDepositsTxBody ∷ EraTxBody era ⇒ PParams era → CertState era → TxBody era → Coin

Compute the total deposits from the Certs of a TxBody.

This is the contribution of a TxBody towards the deposit pot (utxosDeposit field of the UTxOState) of the system

certsTotalRefundsTxBody ∷ EraTxBody era ⇒ PParams era → CertState era → TxBody era → Coin

Compute the total refunds from the Certs of a TxBody.

This is the contribution of a TxBody towards the total Obligations of the system See Obligations and obligationCertState for more information.

data CommitteeAuthorization

Constructors

CommitteeHotCredential !(Credential 'HotCommitteeRole)
Member authorized with a Hot credential acting on behalf of their Cold credential

CommitteeMemberResigned !(StrictMaybe Anchor)
Member resigned with a potential explanation in Anchor

data DState era

The state used by the DELEG rule, which roughly tracks stake delegation and some governance features.

Constructors

DState

Fields

dsAccounts ∷ !(Accounts era)

Keep track of the account state (eg. balance, deposit, stake-pool delegation, etc.) for all registered stake credentials.

dsFutureGenDelegs ∷ !(Map FutureGenDeleg GenDelegPair)

Future genesis key delegations

dsGenDelegs ∷ !GenDelegs

Genesis key delegations

dsIRewards ∷ !InstantaneousRewards

Instantaneous Rewards

data PState era

The state used by the POOL rule, which tracks stake pool information.

Constructors

PState

Fields

psVRFKeyHashes ∷ !(Map (VRFVerKeyHash 'StakePoolVRF) (NonZero Word64))

VRF key hashes that have been registered via PoolParams

psStakePools ∷ !(Map (KeyHash 'StakePool) StakePoolState)

The state of current stake pools.

psFutureStakePools ∷ !(Map (KeyHash 'StakePool) StakePoolState)

The state of future stake pools. Changes to existing stake pool parameters are staged in order to give delegators time to react to changes. See section 11.2, "Example Illustration of the Reward Cycle", of the Shelley Ledger Specification for a sequence diagram.

psRetiring ∷ !(Map (KeyHash 'StakePool) EpochNo)

A map of retiring stake pools to the epoch when they retire.

data InstantaneousRewards

InstantaneousRewards captures the pending changes to the ledger state caused by MIR certificates. It consists of two mappings, the rewards which will be paid out from the reserves and the rewards which will be paid out from the treasury. It also consists of two coin values which represent the transfer of coins from one pot to the other pot. NOTE that the following property should always hold: deltaReserves + deltaTreasury = 0

Constructors

InstantaneousRewards

Fields

iRReserves ∷ !(Map (Credential 'Staking) Coin)

iRTreasury ∷ !(Map (Credential 'Staking) Coin)

deltaReserves ∷ !DeltaCoin

deltaTreasury ∷ !DeltaCoin

data FutureGenDeleg

Constructors

FutureGenDeleg

Fields

fGenDelegSlot ∷ !SlotNo

fGenDelegGenKeyHash ∷ !(KeyHash 'Genesis)

data Anchor

Constructors

Anchor

Fields

anchorUrl ∷ !Url

anchorDataHash ∷ !(SafeHash AnchorData)

data DRepState

Constructors

DRepState

Fields

drepExpiry ∷ !EpochNo

drepAnchor ∷ !(StrictMaybe Anchor)

drepDeposit ∷ !(CompactForm Coin)

drepDelegs ∷ !(Set (Credential 'Staking))

data DRep

Constructors

DRepKeyHash !(KeyHash 'DRepRole)

DRepScriptHash !ScriptHash

DRepAlwaysAbstain

DRepAlwaysNoConfidence

Bundled Patterns

pattern DRepCredential ∷ Credential 'DRepRole → DRep

newtype CommitteeState era

Constructors

CommitteeState

Fields

csCommitteeCreds ∷ Map (Credential 'ColdCommitteeRole) CommitteeAuthorization

authorizedHotCommitteeCredentials ∷ CommitteeState era → Set (Credential 'HotCommitteeRole)

Extract all unique hot credential authorizations for the current committee. Note that there is no unique mapping from Hot to Cold credential, therefore we produce a Set, instead of a Map.

data AnchorData

lookupDepositDState ∷ EraAccounts era ⇒ DState era → StakeCredential → Maybe Coin

Function that looks up the deposit for currently delegated staking credential

lookupRewardDState ∷ EraAccounts era ⇒ DState era → StakeCredential → Maybe Coin

Function that looks up curret reward for the delegated staking credential.

data Obligations

A composite of all the Deposits the system is obligated to eventually pay back.

Constructors

Obligations

Fields

oblStake ∷ !Coin

oblPool ∷ !Coin

oblDRep ∷ !Coin

oblProposal ∷ !Coin

sumObligation ∷ Obligations → Coin

dsGenDelegsL ∷ Lens' (DState era) GenDelegs

dsIRewardsL ∷ Lens' (DState era) InstantaneousRewards

dsFutureGenDelegsL ∷ Lens' (DState era) (Map FutureGenDeleg GenDelegPair)

psStakePoolsL ∷ Lens' (PState era) (Map (KeyHash 'StakePool) StakePoolState)

psFutureStakePoolsL ∷ Lens' (PState era) (Map (KeyHash 'StakePool) StakePoolState)

psRetiringL ∷ Lens' (PState era) (Map (KeyHash 'StakePool) EpochNo)

psVRFKeyHashesL ∷ Lens' (PState era) (Map (VRFVerKeyHash 'StakePoolVRF) (NonZero Word64))

class CanGetUTxO t where

Minimal complete definition

Nothing

Methods

utxoG ∷ SimpleGetter (t era) (UTxO era)

default utxoG ∷ CanSetUTxO t ⇒ SimpleGetter (t era) (UTxO era)

class CanGetUTxO t ⇒ CanSetUTxO t where

Methods

utxoL ∷ Lens' (t era) (UTxO era)

Primitives

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map TxIn (TxOut era)

class EraTx era ⇒ EraUTxO era where

Associated Types

type ScriptsNeeded era = (r ∷ Type) | r → era

A customizable type on per era basis for the information required to find all scripts needed for the transaction.

Methods

consumed ∷ PParams era → CertState era → UTxO era → TxBody era → Value era

getConsumedValue

Arguments

∷ PParams era

→ (Credential 'Staking → Maybe Coin)
Function that can lookup current delegation deposits

→ (Credential 'DRepRole → Maybe Coin)
Function that can lookup current drep deposits

→ UTxO era

→ TxBody era

→ Value era

Calculate all the value that is being consumed by the transaction.

getProducedValue

Arguments

∷ PParams era

→ (KeyHash 'StakePool → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ TxBody era

→ Value era

getScriptsProvided

Arguments

∷ UTxO era
For some era it is necessary to look into the UTxO to find all of the available scripts for the transaction

→ Tx era

→ ScriptsProvided era

Initial eras will look into witness set to find all of the available scripts, but starting with Babbage we can look for available scripts in the UTxO using reference inputs.

getScriptsNeeded ∷ UTxO era → TxBody era → ScriptsNeeded era

Produce all the information required for figuring out which scripts are required for the transaction to be valid, once those scripts are evaluated

getScriptsHashesNeeded ∷ ScriptsNeeded era → Set ScriptHash

Extract the set of all script hashes that are needed for script validation.

getWitsVKeyNeeded ∷ CertState era → UTxO era → TxBody era → Set (KeyHash 'Witness)

Extract all of the KeyHash witnesses that are required for validating the transaction

getMinFeeTxUtxo ∷ PParams era → Tx era → UTxO era → Coin

Minimum fee computation, excluding witnesses and including ref scripts size

newtype ScriptsProvided era

The only reason it is a newtype instead of just a Map is becuase for later eras is expensive to compute the actual map, so we want to use the type safety guidance to avoid redundant work.

Constructors

ScriptsProvided

Fields

unScriptsProvided ∷ Map ScriptHash (Script era)

Functions

txins ∷ EraTxBody era ⇒ TxBody era → Set TxIn

Compute the UTxO inputs of a transaction. txins has the same problems as txouts, see notes below.

txinLookup ∷ TxIn → UTxO era → Maybe (TxOut era)

Lookup a txin for a given UTxO collection

txInsFilter

Arguments

∷ UTxO era
Source UTxO

→ Set TxIn
Which of the TxIns you would like to keep.

→ UTxO era

Filter out TxIn's from the UTxO map

txouts ∷ ∀ era. EraTxBody era ⇒ TxBody era → UTxO era

Compute the transaction outputs of a transaction.

sumUTxO ∷ EraTxOut era ⇒ UTxO era → Value era

Determine the total balance contained in the UTxO.

sumCoinUTxO ∷ EraTxOut era ⇒ UTxO era → Coin

Determine the total Ada only balance contained in the UTxO. This is equivalent to coin . sumUTxO, but it will be more efficient.

Warning - This function cannot be applied to an untrusted UTxO, since it is susceptible to overflow

balance ∷ EraTxOut era ⇒ UTxO era → Value era

Deprecated: In favor of sumUTxO

coinBalance ∷ EraTxOut era ⇒ UTxO era → Coin

Deprecated: In favor of sumCoinUTxO

sumAllValue ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Value era

Sum all the value in any Foldable with TxOuts

sumAllCoin ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Coin

Sum all the Coins in any Foldable with with TxOuts.

Warning - Care should be taken since it is susceptible to integer overflow, therefore make sure this function is not applied to unvalidated TxOuts

areAllAdaOnly ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Bool

Check whether any of the supplied TxOuts contain any MultiAssets. Returns True if non of them do.

verifyWitVKey ∷ Hash HASH EraIndependentTxBody → WitVKey kr → Bool

Verify a transaction body witness

getScriptHash ∷ Addr → Maybe ScriptHash

Extract script hash from value address with script.

data IndividualPoolStake

The IndividualPoolStake contains all the stake controlled by a single stake pool (the combination of owners and delegates) for a given epoch, and also the hash of the stake pool's registered VRF key.

When a stake pool produces a block, the header contains the full VRF verification key and VRF value for leader election. We check the VRF key against the value in IndividualPoolStake and we check the VRF value using the epoch nonce and the relative stake of the pool as given in IndividualPoolStake. The stake is relative to the total amount of active stake in the network. Stake is active if it is both registered and delegated to a registered stake pool.

Constructors

IndividualPoolStake

Fields

individualPoolStake ∷ !Rational

Pool stake distribution. This is a ratio of individualTotalPoolStake/pdTotalActiveStake

individualTotalPoolStake ∷ !(CompactForm Coin)

Total stake delegated to this pool. In addition to all the stake that is part of individualPoolStake we also add proposal-deposits to this field.

individualPoolStakeVrf ∷ !(VRFVerKeyHash 'StakePoolVRF)

data PoolDistr

A map of stake pool IDs (the hash of the stake pool operator's verification key) to IndividualPoolStake. Also holds absolute values necessary for the calculations in the computeDRepDistr.

Constructors

PoolDistr

Fields

unPoolDistr ∷ !(Map (KeyHash 'StakePool) IndividualPoolStake)

pdTotalActiveStake ∷ !(CompactForm Coin)

Total stake delegated to registered stake pools. In addition to the stake considered for the individualPoolStake Rational, we add proposal-deposits to this field.

poolDistrDistrL ∷ Lens' PoolDistr (Map (KeyHash 'StakePool) IndividualPoolStake)

poolDistrTotalL ∷ Lens' PoolDistr (CompactForm Coin)

individualTotalPoolStakeL ∷ Lens' IndividualPoolStake (CompactForm Coin)

newtype Stake

Type of stake as map from staking credential to coins associated. Any staking credential that has no stake will not appear in this Map, even if it is registered. For this reason, this data type should not be used for infering whether credential is registered or not.

Constructors

Stake

Fields

unStake ∷ VMap VB VP (Credential 'Staking) (CompactForm Coin)

sumAllStake ∷ Stake → Coin

sumAllStakeCompact ∷ Stake → CompactForm Coin

sumStakePerPool ∷ VMap VB VB (Credential 'Staking) (KeyHash 'StakePool) → Stake → Map (KeyHash 'StakePool) Coin

Compute amount of stake each pool has. Any registered stake pool that has no stake will not be inlcuded in the resulting map

data SnapShot

Snapshot of the stake distribution.

Constructors

SnapShot

Fields

ssStake ∷ !Stake

ssDelegations ∷ !(VMap VB VB (Credential 'Staking) (KeyHash 'StakePool))

ssPoolParams ∷ !(VMap VB VB (KeyHash 'StakePool) PoolParams)

data SnapShots

Snapshots of the stake distribution.

Note that ssStakeMark and ssStakeMarkPoolDistr are lazy on purpose since we only want to force the thunk after one stability window when we know that they are stable (so that we do not compute them if we do not have to). See more info in the Optimize TICKF ADR

Constructors

SnapShots

Fields

ssStakeMark ∷ SnapShot

ssStakeMarkPoolDistr ∷ PoolDistr

ssStakeSet ∷ !SnapShot

ssStakeGo ∷ !SnapShot

ssFee ∷ !Coin

emptySnapShot ∷ SnapShot

emptySnapShots ∷ SnapShots

poolStake ∷ KeyHash 'StakePool → VMap VB VB (Credential 'Staking) (KeyHash 'StakePool) → Stake → Stake

Get stake of one pool

maxPool ∷ EraPParams era ⇒ PParams era → Coin → Rational → Rational → Coin

Version of maxPool' that extracts ppA0L and ppNOptL from a PParams

maxPool' ∷ NonNegativeInterval → NonZero Word16 → Coin → Rational → Rational → Coin

Calculate maximal pool reward

calculatePoolDistr ∷ SnapShot → PoolDistr

calculatePoolDistr' ∷ (KeyHash 'StakePool → Bool) → SnapShot → PoolDistr

calculatePoolStake ∷ (KeyHash 'StakePool → Bool) → VMap VB VB (Credential 'Staking) (KeyHash 'StakePool) → Stake → Map (KeyHash 'StakePool) Word64

Sum up the Coin (as CompactForm Coin = Word64) for each StakePool

ssStakeMarkL ∷ Lens' SnapShots SnapShot

ssStakeMarkPoolDistrL ∷ Lens' SnapShots PoolDistr

ssStakeSetL ∷ Lens' SnapShots SnapShot

ssStakeGoL ∷ Lens' SnapShots SnapShot

ssFeeL ∷ Lens' SnapShots Coin

ssStakeL ∷ Lens' SnapShot Stake

ssStakeDistrL ∷ Lens' SnapShot (VMap VB VP (Credential 'Staking) (CompactForm Coin))

ssDelegationsL ∷ Lens' SnapShot (VMap VB VB (Credential 'Staking) (KeyHash 'StakePool))

ssPoolParamsL ∷ Lens' SnapShot (VMap VB VB (KeyHash 'StakePool) PoolParams)

class (EraAccounts era, Eq (InstantStake era), Show (InstantStake era), Monoid (InstantStake era), Default (InstantStake era), NFData (InstantStake era), NoThunks (InstantStake era), ToJSON (InstantStake era), EncCBOR (InstantStake era), DecShareCBOR (InstantStake era), Share (InstantStake era) ~ Interns (Credential 'Staking)) ⇒ EraStake era where

Associated Types

type InstantStake era = (r ∷ Type) | r → era

This is the current stake in the system. The important part of this stake is that not all of it is active. Any stake credential that is not registred will not contribute to the active stake, however it will be part of the instant stake. Throughout an epoch it is not relevant which part of the stake is active, because it is only when we take the snaphot that we resolve all the active stake.

Methods

instantStakeCredentialsL ∷ Lens' (InstantStake era) (Map (Credential 'Staking) (CompactForm Coin))

addInstantStake ∷ UTxO era → InstantStake era → InstantStake era

Add new UTxO to the InstantStake. This is invoked for every new TxOut that is added to the ledger state

deleteInstantStake ∷ UTxO era → InstantStake era → InstantStake era

Delete spent UTxO from the InstantStake. This is invoked for every TxOut that is removed from the ledger state

resolveInstantStake ∷ InstantStake era → Accounts era → Stake

Using known stake credential registrations and delegations resolve the instant stake into a Stake that will be used for SnapShot creation by snapShotFromInstantStake.

class CanGetInstantStake t where

Minimal complete definition

Nothing

Methods

instantStakeG ∷ SimpleGetter (t era) (InstantStake era)

default instantStakeG ∷ CanSetInstantStake t ⇒ SimpleGetter (t era) (InstantStake era)

class CanGetInstantStake t ⇒ CanSetInstantStake t where

Methods

instantStakeL ∷ Lens' (t era) (InstantStake era)

snapShotFromInstantStake ∷ ∀ era. EraStake era ⇒ InstantStake era → DState era → PState era → SnapShot

resolveActiveInstantStakeCredentials ∷ EraStake era ⇒ InstantStake era → Accounts era → Map (Credential 'Staking) (CompactForm Coin)

This is the total active stake including the rewards, but ignoring all the stake coming from the pointers. Where "active" stake means any stake credential that is registered and delegated to a stake pool.

Stake Pool State

data StakePoolState

State representation of a stake pool. This type contains all the same information as PoolParams except for the pool ID, which is stored separately as the key in state maps.

Constructors

StakePoolState

Fields

spsVrf ∷ !(VRFVerKeyHash 'StakePoolVRF)

VRF verification key hash for leader election

spsPledge ∷ !Coin

Pledge amount committed by the pool operator

spsCost ∷ !Coin

Fixed operational cost per epoch

spsMargin ∷ !UnitInterval

Pool profit margin (variable fee percentage)

spsRewardAccount ∷ !RewardAccount

Reward account for pool rewards

spsOwners ∷ !(Set (KeyHash 'Staking))

Set of stake key hashes that own this pool

spsRelays ∷ !(StrictSeq StakePoolRelay)

Network relay information for pool connectivity

spsMetadata ∷ !(StrictMaybe PoolMetadata)

Optional metadata (URL and hash)

spsDeposit ∷ !(CompactForm Coin)

Deposit for each pool

Lenses

spsVrfL ∷ Lens' StakePoolState (VRFVerKeyHash 'StakePoolVRF)

spsPledgeL ∷ Lens' StakePoolState Coin

spsCostL ∷ Lens' StakePoolState Coin

spsMarginL ∷ Lens' StakePoolState UnitInterval

spsRewardAccountL ∷ Lens' StakePoolState RewardAccount

spsOwnersL ∷ Lens' StakePoolState (Set (KeyHash 'Staking))

spsRelaysL ∷ Lens' StakePoolState (StrictSeq StakePoolRelay)

spsMetadataL ∷ Lens' StakePoolState (StrictMaybe PoolMetadata)

spsDepositL ∷ Lens' StakePoolState (CompactForm Coin)

Conversions

mkStakePoolState ∷ CompactForm Coin → PoolParams → StakePoolState

Convert PoolParams to StakePoolState by dropping the pool ID. This is the primary way to create a StakePoolState from registration or update parameters.

stakePoolStateToPoolParams ∷ KeyHash 'StakePool → StakePoolState → PoolParams

Convert StakePoolState back to PoolParams by providing the pool ID. This is useful when you need to reconstruct the full parameters from the state representation.

Pool Parameters and Related Types

These types were previously defined in PoolParams. They are now part of this module and re-exported by State.

data PoolParams

A stake pool.

Constructors

PoolParams

Fields

ppId ∷ !(KeyHash 'StakePool)

ppVrf ∷ !(VRFVerKeyHash 'StakePoolVRF)

ppPledge ∷ !Coin

ppCost ∷ !Coin

ppMargin ∷ !UnitInterval

ppRewardAccount ∷ !RewardAccount

ppOwners ∷ !(Set (KeyHash 'Staking))

ppRelays ∷ !(StrictSeq StakePoolRelay)

ppMetadata ∷ !(StrictMaybe PoolMetadata)

data PoolMetadata

Constructors

PoolMetadata

Fields

pmUrl ∷ !Url

pmHash ∷ !ByteString

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

data SizeOfPoolRelays

The size of the ppRelays Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolRelays

data SizeOfPoolOwners

The size of the ppOwners Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolOwners

ppCostL ∷ Lens' PoolParams Coin

ppMetadataL ∷ Lens' PoolParams (StrictMaybe PoolMetadata)

ppVrfL ∷ Lens' PoolParams (VRFVerKeyHash 'StakePoolVRF)

class (EraPParams era, Eq (GovState era), Show (GovState era), NoThunks (GovState era), NFData (GovState era), EncCBOR (GovState era), DecCBOR (GovState era), DecShareCBOR (GovState era), Share (GovState era) ~ (Interns (Credential 'Staking), Interns (KeyHash 'StakePool), Interns (Credential 'DRepRole), Interns (Credential 'HotCommitteeRole)), ToCBOR (GovState era), FromCBOR (GovState era), Default (GovState era), ToJSON (GovState era)) ⇒ EraGov era where

Minimal complete definition

curPParamsGovStateL, prevPParamsGovStateL, futurePParamsGovStateL, obligationGovState

Associated Types

type GovState era = (r ∷ Type) | r → era

Methods

emptyGovState ∷ GovState era

Construct empty governance state

curPParamsGovStateL ∷ Lens' (GovState era) (PParams era)

Lens for accessing current protocol parameters

prevPParamsGovStateL ∷ Lens' (GovState era) (PParams era)

Lens for accessing the previous protocol parameters

futurePParamsGovStateL ∷ Lens' (GovState era) (FuturePParams era)

Lens for accessing the future protocol parameters.

This lens will produce DefinitePParamsUpdate whenever we are absolutely sure that the new PParams will be updated. Which means there will be no chance of a DefinitePParamsUpdate value until we are past the point of no return, which is 2 stability windows before the end of the epoch. This lens is mostly intended for ledger usage and nextEpochUpdatedPParams should be used instead whenever definite results are desired.

obligationGovState ∷ GovState era → Obligations

data FuturePParams era

Constructors

NoPParamsUpdate
This indicates that there is definitely not going to be an update to PParams expected at the next epoch boundary.

DefinitePParamsUpdate !(PParams era)
This case specifies the PParams that will be adopted at the next epoch boundary.

PotentialPParamsUpdate (Maybe (PParams era))
With this case there is no guarantee that these will be the new PParams, users should not rely on this value to be computed efficiently and should use nextEpochPParams instead. The field is lazy on purpose, since we truly need to compute this field only towards the end of the epoch, which is done by solidifyFuturePParams two stability windows before the end of the epoch.

solidifyFuturePParams ∷ FuturePParams era → FuturePParams era

nextEpochPParams ∷ EraGov era ⇒ GovState era → PParams era

This function is guaranteed to produce PParams that will be adopted at the next epoch boundary, whenever this function is applied to the GovState that was produced by ledger at any point that is two stability windows before the end of the epoch. If you need to know if there were actual changes to those PParams then use nextEpochUpdatedPParams instead.

nextEpochUpdatedPParams ∷ EraGov era ⇒ GovState era → StrictMaybe (PParams era)

This function is guaranteed to return updated PParams when it is called during the last two stability windows of the epoch and there were proposals to update PParams that all relevant parties reached consensus on. In other words whenever there is a definite update to PParams coming on the epoch boundary those PParams will be returned, otherwise it will return Nothing. This function is inexpensive and can be invoked at any time without danger of forcing some suspended computation.

knownFuturePParams ∷ FuturePParams era → Maybe (PParams era)

Return new PParams only when it is known that there was an update proposed and it is guaranteed to be applied