Cardano.Ledger.TxIn

Documentation

newtype TxId

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

unTxId ∷ SafeHash EraIndependentTxBody

data TxIn

The input of a UTxO.

Constructors

TxIn !TxId !TxIx

mkTxInPartial ∷ HasCallStack ⇒ TxId → Integer → TxIn

Construct TxIn while throwing an error for an out of range TxIx. Make sure to use it only for testing.

txInToText ∷ TxIn → Text

data TxIx

Transaction index.