Cardano.Ledger.UMap

Description

A UMap (for Unified map) represents

4 Maps with the same domain in one direction, as a single Map and

1 other Map which is an inverse of one of the other 4 Maps.

The advantage of using UMap is that it stores all the information compactly, by exploiting the large amount of sharing in Map #1.

As for the other Map #2, we don't expect it to have much volume.

Constructing a UMap

data RDPair

A Reward-Deposit Pair Used to represent the reward and the deposit for a given (Credential 'Staking c)

Constructors

RDPair

Fields

rdReward ∷ !(CompactForm Coin)

rdDeposit ∷ !(CompactForm Coin)

rdRewardCoin ∷ RDPair → Coin

rdDepositCoin ∷ RDPair → Coin

data UMElem c where

A UMElem compactly represents the range of 4 Maps with the same domain as a single n-tuple.

This space-compacting datatype, and the pattern UMElem are equivalent to: data Elem c = Elem { rdPairT :: !(StrictMaybe RDPair), ptrT :: !(Set Ptr), sPoolT :: !(StrictMaybe (KeyHash 'StakePool c)), -- the stake pool identity dRepT :: !(StrictMaybe (DRep c)), } deriving (Show, Eq, Generic, NoThunks, NFData)

To name the constructors of UMElem we use the notation Txxx where each x is either F for full, i.e. the component is present, or E for empty, i.e. the component is not present.

There are four components: 1) the reward-deposit pair as an RDPair (CompactForm Coin) (CompactForm Coin) as a pair of Word64s, the first x, 2) the set of pointers, the second x, 3) the stake pool id (KeyHash 'StakePool c), the third x, and 4) the voting delegatee id (DRep c), the fourth x.

So, TEEEE means none of the components are present, TFEEE means only the reward-deposit pair is present, TEFEE means only the set of pointers is present, TEEFE means only the stake pool id is present. etc. TEEEF means only the voting delegatee id is present, and

The pattern UMElem will correctly use the optimal constructor.

Bundled Patterns

pattern UMElem ∷ StrictMaybe RDPair → Set Ptr → StrictMaybe (KeyHash 'StakePool c) → StrictMaybe (DRep c) → UMElem c
A UMElem can be extracted and injected into the TEEEE ... TFFFF constructors.

umElemRDPair ∷ UMElem c → Maybe RDPair

Extract the reward-deposit pair if it is present. We can tell that the reward is present when Txxxx has an F in the first position

This is equivalent to the pattern (UMElem (SJust r) _ _ _) -> Just r

umElemRDActive ∷ UMElem c → Maybe RDPair

Extract a delegated reward-deposit pair if it is present. We can tell that the pair is present and active when Txxxx has an F in the 1st position (present) and 3rd position (delegated).

This is equivalent to the pattern (UMElem (SJust r) _ (SJust _) _) -> Just r

data RewardDelegation c

Constructors

RewardDelegationSPO !(KeyHash 'StakePool c) !(CompactForm Coin)

RewardDelegationDRep !(DRep c) !(CompactForm Coin)

RewardDelegationBoth !(KeyHash 'StakePool c) !(DRep c) !(CompactForm Coin)

umElemDRepDelegatedReward ∷ UMElem c → Maybe (CompactForm Coin, DRep c)

Extract a DRep delegated reward if it is present. We can tell that the pair is present and active when Txxxx has an F in the 1st position (present) and 4rd position (DRep delegated).

This is equivalent to the pattern (UMElem (SJust r) _ _ (SJust d)) -> Just (r, d)

umElemDelegations ∷ UMElem c → Maybe (RewardDelegation c)

Extract rewards that are either delegated to a DRep or an SPO (or both). We can tell that the pair is present and active when Txxxx has F's in the 1st and either 3rd or 4th or both positions. If there are no rewards or deposits but the delegations still exist, then we return zero coin as reward.

umElemPtrs ∷ UMElem c → Maybe (Set Ptr)

Extract the set of pointers if it is non-empty. We can tell that the reward is present when Txxxx has an F in the second position

This is equivalent to the pattern (UMElem _ p _ _) -> Just p

umElemSPool ∷ UMElem c → Maybe (KeyHash 'StakePool c)

Extract the stake delegatee pool id, if present. We can tell that the pool id is present when Txxxx has an F in the third position

This is equivalent to the pattern (UMElem _ _ (SJust s) _) -> Just s

umElemDRep ∷ UMElem c → Maybe (DRep c)

Extract the voting delegatee id, if present. We can tell that the delegatee is present when Txxxx has an F in the fourth position

This is equivalent to the pattern (UMElem _ _ _ (SJust d)) -> Just d

umElemAsTuple ∷ UMElem c → (StrictMaybe RDPair, Set Ptr, StrictMaybe (KeyHash 'StakePool c), StrictMaybe (DRep c))

A n-Tuple view of the UMElem. We can view all of the constructors as an UMElem.

nullUMElem ∷ UMElem c → Bool

null for an UMElem

nullUMElemMaybe ∷ UMElem c → Maybe (UMElem c)

null Maybe for an UMElem

data UMap c

A unified map represents 4 Maps with domain (Credential 'Staking c)

1) Map (Credential 'Staking c) RDPair -- (RDPair rewardCoin depositCoin) 2) Map (Credential 'Staking c) (Set Ptr) 3) Map (Credential 'Staking c) (StrictMaybe (KeyHash 'StakePool c)) 4) Map (Credential 'Staking c) (StrictMaybe (DRep c)) and one more map in the inverse direction with Ptr for keys and (Credential 'Staking c) for values.

Constructors

UMap

Fields

umElems ∷ !(Map (Credential 'Staking c) (UMElem c))

umPtrs ∷ !(Map Ptr (Credential 'Staking c))

umElemsL ∷ Lens' (UMap c) (Map (Credential 'Staking c) (UMElem c))

empty ∷ UMap c

Construct an empty UMap

umInvariant ∷ Credential 'Staking c → Ptr → UMap c → Bool

It is worthwhile stating the invariant that holds on a Unified Map. The umPtrs and the ptrT field of the umElems are inverses.

StakeCredentials

data StakeCredentials c

All maps unrolled. It is important to note that all fields are lazy, because conversion from UMap can be expensive, thus only fields that are forced will incur that conversion overhead.

Constructors

StakeCredentials

Fields

scRewards ∷ Map (Credential 'Staking c) Coin

scDeposits ∷ Map (Credential 'Staking c) Coin

scSPools ∷ Map (Credential 'Staking c) (KeyHash 'StakePool c)

scDReps ∷ Map (Credential 'Staking c) (DRep c)

scPtrs ∷ Map Ptr (Credential 'Staking c)

scPtrsInverse ∷ Map (Credential 'Staking c) (Set Ptr)

There will be no empty sets in the range

toStakeCredentials ∷ UMap c → StakeCredentials c

domRestrictedStakeCredentials ∷ Set (Credential 'Staking c) → UMap c → StakeCredentials c

UView and its components

data UView c k v where

A UView lets one view a UMap in n different ways, one for each of the elements in a Unified Element UMElem (4) A (UView c key value) can be used like a (Map key value). It acts like a map, supporting efficient insert, delete, and lookup operations.

Constructors

RewDepUView ∷ !(UMap c) → UView c (Credential 'Staking c) RDPair

PtrUView ∷ !(UMap c) → UView c Ptr (Credential 'Staking c)

SPoolUView ∷ !(UMap c) → UView c (Credential 'Staking c) (KeyHash 'StakePool c)

DRepUView ∷ !(UMap c) → UView c (Credential 'Staking c) (DRep c)

rewDepUView ∷ Map (Credential 'Staking c) (UMElem c) → Map Ptr (Credential 'Staking c) → UView c (Credential 'Staking c) RDPair

Construct a RewDepUView from the two maps that make up a UMap

ptrUView ∷ Map (Credential 'Staking c) (UMElem c) → Map Ptr (Credential 'Staking c) → UView c Ptr (Credential 'Staking c)

Construct a PtrUView from the two maps that make up a UMap

sPoolUView ∷ Map (Credential 'Staking c) (UMElem c) → Map Ptr (Credential 'Staking c) → UView c (Credential 'Staking c) (KeyHash 'StakePool c)

Construct a SPoolUView from the two maps that make up a UMap

dRepUView ∷ Map (Credential 'Staking c) (UMElem c) → Map Ptr (Credential 'Staking c) → UView c (Credential 'Staking c) (DRep c)

Construct a DRepUView from the two maps that make up a UMap

unUView ∷ UView c k v → UMap c

Extract the underlying UMap from a UView

unUnifyToVMap ∷ UView c k v → VMap VB VB k v

Materialize a real VMap (Vector Map) from a UView This is expensive, use it wisely (like maybe once per epoch boundary to make a SnapShot)

rdPairMap ∷ UMap c → Map (Credential 'Staking c) RDPair

Extract a reward-deposit pairs Map from a UMap

rewardMap ∷ UMap c → Map (Credential 'Staking c) Coin

Extract a rewards Map from a UMap

compactRewardMap ∷ UMap c → Map (Credential 'Staking c) (CompactForm Coin)

Extract a compact rewards Map from a UMap

depositMap ∷ UMap c → Map (Credential 'Staking c) Coin

Extract a deposits Map from a UMap

ptrMap ∷ UMap c → Map Ptr (Credential 'Staking c)

Extract a pointers Map from a UMap

invPtrMap ∷ UMap c → Map (Credential 'Staking c) (Set Ptr)

Extract a pointers Map from a UMap

sPoolMap ∷ UMap c → Map (Credential 'Staking c) (KeyHash 'StakePool c)

Extract a stake pool delegations Map from a UMap

dRepMap ∷ UMap c → Map (Credential 'Staking c) (DRep c)

Extract a delegated-representatives Map from a UMap

domRestrictedMap ∷ Set k → UView c k v → Map k v

Extract a domain-restricted Map of a UMap. If `Set k` is small this should be efficient.

data family CompactForm a ∷ Type

toCompact ∷ Compactible a ⇒ a → Maybe (CompactForm a)

fromCompact ∷ Compactible a ⇒ CompactForm a → a

addCompact ∷ CompactForm Coin → CompactForm Coin → CompactForm Coin

sumCompactCoin ∷ Foldable t ⇒ t (CompactForm Coin) → CompactForm Coin

sumRewardsUView ∷ UView c k RDPair → CompactForm Coin

sumDepositUView ∷ UView c k RDPair → CompactForm Coin

compactCoinOrError ∷ HasCallStack ⇒ Coin → CompactForm Coin

unify ∷ Map (Credential 'Staking c) RDPair → Map Ptr (Credential 'Staking c) → Map (Credential 'Staking c) (KeyHash 'StakePool c) → Map (Credential 'Staking c) (DRep c) → UMap c

Create a UMap from 4 separate maps. NOTE: For use in tests only.

unUnify ∷ UView c k v → Map k v

Materialize a real Map from a View This is expensive, use it wisely (like maybe once per epoch boundary to make a SnapShot) See also domRestrictedMap, which domain-restricts before computing a view.

Set and Map operations on UViews

nullUView ∷ UView c k v → Bool

null for a UView, just like null

member ∷ k → UView c k v → Bool

Membership check for a UView, just like member

Spec: eval (k ∈ dom (rewards dState)) eval (k ∈ dom (rewards ds))) eval (hk ∈ dom (rewards ds)) eval (hk ∉ dom (rewards ds))

member' ∷ Credential 'Staking c → UMap c → Bool

Checks for membership directly against umElems instead of a UView.

notMember ∷ k → UView c k v → Bool

Membership check for a UView, just like member

Spec: eval (k ∈ dom (rewards dState)) eval (k ∈ dom (rewards ds))) eval (hk ∈ dom (rewards ds)) eval (hk ∉ dom (rewards ds))

delete ∷ k → UView c k v → UMap c

delete' ∷ k → UView c k v → UView c k v

Delete a key and its value from the map-like UView, returning a version of the same UView.

In the case of a PtrUView we maintain the umInvariant and delete the pairs from both umElems as well as umPtrs of the UMap.

insertWith ∷ (v → v → v) → k → v → UView c k v → UMap c

insertWith' ∷ (v → v → v) → k → v → UView c k v → UView c k v

Insert with combination

If k exists as a key in the (map-like) UView:

to keep the old value > insertWith' ( old new -> old) k v view

to replace the old value with the new value > insertWith' ( old new -> new) k v view

to combine the old and new values with summation > insertWith' ( old new -> old + new) k v view

If k does not exist as a key in the UView, the combining function is ignored, and the key k and the value v are inserted into the map-like UView > insertWith' ignoredCombiningFunction k v view

insert ∷ k → v → UView c k v → UMap c

insert' ∷ k → v → UView c k v → UView c k v

adjust ∷ (RDPair → RDPair) → k → UView c k RDPair → UMap c

Adjust a UView, just like adjust. This is implemented only for reward-deposit pairs.

lookup ∷ k → UView c k v → Maybe v

Lookup a UView, just like lookup.

domain ∷ UView c k v → Set k

Get the domain of the Map-like UView

range ∷ UView c k v → Set v

Get the range of the Map-like UView

(∪) ∷ UView c k v → (k, v) → UMap c

Union with left preference. So if k, already exists, do nothing, if it doesn't exist insert it.

Spec: evalUnified (RewDepUView u1 ∪ singleton hk mempty) evalUnified (Ptrs u2 ∪ singleton ptr hk)

unionL ∷ UView c k v → (k, v) → UMap c

Union with left preference. So if k, already exists, do nothing, if it doesn't exist insert it.

Spec: evalUnified (RewDepUView u1 ∪ singleton hk mempty) evalUnified (Ptrs u2 ∪ singleton ptr hk)

(⨃) ∷ UView c k v → Map k v → UMap c

Union with right preference. So if k, already exists, then old v is overwritten with the new v.

Special rules apply for the RewDepUView, where only the rdReward field of the RDPair is overwritten, and the old rdDeposit value persists.

Note: In this case it is an invariant that the domain of the Map on the right side is a subset of the domain of the RewDepUView. See the single case in module Cardano.Ledger.Shelley.Rules.Delegs, in the dealing with Withdrawals's where it is used at this type.

Spec: evalUnified (delegations ds ⨃ singleton hk dpool) evalUnified (rewards' ⨃ wdrls_')

unionR ∷ UView c k v → Map k v → UMap c

Union with right preference. So if k, already exists, then old v is overwritten with the new v.

Special rules apply for the RewDepUView, where only the rdReward field of the RDPair is overwritten, and the old rdDeposit value persists.

Note: In this case it is an invariant that the domain of the Map on the right side is a subset of the domain of the RewDepUView. See the single case in module Cardano.Ledger.Shelley.Rules.Delegs, in the dealing with Withdrawals's where it is used at this type.

Spec: evalUnified (delegations ds ⨃ singleton hk dpool) evalUnified (rewards' ⨃ wdrls_')

(∪+) ∷ UView c (Credential 'Staking c) RDPair → Map (Credential 'Staking c) (CompactForm Coin) → UMap c

Add the reward from the Map on the right side to the reward in the UView on the left. This is only implemented and is applicable to RewDepUViews.

We presume that the domain of the Map on the right, is a subset of the domain of the UView on the left.

Spec: evalUnified (rewards dState ∪+ registeredAggregated) evalUnified (rewards' ∪+ update) evalUnified (RewDepUView u0 ∪+ refunds)

unionRewAgg ∷ UView c (Credential 'Staking c) RDPair → Map (Credential 'Staking c) (CompactForm Coin) → UMap c

Add the reward from the Map on the right side to the reward in the UView on the left. This is only implemented and is applicable to RewDepUViews.

We presume that the domain of the Map on the right, is a subset of the domain of the UView on the left.

Spec: evalUnified (rewards dState ∪+ registeredAggregated) evalUnified (rewards' ∪+ update) evalUnified (RewDepUView u0 ∪+ refunds)

unionKeyDeposits ∷ UView c k RDPair → Map k (CompactForm Coin) → UMap c

Add the deposit from the Map on the right side to the deposit in the UView on the left. This is only implemented and is applicable to RewDepUViews.

(⋪) ∷ Set k → UView c k v → UMap c

Delete all keys in the given Set from the domain of the given map-like UView.

Spec: evalUnified (setSingleton hk ⋪ RewDepUView u0) evalUnified (setSingleton hk ⋪ SPoolUView u1)

domDelete ∷ Set k → UView c k v → UMap c

Delete all keys in the given Set from the domain of the given map-like UView.

Spec: evalUnified (setSingleton hk ⋪ RewDepUView u0) evalUnified (setSingleton hk ⋪ SPoolUView u1)

(⋫) ∷ UView c k v → Set v → UMap c

Delete all elements in the given Set from the range of the given map-like UView. This is slow for SPoolUView, RewDepUView, and DReps UViews, better hope the sets are small

Spec: evalUnified (Ptrs u2 ⋫ setSingleton hk) evalUnified (SPoolUView u1 ⋫ retired)

rngDelete ∷ UView c k v → Set v → UMap c

Delete all elements in the given Set from the range of the given map-like UView. This is slow for SPoolUView, RewDepUView, and DReps UViews, better hope the sets are small

Spec: evalUnified (Ptrs u2 ⋫ setSingleton hk) evalUnified (SPoolUView u1 ⋫ retired)

(◁) ∷ UView c k v → Map k u → Map k u

Domain restriction.

Spec: eval (dom rewards' ◁ iRReserves (_irwd ds) :: RewardAccounts (Crypto era)) eval (dom rewards' ◁ iRTreasury (_irwd ds) :: RewardAccounts (Crypto era))

domRestrict ∷ UView c k v → Map k u → Map k u

Domain restriction.

Spec: eval (dom rewards' ◁ iRReserves (_irwd ds) :: RewardAccounts (Crypto era)) eval (dom rewards' ◁ iRTreasury (_irwd ds) :: RewardAccounts (Crypto era))

Derived functions

findWithDefault ∷ v → k → UView c k v → v

Find the value associated with a key from a UView, return the default if the key is not there.

size ∷ UView c k v → Int

A UView is a view, so the size of the view is NOT the same as the size of the underlying UMElem map.

domDeleteAll ∷ Set (Credential 'Staking c) → UMap c → UMap c

Delete the stake credentials in the domain and all associated ranges from the UMap This can be expensive when there are many pointers associated with the credential.

deleteStakingCredential ∷ Credential 'Staking c → UMap c → UMap c

Completely remove the staking credential from the UMap, including all associated pointers.

extractStakingCredential ∷ Credential 'Staking c → UMap c → (Maybe (UMElem c), UMap c)

Just like deleteStakingCredential, but also returned the removed element.