Cardano.Ledger.UnifiedMap

Description

A Monomorphic version of UMap specialized to the needs of Cardano

Documentation

type ViewMap crypto = View Coin (Credential 'Staking crypto) (KeyHash 'StakePool crypto) Ptr

type UnifiedMap crypto = UMap Coin (Credential 'Staking crypto) (KeyHash 'StakePool crypto) Ptr

type Triple crypto = Trip Coin Ptr (KeyHash 'StakePool crypto)

data View coin cr pl ptr k v where

Constructors

Rewards :: forall coin cr pl ptr. !(UMap coin cr pl ptr) -> View coin cr pl ptr cr coin

Delegations :: forall coin cr pl ptr. !(UMap coin cr pl ptr) -> View coin cr pl ptr cr pl

Ptrs :: forall coin cr pl ptr. !(UMap coin cr pl ptr) -> View coin cr pl ptr ptr cr

data UMap coin cred pool ptr

Constructors

UnifiedMap !(Map cred (Trip coin ptr pool)) !(Map ptr cred)

data Trip coin ptr pool where

Bundled Patterns

pattern Triple :: StrictMaybe coin -> Set ptr -> StrictMaybe pool -> Trip coin ptr pool

class UnifiedView coin cred pool ptr k v where

Methods

tag :: Tag coin cred pool ptr k v

data Tag coin cred pool ptr k v where

Constructors

Rew :: forall coin cred pool ptr. Tag coin cred pool ptr cred coin

Del :: forall coin cred pool ptr. Tag coin cred pool ptr cred pool

Ptr :: forall coin cred pool ptr. Tag coin cred pool ptr ptr cred

Orphan instances

UnifiedView Coin (Credential 'Staking crypto) (KeyHash 'StakePool crypto) Ptr Ptr (Credential 'Staking crypto)

Instance details

Methods

tag :: Tag Coin (Credential 'Staking crypto) (KeyHash 'StakePool crypto) Ptr Ptr (Credential 'Staking crypto)

UnifiedView Coin (Credential 'Staking crypto) (KeyHash 'StakePool crypto) Ptr (Credential 'Staking crypto) Coin

Instance details

Methods

tag :: Tag Coin (Credential 'Staking crypto) (KeyHash 'StakePool crypto) Ptr (Credential 'Staking crypto) Coin

UnifiedView Coin (Credential 'Staking crypto) (KeyHash 'StakePool crypto) Ptr (Credential 'Staking crypto) (KeyHash 'StakePool crypto)

Instance details

Methods

tag :: Tag Coin (Credential 'Staking crypto) (KeyHash 'StakePool crypto) Ptr (Credential 'Staking crypto) (KeyHash 'StakePool crypto)