Cardano.Ledger.Val

Description

This module defines a generalised notion of a "value" - that is, something with which we may quantify a transaction output.

Documentation

class (Compactible t, Inject Coin t, EncCBOR (CompactForm t), DecCBOR (CompactForm t), Abelian t, NoThunks t, EncCBOR t, DecCBOR t, ToJSON t, NFData t, Show t, Eq t) => Val t where

Minimal complete definition

(<×>), coin, modifyCoin, size, pointwise, isAdaOnly, isAdaOnlyCompact, coinCompact, injectCompact, modifyCompactCoin

Methods

zero :: t

the value with nothing in it

(<+>) :: t -> t -> t infixl 6

add two value

(<×>) :: Integral i => i -> t -> t infixl 7

scale a value by an Integral constant

(<->) :: t -> t -> t infixl 6

subtract two values

isZero :: t -> Bool

Is the argument zero?

coin :: t -> Coin

Get the ADA present in the value (since ADA is our "blessed" currency)

modifyCoin :: (Coin -> Coin) -> t -> t

modify the blessed Coin part of t

size :: t -> Integer

pointwise :: (Integer -> Integer -> Bool) -> t -> t -> Bool

used to compare values pointwise. Rather than using: (v1 <= v2) use: pointwise (<=) v1 v2 | If a quantity is stored in only one of v1 or v2, we use 0 for the missing quantity.

isAdaOnly :: t -> Bool

Check if value contains only ADA. Must hold property:

inject (coin v) == v

isAdaOnlyCompact :: CompactForm t -> Bool

coinCompact :: CompactForm t -> CompactForm Coin

injectCompact :: CompactForm Coin -> CompactForm t

modifyCompactCoin :: (CompactForm Coin -> CompactForm Coin) -> CompactForm t -> CompactForm t

inject :: Inject t s => t -> s

scale :: (Val t, Integral i) => i -> t -> t

invert :: Val t => t -> t

sumVal :: (Foldable t, Val v) => t v -> v

adaOnly :: Val v => v -> Bool

Deprecated: In favor of isAdaOnly