Test.Cardano.Ledger.Common

Documentation

unless ∷ Applicative f ⇒ Bool → f () → f ()

The reverse of when.

forM_ ∷ (Foldable t, Monad m) ⇒ t a → (a → m b) → m ()

forM_ is mapM_ with its arguments flipped. For a version that doesn't ignore the results see forM.

forM_ is just like for_, but specialised to monadic actions.

when ∷ Applicative f ⇒ Bool → f () → f ()

Conditional execution of Applicative expressions. For example,

when debug (putStrLn "Debugging")

will output the string Debugging if the Boolean value debug is True, and otherwise do nothing.

type HasCallStack = ?callStack ∷ CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.9.0.0

data Gen a

class Arbitrary a where

Minimal complete definition

arbitrary

Methods

arbitrary ∷ Gen a

shrink ∷ a → [a]

example ∷ Expectation → Expectation

after ∷ ActionWith a → SpecWith a → SpecWith a

afterAll ∷ HasCallStack ⇒ ActionWith a → SpecWith a → SpecWith a

afterAll_ ∷ HasCallStack ⇒ IO () → SpecWith a → SpecWith a

after_ ∷ IO () → SpecWith a → SpecWith a

around ∷ (ActionWith a → IO ()) → SpecWith a → Spec

aroundAll ∷ HasCallStack ⇒ (ActionWith a → IO ()) → SpecWith a → Spec

aroundAllWith ∷ HasCallStack ⇒ (ActionWith a → ActionWith b) → SpecWith a → SpecWith b

aroundAll_ ∷ HasCallStack ⇒ (IO () → IO ()) → SpecWith a → SpecWith a

aroundWith ∷ (ActionWith a → ActionWith b) → SpecWith a → SpecWith b

around_ ∷ (IO () → IO ()) → SpecWith a → SpecWith a

before ∷ IO a → SpecWith a → Spec

beforeAll ∷ HasCallStack ⇒ IO a → SpecWith a → Spec

beforeAllWith ∷ HasCallStack ⇒ (b → IO a) → SpecWith a → SpecWith b

beforeAll_ ∷ HasCallStack ⇒ IO () → SpecWith a → SpecWith a

beforeWith ∷ (b → IO a) → SpecWith a → SpecWith b

before_ ∷ IO () → SpecWith a → SpecWith a

ignoreSubject ∷ SpecWith () → SpecWith a

mapSubject ∷ (b → a) → SpecWith a → SpecWith b

hspec ∷ Spec → IO ()

context ∷ HasCallStack ⇒ String → SpecWith a → SpecWith a

describe ∷ HasCallStack ⇒ String → SpecWith a → SpecWith a

fcontext ∷ HasCallStack ⇒ String → SpecWith a → SpecWith a

fdescribe ∷ HasCallStack ⇒ String → SpecWith a → SpecWith a

fit ∷ (HasCallStack, Example a) ⇒ String → a → SpecWith (Arg a)

focus ∷ SpecWith a → SpecWith a

fspecify ∷ (HasCallStack, Example a) ⇒ String → a → SpecWith (Arg a)

it ∷ (HasCallStack, Example a) ⇒ String → a → SpecWith (Arg a)

parallel ∷ SpecWith a → SpecWith a

pending ∷ HasCallStack ⇒ Expectation

pendingWith ∷ HasCallStack ⇒ String → Expectation

specify ∷ (HasCallStack, Example a) ⇒ String → a → SpecWith (Arg a)

xcontext ∷ HasCallStack ⇒ String → SpecWith a → SpecWith a

xdescribe ∷ HasCallStack ⇒ String → SpecWith a → SpecWith a

xit ∷ (HasCallStack, Example a) ⇒ String → a → SpecWith (Arg a)

xspecify ∷ (HasCallStack, Example a) ⇒ String → a → SpecWith (Arg a)

runIO ∷ IO r → SpecM a r

anyArithException ∷ Selector ArithException

anyErrorCall ∷ Selector ErrorCall

anyException ∷ Selector SomeException

anyIOException ∷ Selector IOException

errorCall ∷ String → Selector ErrorCall

expectationFailure ∷ HasCallStack ⇒ String → Expectation

shouldBe ∷ (HasCallStack, Show a, Eq a) ⇒ a → a → Expectation

shouldContain ∷ (HasCallStack, Show a, Eq a) ⇒ [a] → [a] → Expectation

shouldEndWith ∷ (HasCallStack, Show a, Eq a) ⇒ [a] → [a] → Expectation

shouldMatchList ∷ (HasCallStack, Show a, Eq a) ⇒ [a] → [a] → Expectation

shouldNotBe ∷ (HasCallStack, Show a, Eq a) ⇒ a → a → Expectation

shouldNotContain ∷ (HasCallStack, Show a, Eq a) ⇒ [a] → [a] → Expectation

shouldNotReturn ∷ (HasCallStack, Show a, Eq a) ⇒ IO a → a → Expectation

shouldNotSatisfy ∷ (HasCallStack, Show a) ⇒ a → (a → Bool) → Expectation

shouldReturn ∷ (HasCallStack, Show a, Eq a) ⇒ IO a → a → Expectation

shouldSatisfy ∷ (HasCallStack, Show a) ⇒ a → (a → Bool) → Expectation

shouldStartWith ∷ (HasCallStack, Show a, Eq a) ⇒ [a] → [a] → Expectation

shouldThrow ∷ (HasCallStack, Exception e) ⇒ IO a → Selector e → Expectation

type ActionWith a = a → IO ()

class Example e

Minimal complete definition

evaluateExample

Associated Types

type Arg e

type Arg e = ()

type family Arg e

type Spec = SpecWith ()

type SpecWith a = SpecM a ()

type Expectation = Assertion

type Selector a = a → Bool

fprop ∷ (HasCallStack, Testable prop) ⇒ String → prop → Spec

prop ∷ (HasCallStack, Testable prop) ⇒ String → prop → Spec

xprop ∷ (HasCallStack, Testable prop) ⇒ String → prop → Spec

modifyArgs ∷ (Args → Args) → SpecWith a → SpecWith a

modifyMaxDiscardRatio ∷ (Int → Int) → SpecWith a → SpecWith a

modifyMaxShrinks ∷ (Int → Int) → SpecWith a → SpecWith a

modifyMaxSize ∷ (Int → Int) → SpecWith a → SpecWith a

modifyMaxSuccess ∷ (Int → Int) → SpecWith a → SpecWith a

class Testable prop where

Minimal complete definition

property

Methods

property ∷ prop → Property

propertyForAllShrinkShow ∷ Gen a → (a → [a]) → (a → [String]) → (a → prop) → Property

data Property

allProperties ∷ Q Exp

forAllProperties ∷ Q Exp

monomorphic ∷ Name → ExpQ

polyQuickCheck ∷ Name → ExpQ

polyVerboseCheck ∷ Name → ExpQ

quickCheckAll ∷ Q Exp

verboseCheckAll ∷ Q Exp

(><) ∷ (Gen a → Gen a) → (Gen a → Gen a) → Gen a → Gen a

applyArbitrary2 ∷ (Arbitrary a, Arbitrary b) ⇒ (a → b → r) → Gen r

applyArbitrary3 ∷ (Arbitrary a, Arbitrary b, Arbitrary c) ⇒ (a → b → c → r) → Gen r

applyArbitrary4 ∷ (Arbitrary a, Arbitrary b, Arbitrary c, Arbitrary d) ⇒ (a → b → c → d → r) → Gen r

arbitrary1 ∷ (Arbitrary1 f, Arbitrary a) ⇒ Gen (f a)

arbitrary2 ∷ (Arbitrary2 f, Arbitrary a, Arbitrary b) ⇒ Gen (f a b)

arbitraryASCIIChar ∷ Gen Char

arbitraryBoundedEnum ∷ (Bounded a, Enum a) ⇒ Gen a

arbitraryBoundedIntegral ∷ (Bounded a, Integral a) ⇒ Gen a

arbitraryBoundedRandom ∷ (Bounded a, Random a) ⇒ Gen a

arbitraryPrintableChar ∷ Gen Char

arbitrarySizedBoundedIntegral ∷ (Bounded a, Integral a) ⇒ Gen a

arbitrarySizedFractional ∷ Fractional a ⇒ Gen a

arbitrarySizedIntegral ∷ Integral a ⇒ Gen a

arbitrarySizedNatural ∷ Integral a ⇒ Gen a

arbitraryUnicodeChar ∷ Gen Char

coarbitraryEnum ∷ Enum a ⇒ a → Gen b → Gen b

coarbitraryIntegral ∷ Integral a ⇒ a → Gen b → Gen b

coarbitraryReal ∷ Real a ⇒ a → Gen b → Gen b

coarbitraryShow ∷ Show a ⇒ a → Gen b → Gen b

genericCoarbitrary ∷ (Generic a, GCoArbitrary (Rep a)) ⇒ a → Gen b → Gen b

genericShrink ∷ (Generic a, RecursivelyShrink (Rep a), GSubterms (Rep a) a) ⇒ a → [a]

infiniteList ∷ Arbitrary a ⇒ Gen [a]

orderedList ∷ (Ord a, Arbitrary a) ⇒ Gen [a]

recursivelyShrink ∷ (Generic a, RecursivelyShrink (Rep a)) ⇒ a → [a]

shrink1 ∷ (Arbitrary1 f, Arbitrary a) ⇒ f a → [f a]

shrink2 ∷ (Arbitrary2 f, Arbitrary a, Arbitrary b) ⇒ f a b → [f a b]

shrinkDecimal ∷ RealFrac a ⇒ a → [a]

shrinkIntegral ∷ Integral a ⇒ a → [a]

shrinkList ∷ (a → [a]) → [a] → [[a]]

shrinkMap ∷ Arbitrary a ⇒ (a → b) → (b → a) → b → [b]

shrinkMapBy ∷ (a → b) → (b → a) → (a → [a]) → b → [b]

shrinkNothing ∷ a → [a]

shrinkRealFrac ∷ RealFrac a ⇒ a → [a]

subterms ∷ (Generic a, GSubterms (Rep a) a) ⇒ a → [a]

vector ∷ Arbitrary a ⇒ Int → Gen [a]

discard ∷ a

labelledExamples ∷ Testable prop ⇒ prop → IO ()

labelledExamplesResult ∷ Testable prop ⇒ prop → IO Result

labelledExamplesWith ∷ Testable prop ⇒ Args → prop → IO ()

labelledExamplesWithResult ∷ Testable prop ⇒ Args → prop → IO Result

pattern Fn ∷ (a → b) → Fun a b

pattern Fn2 ∷ (a → b → c) → Fun (a, b) c

pattern Fn3 ∷ (a → b → c → d) → Fun (a, b, c) d

applyFun ∷ Fun a b → a → b

applyFun2 ∷ Fun (a, b) c → a → b → c

applyFun3 ∷ Fun (a, b, c) d → a → b → c → d

functionBoundedEnum ∷ (Eq a, Bounded a, Enum a) ⇒ (a → b) → a :-> b

functionIntegral ∷ Integral a ⇒ (a → b) → a :-> b

functionMap ∷ Function b ⇒ (a → b) → (b → a) → (a → c) → a :-> c

functionRealFrac ∷ RealFrac a ⇒ (a → b) → a :-> b

functionShow ∷ (Show a, Read a) ⇒ (a → c) → a :-> c

functionVoid ∷ (∀ b. void → b) → void :-> c

choose ∷ Random a ⇒ (a, a) → Gen a

chooseAny ∷ Random a ⇒ Gen a

chooseBoundedIntegral ∷ (Bounded a, Integral a) ⇒ (a, a) → Gen a

chooseEnum ∷ Enum a ⇒ (a, a) → Gen a

chooseInt ∷ (Int, Int) → Gen Int

chooseInteger ∷ (Integer, Integer) → Gen Integer

elements ∷ [a] → Gen a

frequency ∷ [(Int, Gen a)] → Gen a

generate ∷ Gen a → IO a

getSize ∷ Gen Int

growingElements ∷ [a] → Gen a

infiniteListOf ∷ Gen a → Gen [a]

listOf ∷ Gen a → Gen [a]

listOf1 ∷ Gen a → Gen [a]

oneof ∷ [Gen a] → Gen a

resize ∷ Int → Gen a → Gen a

sample ∷ Show a ⇒ Gen a → IO ()

sample' ∷ Gen a → IO [a]

scale ∷ (Int → Int) → Gen a → Gen a

shuffle ∷ [a] → Gen [a]

sized ∷ (Int → Gen a) → Gen a

sublistOf ∷ [a] → Gen [a]

suchThat ∷ Gen a → (a → Bool) → Gen a

suchThatMap ∷ Gen a → (a → Maybe b) → Gen b

suchThatMaybe ∷ Gen a → (a → Bool) → Gen (Maybe a)

variant ∷ Integral n ⇒ n → Gen a → Gen a

vectorOf ∷ Int → Gen a → Gen [a]

(.&&.) ∷ (Testable prop1, Testable prop2) ⇒ prop1 → prop2 → Property

(.&.) ∷ (Testable prop1, Testable prop2) ⇒ prop1 → prop2 → Property

(.||.) ∷ (Testable prop1, Testable prop2) ⇒ prop1 → prop2 → Property

(=/=) ∷ (Eq a, Show a) ⇒ a → a → Property

(===) ∷ (Eq a, Show a) ⇒ a → a → Property

(==>) ∷ Testable prop ⇒ Bool → prop → Property

again ∷ Testable prop ⇒ prop → Property

checkCoverage ∷ Testable prop ⇒ prop → Property

checkCoverageWith ∷ Testable prop ⇒ Confidence → prop → Property

classify ∷ Testable prop ⇒ Bool → String → prop → Property

collect ∷ (Show a, Testable prop) ⇒ a → prop → Property

conjoin ∷ Testable prop ⇒ [prop] → Property

counterexample ∷ Testable prop ⇒ String → prop → Property

cover ∷ Testable prop ⇒ Double → Bool → String → prop → Property

coverTable ∷ Testable prop ⇒ String → [(String, Double)] → prop → Property

disjoin ∷ Testable prop ⇒ [prop] → Property

expectFailure ∷ Testable prop ⇒ prop → Property

forAll ∷ (Show a, Testable prop) ⇒ Gen a → (a → prop) → Property

forAllBlind ∷ Testable prop ⇒ Gen a → (a → prop) → Property

forAllShow ∷ Testable prop ⇒ Gen a → (a → String) → (a → prop) → Property

forAllShrink ∷ (Show a, Testable prop) ⇒ Gen a → (a → [a]) → (a → prop) → Property

forAllShrinkBlind ∷ Testable prop ⇒ Gen a → (a → [a]) → (a → prop) → Property

forAllShrinkShow ∷ Testable prop ⇒ Gen a → (a → [a]) → (a → String) → (a → prop) → Property

idempotentIOProperty ∷ Testable prop ⇒ IO prop → Property

ioProperty ∷ Testable prop ⇒ IO prop → Property

label ∷ Testable prop ⇒ String → prop → Property

mapSize ∷ Testable prop ⇒ (Int → Int) → prop → Property

noShrinking ∷ Testable prop ⇒ prop → Property

once ∷ Testable prop ⇒ prop → Property

printTestCase ∷ Testable prop ⇒ String → prop → Property

shrinking ∷ Testable prop ⇒ (a → [a]) → a → (a → prop) → Property

stdConfidence ∷ Confidence

tabulate ∷ Testable prop ⇒ String → [String] → prop → Property

total ∷ NFData a ⇒ a → Property

verbose ∷ Testable prop ⇒ prop → Property

verboseShrinking ∷ Testable prop ⇒ prop → Property

whenFail ∷ Testable prop ⇒ IO () → prop → Property

whenFail' ∷ Testable prop ⇒ IO () → prop → Property

withMaxSuccess ∷ Testable prop ⇒ Int → prop → Property

within ∷ Testable prop ⇒ Int → prop → Property

isSuccess ∷ Result → Bool

quickCheck ∷ Testable prop ⇒ prop → IO ()

quickCheckResult ∷ Testable prop ⇒ prop → IO Result

quickCheckWith ∷ Testable prop ⇒ Args → prop → IO ()

quickCheckWithResult ∷ Testable prop ⇒ Args → prop → IO Result

stdArgs ∷ Args

verboseCheck ∷ Testable prop ⇒ prop → IO ()

verboseCheckResult ∷ Testable prop ⇒ prop → IO Result

verboseCheckWith ∷ Testable prop ⇒ Args → prop → IO ()

verboseCheckWithResult ∷ Testable prop ⇒ Args → prop → IO Result

class Arbitrary1 (f ∷ Type → Type) where

Minimal complete definition

liftArbitrary

Methods

liftArbitrary ∷ Gen a → Gen (f a)

liftShrink ∷ (a → [a]) → f a → [f a]

class Arbitrary2 (f ∷ Type → Type → Type) where

Minimal complete definition

liftArbitrary2

Methods

liftArbitrary2 ∷ Gen a → Gen b → Gen (f a b)

liftShrink2 ∷ (a → [a]) → (b → [b]) → f a b → [f a b]

class CoArbitrary a where

Minimal complete definition

Nothing

Methods

coarbitrary ∷ a → Gen b → Gen b

data Fun a b

Constructors

Fun (a :-> b, b, Shrunk) (a → b)

class Function a where

Minimal complete definition

Nothing

Methods

function ∷ (a → b) → a :-> b

newtype ASCIIString

Constructors

ASCIIString

Fields

getASCIIString ∷ String

newtype Blind a

Constructors

Blind

Fields

getBlind ∷ a

newtype Fixed a

Constructors

Fixed

Fields

getFixed ∷ a

data InfiniteList a

Constructors

InfiniteList

Fields

getInfiniteList ∷ [a]

infiniteListInternalData ∷ InfiniteListInternalData a

newtype Large a

Constructors

Large

Fields

getLarge ∷ a

newtype Negative a

Constructors

Negative

Fields

getNegative ∷ a

newtype NonEmptyList a

Constructors

NonEmpty

Fields

getNonEmpty ∷ [a]

newtype NonNegative a

Constructors

NonNegative

Fields

getNonNegative ∷ a

newtype NonPositive a

Constructors

NonPositive

Fields

getNonPositive ∷ a

newtype NonZero a

Constructors

NonZero

Fields

getNonZero ∷ a

newtype OrderedList a

Constructors

Ordered

Fields

getOrdered ∷ [a]

newtype Positive a

Constructors

Positive

Fields

getPositive ∷ a

newtype PrintableString

Constructors

PrintableString

Fields

getPrintableString ∷ String

newtype Shrink2 a

Constructors

Shrink2

Fields

getShrink2 ∷ a

class ShrinkState s a where

Methods

shrinkInit ∷ a → s

shrinkState ∷ a → s → [(a, s)]

data Shrinking s a

Constructors

Shrinking s a

newtype Small a

Constructors

Small

Fields

getSmall ∷ a

data Smart a

Constructors

Smart Int a

newtype SortedList a

Constructors

Sorted

Fields

getSorted ∷ [a]

newtype UnicodeString

Constructors

UnicodeString

Fields

getUnicodeString ∷ String

data Discard

Constructors

Discard

data Confidence

Constructors

Confidence

Fields

certainty ∷ Integer

tolerance ∷ Double

data Args

Constructors

Args

Fields

replay ∷ Maybe (QCGen, Int)

maxSuccess ∷ Int

maxDiscardRatio ∷ Int

maxSize ∷ Int

chatty ∷ Bool

maxShrinks ∷ Int

data Result

Constructors

Success

Fields

numTests ∷ Int

numDiscarded ∷ Int

labels ∷ !(Map [String] Int)

classes ∷ !(Map String Int)

tables ∷ !(Map String (Map String Int))

output ∷ String

GaveUp

Fields

numTests ∷ Int

numDiscarded ∷ Int

labels ∷ !(Map [String] Int)

classes ∷ !(Map String Int)

tables ∷ !(Map String (Map String Int))

output ∷ String

Failure

Fields

numTests ∷ Int

numDiscarded ∷ Int

numShrinks ∷ Int

numShrinkTries ∷ Int

numShrinkFinal ∷ Int

usedSeed ∷ QCGen

usedSize ∷ Int

reason ∷ String

theException ∷ Maybe AnException

output ∷ String

failingTestCase ∷ [String]

failingLabels ∷ [String]

failingClasses ∷ Set String

NoExpectedFailure

Fields

numTests ∷ Int

numDiscarded ∷ Int

labels ∷ !(Map [String] Int)

classes ∷ !(Map String Int)

tables ∷ !(Map String (Map String Int))

output ∷ String

ledgerTestMain ∷ Spec → IO ()

ledgerTestMainWith ∷ Config → Spec → IO ()

ledgerHspecConfig ∷ Config