Test.Cardano.Ledger.Core.KeyPair

Documentation

class MakeCredential c r where

Methods

mkCredential ∷ c → Credential r

class MakeStakeReference c where

Minimal complete definition

Nothing

Methods

mkStakeRef ∷ c → StakeReference

default mkStakeRef ∷ MakeCredential c 'Staking ⇒ c → StakeReference

mkAddr ∷ (MakeCredential p 'Payment, MakeStakeReference s) ⇒ p → s → Addr

Construct a Testnet address from payment and staking components

mkScriptAddr ∷ ScriptHash → KeyPair 'Staking → Addr

Deprecated: In favor of mkAddr

mkCred ∷ KeyPair kr → Credential kr

Deprecated: In favor of mkCredential

data KeyPair (kd ∷ KeyRole)

Constructors

KeyPair

Fields

vKey ∷ !(VKey kd)

sKey ∷ !(SignKeyDSIGN DSIGN)

type KeyPairs = [(KeyPair 'Payment, KeyPair 'Staking)]

Representation of a list of pairs of key pairs, e.g., pay and stake keys

mkWitnessVKey ∷ SafeHash EraIndependentTxBody → KeyPair kr → WitVKey 'Witness

Create a witness for transaction

mkWitnessesVKey ∷ SafeHash EraIndependentTxBody → [KeyPair kr] → Set (WitVKey 'Witness)

Create witnesses for transaction

makeWitnessesFromScriptKeys ∷ SafeHash EraIndependentTxBody → Map (KeyHash kr) (KeyPair kr) → Set (KeyHash kr) → Set (WitVKey 'Witness)

From a list of key pairs and a set of key hashes required for a multi-sig scripts, return the set of required keys.

mkKeyHashWitFunPair ∷ ∀ kr. Gen (KeyHash kr, SafeHash EraIndependentTxBody → WitVKey 'Witness)

When wrting a test which needs a KeyHash, and you will also later need a witness for that function, use keyHashWitFunPair. Since one cannot make a witness until one has SafeHash of the TxBody that the KeyHash is embedded. The second part of the pair is a (SafeHash to WitVKey) function. Use it something like this do (key,witfun) <- keyHashWitFunPair txbody <- ... key ... let safehash = hashAnnotated txbody tx = ... txbody ... (witfun safehash) ...

mkVKeyRewardAccount ∷ Network → KeyPair 'Staking → RewardAccount

mkKeyPair ∷ Int → KeyPair r

mkKeyPairWithSeed ∷ ByteString → KeyPair r

mkKeyHash ∷ Int → KeyHash kd

data ByronKeyPair

Constructors

ByronKeyPair

Fields

bkpVerificationKey ∷ !VerificationKey

bkpSigningKey ∷ !SigningKey

mkBootKeyPairWithSeed ∷ ByteString → ByronKeyPair

genByronVKeyAddr ∷ Gen (VerificationKey, Address)

genByronAddrFromVKey ∷ VerificationKey → Gen Address