Test.Cardano.Ledger.Era

Documentation

class (EraTx era, EraTxOut era, EraTxBody era, EraTxAuxData era, EraTxWits era, EraScript era, EraPParams era, EraBlockBody era, EraTxCert era, EraCertState era, EraGov era, EraStake era, EraUTxO era, EraAccounts era, EraGenesis era, Arbitrary (Tx era), Arbitrary (TxOut era), Arbitrary (TxBody era), Arbitrary (TxAuxData era), Arbitrary (TxWits era), Arbitrary (Script era), Arbitrary (PParamsHKD Identity era), Arbitrary (PParamsHKD StrictMaybe era), Arbitrary (TxCert era), Arbitrary (Value era), Arbitrary (CompactForm (Value era)), Arbitrary (CertState era), Arbitrary (GovState era), Arbitrary (InstantStake era), Arbitrary (Accounts era), Arbitrary (AccountState era), ToExpr (Tx era), ToExpr (TxOut era), ToExpr (TxBody era), ToExpr (TxAuxData era), ToExpr (TxWits era), ToExpr (Script era), ToExpr (PParamsHKD Identity era), ToExpr (PParamsHKD StrictMaybe era), ToExpr (TxCert era), ToExpr (Value era), ToExpr (CompactForm (Value era)), ToExpr (CertState era), ToExpr (GovState era), ToExpr (InstantStake era), ToExpr (Accounts era), ToExpr (AccountState era), DecCBOR (Script era), DecCBOR (NativeScript era), DecCBOR (TxAuxData era), DecCBOR (TxWits era), DecCBOR (TxBody era), DecCBOR (Tx era), Eq (TranslationContext era), Show (TranslationContext era), Typeable (TranslationContext era), ToJSON (TranslationContext era), FromJSON (TranslationContext era), Arbitrary (TranslationContext era)) ⇒ EraTest era where

Methods

zeroCostModels ∷ CostModels

mkTestAccountState ∷ HasCallStack ⇒ Maybe Ptr → CompactForm Coin → Maybe (KeyHash 'StakePool) → Maybe DRep → AccountState era

This is a helper function that allows for creation of an EraAccounts in era agnostic fashion. There is no equivalent function outside of testing since arguments required for creation of EraAccounts varies between eras and we can get away with such function in testing because we allow for such function to be partial.

accountsFromAccountsMap ∷ Map (Credential 'Staking) (AccountState era) → Accounts era

registerTestAccount ∷ (HasCallStack, EraTest era) ⇒ Credential 'Staking → Maybe Ptr → CompactForm Coin → Maybe (KeyHash 'StakePool) → Maybe DRep → Accounts era → Accounts era

This is a helper function that uses mkTestAccountState to register an account.