Test.Cardano.Ledger.Plutus

Documentation

data PlutusArgs

Constructors

PlutusArgs

Fields

paData ∷ Data

paSpendDatum ∷ Maybe Data

data ScriptTestContext

Constructors

∀ l.PlutusLanguage l ⇒ ScriptTestContext

Fields

stcScript ∷ Plutus l

stcArgs ∷ PlutusArgs

Plutus

alwaysSucceedsPlutus ∷ Natural → Plutus l

alwaysFailsPlutus ∷ Natural → Plutus l

CostModel

mkCostModelConst ∷ HasCallStack ⇒ Language → Int64 → CostModel

Construct a test cost model where all parameters are set to the same value

zeroTestingCostModel ∷ HasCallStack ⇒ Language → CostModel

zeroTestingCostModelV1 ∷ HasCallStack ⇒ CostModel

zeroTestingCostModelV2 ∷ HasCallStack ⇒ CostModel

zeroTestingCostModelV3 ∷ HasCallStack ⇒ CostModel

testingCostModel ∷ HasCallStack ⇒ Language → CostModel

testingCostModelV1 ∷ HasCallStack ⇒ CostModel

testingCostModelV2 ∷ HasCallStack ⇒ CostModel

testingCostModelV3 ∷ HasCallStack ⇒ CostModel

testingEvaluationContext ∷ Language → EvaluationContext

CostModels

testingCostModels ∷ HasCallStack ⇒ [Language] → CostModels

Test CostModels for all available languages

zeroTestingCostModels ∷ HasCallStack ⇒ [Language] → CostModels

Test CostModels for all available languages with zero values for all parameters