Test.Cardano.Ledger.Plutus.Examples

Description

This file is generated by "plutus-preprocessor:plutus-preprocessor"

Documentation

alwaysSucceedsNoDatum ∷ SLanguage l → Plutus l

Script that always succeeds, unless arguments are malformed or context contains a datum

alwaysSucceedsWithDatum ∷ SLanguage l → Plutus l

Script that always succeeds, unless arguments are malformed or context does not contain a datum

alwaysFailsNoDatum ∷ SLanguage l → Plutus l

Script that always fails, unless arguments are malformed or context contains a datum

alwaysFailsWithDatum ∷ SLanguage l → Plutus l

Script that always fails, unless arguments are malformed or context does not contain a datum

redeemerSameAsDatum ∷ SLanguage l → Plutus l

Script that succeeds whenever redeemer equals to the datum Fails on malformed arguments

evenDatum ∷ SLanguage l → Plutus l

Script that succeeds whenever Integer datum is supplied and it's value is even. Fails on malformed arguments

evenRedeemerNoDatum ∷ SLanguage l → Plutus l

Script that succeeds whenever Integer redeemer is supplied and it's value is even Fails on malformed arguments or whenever datum is present in the context

evenRedeemerWithDatum ∷ SLanguage l → Plutus l

Script that succeeds whenever Integer redeemer is supplied and it's value is even Fails on malformed arguments or whenever datum is missing from the context

purposeIsWellformedNoDatum ∷ SLanguage l → Plutus l

Script that succeeds when datum is not expected and purpose arguments are validated against txInfo Fails on malformed arguments

purposeIsWellformedWithDatum ∷ SLanguage l → Plutus l

Script that succeeds when datum is expected and purpose arguments are validated against txInfo Fails on malformed arguments

datumIsWellformed ∷ SLanguage l → Plutus l

Script that succeeds when datum is expected and datum is validated against txInfo Fails on malformed arguments

inputsOutputsAreNotEmptyNoDatum ∷ SLanguage l → Plutus l

Script that succeeds when inputs and outputs are not empty validated against txInfo Fails on malformed arguments and also if inputs or outputs are empty

inputsOutputsAreNotEmptyWithDatum ∷ SLanguage l → Plutus l

Script that succeeds when inputs and outputs are not empty validated against txInfo Fails on malformed arguments and also if inputs or outputs are empty