Test.Cardano.Ledger.Plutus.ToPlutusData

Documentation

roundTripPlutusDataSpec ∷ ∀ x. (HasCallStack, Typeable x, Show x, Eq x, Arbitrary x, ToPlutusData x) ⇒ Spec

Test the roundtrip property via fromPlutusData and toPlutusData for any type meeting all the constraints below.