Cardano.Ledger.Address

Documentation

mkVKeyRwdAcnt :: Crypto crypto => Network -> KeyPair 'Staking crypto -> RewardAcnt crypto

mkRwdAcnt :: Network -> Credential 'Staking crypto -> RewardAcnt crypto

toAddr :: Crypto crypto => Network -> (KeyPair 'Payment crypto, KeyPair 'Staking crypto) -> Addr crypto

toCred :: Crypto crypto => KeyPair kr crypto -> Credential kr crypto

serialiseAddr :: Addr crypto -> ByteString

Serialise an address to the external format.

deserialiseAddr :: Crypto crypto => ByteString -> Maybe (Addr crypto)

Deserialise an address from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

deserialiseAddrStakeRef :: Crypto crypto => ByteString -> Maybe (StakeReference crypto)

Deserialise a stake refence from a address. This will fail if this is a Bootstrap address (or malformed).

data Addr crypto

An address for UTxO.

Contents of Addr data type are intentionally left as lazy, otherwise operating on compact form of an address will result in redundant work.

Constructors

Addr Network (PaymentCredential crypto) (StakeReference crypto)

AddrBootstrap (BootstrapAddress crypto)

newtype BootstrapAddress crypto

Constructors

BootstrapAddress

Fields

unBootstrapAddress :: Address

bootstrapAddressAttrsSize :: BootstrapAddress crypto -> Int

The size of the extra attributes in a bootstrp (ie Byron) address. Used to help enforce that people do not post huge ones on the chain.

isBootstrapRedeemer :: BootstrapAddress crypto -> Bool

Return True if a given address is a redeemer address from the Byron Era

getNetwork :: Addr crypto -> Network

data RewardAcnt crypto

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork :: !Network

getRwdCred :: !(Credential 'Staking crypto)

serialiseRewardAcnt :: RewardAcnt crypto -> ByteString

Serialise a reward account to the external format.

deserialiseRewardAcnt :: Crypto crypto => ByteString -> Maybe (RewardAcnt crypto)

Deserialise an reward account from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

byron :: Int

notBaseAddr :: Int

isEnterpriseAddr :: Int

stakeCredIsScript :: Int

getAddr :: Crypto crypto => Get (Addr crypto)

getKeyHash :: Crypto crypto => Get (Credential kr crypto)

bootstrapKeyHash :: forall crypto. Crypto crypto => BootstrapAddress crypto -> KeyHash 'Payment crypto

getPtr :: Get Ptr

getRewardAcnt :: Crypto crypto => Get (RewardAcnt crypto)

getScriptHash :: Crypto crypto => Get (Credential kr crypto)

getVariableLengthWord64 :: Get Word64

payCredIsScript :: Int

putAddr :: Addr crypto -> Put

putCredential :: Credential kr crypto -> Put

putPtr :: Ptr -> Put

putRewardAcnt :: RewardAcnt crypto -> Put

putVariableLengthWord64 :: Word64 -> Put

word64ToWord7s :: Word64 -> [Word7]

word7sToWord64 :: [Word7] -> Word64

newtype Word7

Constructors

Word7 Word8

toWord7 :: Word8 -> Word7