Cardano.Ledger.Address

Documentation

mkRwdAcnt ∷ Network → Credential 'Staking c → RewardAcnt c

serialiseAddr ∷ Addr c → ByteString

Serialise an address to the external format.

deserialiseAddr ∷ Crypto c ⇒ ByteString → Maybe (Addr c)

Deserialise an address from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

data Addr c

An address for UTxO.

Contents of Addr data type are intentionally left as lazy, otherwise operating on compact form of an address will result in redundant work.

Constructors

Addr Network (PaymentCredential c) (StakeReference c)

AddrBootstrap (BootstrapAddress c)

newtype BootstrapAddress c

Constructors

BootstrapAddress

Fields

unBootstrapAddress ∷ Address

bootstrapAddressAttrsSize ∷ BootstrapAddress c → Int

The size of the extra attributes in a bootstrp (ie Byron) address. Used to help enforce that people do not post huge ones on the chain.

isBootstrapRedeemer ∷ BootstrapAddress c → Bool

Return True if a given address is a redeemer address from the Byron Era

getNetwork ∷ Addr c → Network

data RewardAcnt c

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork ∷ !Network

getRwdCred ∷ !(Credential 'Staking c)

serialiseRewardAcnt ∷ RewardAcnt c → ByteString

Serialise a reward account to the external format.

deserialiseRewardAcnt ∷ Crypto c ⇒ ByteString → Maybe (RewardAcnt c)

Deserialise a reward account from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

bootstrapKeyHash ∷ ∀ c. Crypto c ⇒ BootstrapAddress c → KeyHash 'Payment c

putAddr ∷ Addr c → Put

putCredential ∷ Credential kr c → Put

putPtr ∷ Ptr → Put

putRewardAcnt ∷ RewardAcnt c → Put

putVariableLengthWord64 ∷ Word64 → Put

newtype Word7

Constructors

Word7 Word8

toWord7 ∷ Word8 → Word7

Compact Address

fromBoostrapCompactAddress ∷ CompactAddress → CompactAddr c

Convert Byron's comapct address into CompactAddr. This is just an efficient type cast.

compactAddr ∷ Addr c → CompactAddr c

decompactAddr ∷ ∀ c. (HasCallStack, Crypto c) ⇒ CompactAddr c → Addr c

data CompactAddr c

unCompactAddr ∷ CompactAddr c → ShortByteString

Unwrap the compact address and get to the address' binary representation.

isPayCredScriptCompactAddr ∷ CompactAddr c → Bool

Efficiently check whether compated adddress is an address with a credential that is a payment script.

isBootstrapCompactAddr ∷ CompactAddr c → Bool

Efficiently check whether compated adddress is a Byron address.

decodeAddr ∷ ∀ c m. (Crypto c, MonadFail m) ⇒ ByteString → m (Addr c)

decodeAddrShort ∷ ∀ c m. (Crypto c, MonadFail m) ⇒ ShortByteString → m (Addr c)

decodeAddrEither ∷ ∀ c. Crypto c ⇒ ByteString → Either String (Addr c)

decodeAddrShortEither ∷ ∀ c. Crypto c ⇒ ShortByteString → Either String (Addr c)

fromCborAddr ∷ ∀ c s. Crypto c ⇒ Decoder s (Addr c)

Decoder for an Addr. Works in all eras

fromCborBothAddr ∷ ∀ c s. Crypto c ⇒ Decoder s (Addr c, CompactAddr c)

This is the decoder for an address that returns both the actual Addr and the bytes, that it was encoded as.

fromCborCompactAddr ∷ ∀ c s. Crypto c ⇒ Decoder s (CompactAddr c)

Returns the actual bytes that represent an addres, while ensuring that they can be decoded in any era as an Addr when need be.

fromCborBackwardsBothAddr ∷ ∀ c s. Crypto c ⇒ Decoder s (Addr c, CompactAddr c)

Prior to Babbage era we did not check if a binary blob representing an address was fully consumed, so unfortunately we must preserve this behavior. However, we do not need to preserve the unconsumed bytes in memory, therefore we can to drop the garbage after we successfully decoded the malformed address. We also need to allow bogus pointer address to be deserializeable prior to Babbage era.

decodeRewardAcnt ∷ ∀ c b m. (Crypto c, AddressBuffer b, MonadFail m) ⇒ b → m (RewardAcnt c)

fromCborRewardAcnt ∷ ∀ c s. Crypto c ⇒ Decoder s (RewardAcnt c)

newtype Withdrawals c

This is called wdrl in the spec.

Constructors

Withdrawals

Fields

unWithdrawals ∷ Map (RewardAcnt c) Coin