Cardano.Ledger.BaseTypes

Documentation

newtype BlockNo

The 0-based index of the block in the blockchain. BlockNo is <= SlotNo and is only equal at slot N if there is a block for every slot where N <= SlotNo.

Constructors

BlockNo

Fields

unBlockNo ∷ Word64

newtype EpochInterval

Constructors

EpochInterval

Fields

unEpochInterval ∷ Word32

newtype EpochSize

Constructors

EpochSize

Fields

unEpochSize ∷ Word64

newtype EpochNo

An epoch, i.e. the number of the epoch.

Constructors

EpochNo

Fields

unEpochNo ∷ Word64

data WithOrigin t

Constructors

Origin

At !t

newtype SlotNo

The 0-based index for the Ourboros time slot.

Constructors

SlotNo

Fields

unSlotNo ∷ Word64

binOpEpochNo ∷ (Word64 → Word64 → Word64) → EpochNo → EpochNo → EpochNo

Convenience function for doing binary operations on two EpochNos

addEpochInterval ∷ EpochNo → EpochInterval → EpochNo

Add a EpochInterval (a positive change) to an EpochNo to get a new EpochNo

module Cardano.Ledger.BaseTypes.NonZero

data ProtVer

Constructors

ProtVer

Fields

pvMajor ∷ !Version

pvMinor ∷ !Natural

module Cardano.Ledger.Binary.Version

type FixedPoint = Digits34

(==>) ∷ Bool → Bool → Bool infix 1

(⭒) ∷ Nonce → Nonce → Nonce

Evolve the nonce

data Network

Constructors

Testnet

Mainnet

networkToWord8 ∷ Network → Word8

word8ToNetwork ∷ Word8 → Maybe Network

data Nonce

Evolving nonce type.

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce
Identity element

newtype Seed

Seed to the verifiable random function.

Constructors

Seed (Hash Blake2b_256 Seed)

data UnitInterval

Type to represent a value in the unit interval [0; 1]

data PositiveUnitInterval

Type to represent a value in the unit interval (0; 1]

data PositiveInterval

Type to represent a value in the interval (0; +∞)

data NonNegativeInterval

Type to represent a value in the interval [0; +∞)

class Bounded r ⇒ BoundedRational r where

Type clases that allows conversion between Rational and some form of bounded rational type. Bounds can be restricted by both the Bounded type class and underlyng representation.

maybe True (\br -> minBound <= br && br <= maxBound) . boundRational

Roundtrip properties must hold:

\r -> maybe True ((r ==) . unboundRational) (boundRational r)
\br -> Just br == boundRational (unboundRational br)

Methods

boundRational ∷ Rational → Maybe r

Returns Nothing when supplied value is not within bounds or when precision is too high to be represented by the underlying type

Example

Expand

>>> :set -XTypeApplications
>>> import Data.Ratio
>>> boundRational @UnitInterval $ 2 % 3
Just (2 % 3)
>>> boundRational @UnitInterval (-0.5)
Nothing
>>> boundRational @UnitInterval (1.5)
Nothing
>>> boundRational @UnitInterval 0
Just (0 % 1)
>>> boundRational @PositiveUnitInterval 0
Nothing

unboundRational ∷ r → Rational

Promote bounded rational type into the unbounded Rational.

fpPrecision ∷ FixedPoint

integralToBounded ∷ ∀ i b m. (Integral i, Integral b, Bounded b, MonadFail m) ⇒ i → m b

promoteRatio ∷ Integral a ⇒ Ratio a → Rational

invalidKey ∷ (Typeable a, MonadFail m) ⇒ Word → m a

Report an error when a numeric key of the type constructor doesn't match.

mkNonceFromOutputVRF ∷ OutputVRF v → Nonce

Make a nonce from the VRF output bytes

mkNonceFromNumber ∷ Word64 → Nonce

Make a nonce from a number.

data Url

urlToText ∷ Url → Text

textToUrl ∷ MonadFail m ⇒ Int → Text → m Url

Turn a Text into a Url, fail if the Text has more than n Bytes

data DnsName

dnsToText ∷ DnsName → Text

textToDns ∷ MonadFail m ⇒ Int → Text → m DnsName

Turn a Text into a DnsName, fail if the Text has more than n Bytes

newtype Port

Constructors

Port

Fields

portToWord16 ∷ Word16

data ActiveSlotCoeff

mkActiveSlotCoeff ∷ PositiveUnitInterval → ActiveSlotCoeff

activeSlotVal ∷ ActiveSlotCoeff → PositiveUnitInterval

activeSlotLog ∷ ActiveSlotCoeff → FixedPoint

module Data.Maybe.Strict

newtype BlocksMade

Number of blocks which have been created by stake pools in the current epoch.

Constructors

BlocksMade

Fields

unBlocksMade ∷ Map (KeyHash 'StakePool) Natural

kindObject ∷ Text → [Pair] → Value

Helper function for a common pattern of creating objects

Indices

newtype TxIx

Transaction index.

Constructors

TxIx

Fields

unTxIx ∷ Word16

txIxToInt ∷ TxIx → Int

txIxFromIntegral ∷ ∀ a m. (Integral a, MonadFail m) ⇒ a → m TxIx

mkTxIx ∷ Word16 → TxIx

Deprecated: In favor of TxIx

Construct a TxIx from a 16 bit unsigned integer

mkTxIxPartial ∷ HasCallStack ⇒ Integer → TxIx

Construct a TxIx from an arbitrary precision Integer. Throws an error for values out of range. Make sure to use it only for testing.

newtype CertIx

Certificate index. There is mkCertIxPartial that can be used for testing when constructing from other integral types that are larger than Word16

Constructors

CertIx

Fields

unCertIx ∷ Word16

certIxToInt ∷ CertIx → Int

certIxFromIntegral ∷ ∀ a m. (Integral a, MonadFail m) ⇒ a → m CertIx

mkCertIx ∷ Word16 → CertIx

Deprecated: In favor of CertIx

Construct a CertIx from a 16 bit unsigned integer

mkCertIxPartial ∷ HasCallStack ⇒ Integer → CertIx

Construct a CertIx from an arbitrary precision Integer. Throws an error for values out of range. Make sure to use it only for testing.

data Anchor

Constructors

Anchor

Fields

anchorUrl ∷ !Url

anchorDataHash ∷ !(SafeHash AnchorData)

newtype AnchorData

Constructors

AnchorData ByteString

hashAnchorData ∷ AnchorData → SafeHash AnchorData

Deprecated: In favor of hashAnnotated

Hash AnchorData

STS Base

data Globals

Constructors

Globals

Fields

epochInfo ∷ !(EpochInfo (Either Text))

slotsPerKESPeriod ∷ !Word64

stabilityWindow ∷ !Word64

The window size in which our chosen chain growth property guarantees at least k blocks. From the paper "Ouroboros praos: An adaptively-secure, semi-synchronous proof-of-stake protocol". The stabilityWindow constant is used in a number of places; for example, protocol updates must be submitted at least twice this many slots before an epoch boundary.

randomnessStabilisationWindow ∷ !Word64

Number of slots before the end of the epoch at which we stop updating the candidate nonce for the next epoch.

securityParameter ∷ !(NonZero Word64)

Maximum number of blocks we are allowed to roll back

maxKESEvo ∷ !Word64

Maximum number of KES iterations

quorum ∷ !Word64

Quorum for update system votes and MIR certificates

maxLovelaceSupply ∷ !Word64

Maximum number of lovelace in the system

activeSlotCoeff ∷ !ActiveSlotCoeff

Active Slot Coefficient, named f in "Ouroboros Praos: An adaptively-secure, semi-synchronous proof-of-stake protocol"

networkId ∷ !Network

The network ID

systemStart ∷ !SystemStart

System start time

epochInfoPure ∷ Globals → EpochInfo Identity

Pure epoch info via throw. Note that this should only be used when we can guarantee the validity of the translation; in particular, the EpochInfo used here should never be applied to user-supplied input.

type ShelleyBase = ReaderT Globals Identity

data Relation

Relationship descriptor for the expectation in the Mismatch type.

Constructors

RelEQ
Equal

RelLT
Less then

RelGT
Greater then

RelLTEQ
Less then or equal

RelGTEQ
Greater then or equal

RelSubset
Is subset of

data Mismatch (r ∷ Relation) a

This is intended to help clarify supplied and expected values reported by predicate-failures in all eras.

Constructors

Mismatch

Fields

mismatchSupplied ∷ !a

mismatchExpected ∷ !a

swapMismatch ∷ Mismatch r a → (a, a)

Convert a Mismatch to a tuple that has "supplied" and "expected" swapped places

unswapMismatch ∷ (a, a) → Mismatch r a

Convert a tuple that has "supplied" and "expected" swapped places to a Mismatch type.

Injection

class Inject t s where

Methods

inject ∷ t → s

positiveUnitIntervalNonZeroRational ∷ PositiveUnitInterval → NonZero Rational

Aeson helpers

newtype KeyValuePairs a

Constructors

KeyValuePairs

Fields

unKeyValuePairs ∷ a

class ToKeyValuePairs a where

Methods

toKeyValuePairs ∷ KeyValue e kv ⇒ a → [kv]