Cardano.Ledger.BaseTypes

Documentation

data ProtVer

Constructors

ProtVer

Fields

pvMajor :: !Natural

pvMinor :: !Natural

type FixedPoint = Digits34

(==>) :: Bool -> Bool -> Bool infix 1

(⭒) :: Nonce -> Nonce -> Nonce

Evolve the nonce

data Network

Constructors

Testnet

Mainnet

networkToWord8 :: Network -> Word8

word8ToNetwork :: Word8 -> Maybe Network

data Nonce

Evolving nonce type.

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce
Identity element

newtype Seed

Seed to the verifiable random function.

We do not expose the constructor to Seed. Instead, a Seed should be created using mkSeed for a VRF calculation.

Constructors

Seed (Hash Blake2b_256 Seed)

data UnitInterval

Type to represent a value in the unit interval [0; 1]

data PositiveUnitInterval

Type to represent a value in the unit interval (0; 1]

data PositiveInterval

Type to represent a value in the interval (0; +∞)

data NonNegativeInterval

Type to represent a value in the interval [0; +∞)

class Bounded r => BoundedRational r where

Type clases that allows conversion between Rational and some form of bounded rational type. Bounds can be restricted by both the Bounded type class and underlyng representation.

maybe True (\br -> minBound <= br && br <= maxBound) . boundRational

Roundtrip properties must hold:

\r -> maybe True ((r ==) . unboundRational) (boundRational r)
\br -> Just br == boundRational (unboundRational br)

Methods

boundRational :: Rational -> Maybe r

Returns Nothing when supplied value is not within bounds or when precision is too high to be represented by the underlying type

Example

Expand

>>> :set -XTypeApplications
>>> import Data.Ratio
>>> boundRational @UnitInterval $ 2 % 3
Just (2 % 3)
>>> boundRational @UnitInterval (-0.5)
Nothing
>>> boundRational @UnitInterval (1.5)
Nothing
>>> boundRational @UnitInterval 0
Just (0 % 1)
>>> boundRational @PositiveUnitInterval 0
Nothing

unboundRational :: r -> Rational

Promote bounded rational type into the unbounded Rational.

boundedRationalFromCBOR :: BoundedRational r => Decoder s r

Deserialize BoundedRational type using Rational deserialization and fail when bounds are violated.

boundedRationalToCBOR :: BoundedRational r => r -> Encoding

Serialize BoundedRational type in the same way Rational is serialized.

fpPrecision :: FixedPoint

promoteRatio :: Integral a => Ratio a -> Rational

invalidKey :: Word -> Decoder s a

mkNonceFromOutputVRF :: OutputVRF v -> Nonce

Make a nonce from the VRF output bytes

mkNonceFromNumber :: Word64 -> Nonce

Make a nonce from a number.

data Url

urlToText :: Url -> Text

textToUrl :: Text -> Maybe Url

data DnsName

dnsToText :: DnsName -> Text

textToDns :: Text -> Maybe DnsName

newtype Port

Constructors

Port

Fields

portToWord16 :: Word16

data ActiveSlotCoeff

mkActiveSlotCoeff :: PositiveUnitInterval -> ActiveSlotCoeff

activeSlotVal :: ActiveSlotCoeff -> PositiveUnitInterval

activeSlotLog :: ActiveSlotCoeff -> FixedPoint

module Data.Maybe.Strict

newtype BlocksMade crypto

Blocks made

Constructors

BlocksMade

Fields

unBlocksMade :: Map (KeyHash 'StakePool crypto) Natural

Indices

newtype TxIx

Transaction index.

Constructors

TxIx Word64

txIxToInt :: TxIx -> Int

txIxFromIntegral :: Integral a => a -> Maybe TxIx

mkTxIxPartial :: HasCallStack => Integer -> TxIx

Construct a TxIx from an arbitrary precision Integer. Throws an error for values out of range. Make sure to use it only for testing.

newtype CertIx

Certificate index. Use certIxFromIntegral in order to construct this index safely from anything other than Word16. There is also mkCertIxPartial that can be used for testing.

Constructors

CertIx Word64

certIxToInt :: CertIx -> Int

certIxFromIntegral :: Integral a => a -> Maybe CertIx

mkCertIxPartial :: HasCallStack => Integer -> CertIx

Construct a CertIx from an arbitrary precision Integer. Throws an error for values out of range. Make sure to use it only for testing.

STS Base

data Globals

Constructors

Globals

Fields

epochInfo :: !(EpochInfo (Either Text))

slotsPerKESPeriod :: !Word64

stabilityWindow :: !Word64

The window size in which our chosen chain growth property guarantees at least k blocks. From the paper "Ouroboros praos: An adaptively-secure, semi-synchronous proof-of-stake protocol". The stabilityWindow constant is used in a number of places; for example, protocol updates must be submitted at least twice this many slots before an epoch boundary.

randomnessStabilisationWindow :: !Word64

Number of slots before the end of the epoch at which we stop updating the candidate nonce for the next epoch.

securityParameter :: !Word64

Maximum number of blocks we are allowed to roll back

maxKESEvo :: !Word64

Maximum number of KES iterations

quorum :: !Word64

Quorum for update system votes and MIR certificates

maxMajorPV :: !Natural

All blocks invalid after this protocol version

maxLovelaceSupply :: !Word64

Maximum number of lovelace in the system

activeSlotCoeff :: !ActiveSlotCoeff

Active Slot Coefficient, named f in "Ouroboros Praos: An adaptively-secure, semi-synchronous proof-of-stake protocol"

networkId :: !Network

The network ID

systemStart :: !SystemStart

System start time

epochInfoPure :: Globals -> EpochInfo Identity

Pure epoch info via throw. Note that this should only be used when we can guarantee the validity of the translation; in particular, the EpochInfo used here should never be applied to user-supplied input.

type ShelleyBase = ReaderT Globals Identity