Cardano.Ledger.Block

Documentation

data Block h era

Constructors

Block !h !(TxSeq era)

bheader ∷ Block h era → h

bbody ∷ Block h era → TxSeq era

neededTxInsForBlock ∷ ∀ h era. EraSegWits era ⇒ Block h era → Set TxIn

The validity of any individual block depends only on a subset of the UTxO stored in the ledger state. This function returns the transaction inputs corresponding to the required UTxO for a given Block.

This function will be used by the consensus layer to enable storing the UTxO on disk. In particular, given a block, the consensus layer will use neededTxInsForBlock to retrieve the needed UTxO from disk and present only those to the ledger.