Cardano.Ledger.Block

Documentation

data Block h era

Constructors

Block

Fields

blockHeader ∷ !h

blockBody ∷ !(BlockBody era)

bheader ∷ Block h era → h

Deprecated: In favor of blockHeader

bbody ∷ Block h era → BlockBody era

Deprecated: In favor of blockBody

neededTxInsForBlock ∷ ∀ h era. EraBlockBody era ⇒ Block h era → Set TxIn

The validity of any individual block depends only on a subset of the UTxO stored in the ledger state. This function returns the transaction inputs corresponding to the required UTxO for a given Block.

This function will be used by the consensus layer to enable storing the UTxO on disk. In particular, given a block, the consensus layer will use neededTxInsForBlock to retrieve the needed UTxO from disk and present only those to the ledger.