Cardano.Ledger.CertState

Documentation

data CertState era

The state associated with the DELPL rule, which combines the DELEG rule and the POOL rule.

Constructors

CertState

Fields

certVState ∷ !(VState era)

certPState ∷ !(PState era)

certDState ∷ !(DState era)

data CommitteeAuthorization c

Constructors

CommitteeHotCredential !(Credential 'HotCommitteeRole c)
Member authorized with a Hot credential acting on behalf of their Cold credential

CommitteeMemberResigned !(StrictMaybe (Anchor c))
Member resigned with a potential explanation in Anchor

data DState era

The state used by the DELEG rule, which roughly tracks stake delegation and some governance features.

Constructors

DState

Fields

dsUnified ∷ !(UMap (EraCrypto era))

Unified Reward Maps. This contains the reward map (which is the source of truth regarding the registered stake credentials, the deposit map, the delegation map, and the stake credential pointer map.

dsFutureGenDelegs ∷ !(Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

Future genesis key delegations

dsGenDelegs ∷ !(GenDelegs (EraCrypto era))

Genesis key delegations

dsIRewards ∷ !(InstantaneousRewards (EraCrypto era))

Instantaneous Rewards

data PState era

The state used by the POOL rule, which tracks stake pool information.

Constructors

PState

Fields

psStakePoolParams ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

The stake pool parameters.

psFutureStakePoolParams ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

The future stake pool parameters. Changes to existing stake pool parameters are staged in order to give delegators time to react to changes. See section 11.2, "Example Illustration of the Reward Cycle", of the Shelley Ledger Specification for a sequence diagram.

psRetiring ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

A map of retiring stake pools to the epoch when they retire.

psDeposits ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) Coin)

A map of the deposits for each pool

data VState era

The state that tracks the voting entities (DReps and Constitutional Committee members). In the formal ledger specification this type is called GState

Constructors

VState

Fields

vsDReps ∷ !(Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

vsCommitteeState ∷ !(CommitteeState era)

vsNumDormantEpochs ∷ !EpochNo

Number of contiguous epochs in which there are exactly zero active governance proposals to vote on. It is incremented in every EPOCH rule if the number of active governance proposals to vote on continues to be zero. It is reset to zero when a new governance action is successfully proposed. We need this counter in order to bump DRep expiries through dormant periods when DReps do not have an opportunity to vote on anything.

data InstantaneousRewards c

InstantaneousRewards captures the pending changes to the ledger state caused by MIR certificates. It consists of two mappings, the rewards which will be paid out from the reserves and the rewards which will be paid out from the treasury. It also consists of two coin values which represent the transfer of coins from one pot to the other pot. NOTE that the following property should always hold: deltaReserves + deltaTreasury = 0

Constructors

InstantaneousRewards

Fields

iRReserves ∷ !(Map (Credential 'Staking c) Coin)

iRTreasury ∷ !(Map (Credential 'Staking c) Coin)

deltaReserves ∷ !DeltaCoin

deltaTreasury ∷ !DeltaCoin

data FutureGenDeleg c

Constructors

FutureGenDeleg

Fields

fGenDelegSlot ∷ !SlotNo

fGenDelegGenKeyHash ∷ !(KeyHash 'Genesis c)

data Anchor c

Constructors

Anchor

Fields

anchorUrl ∷ !Url

anchorDataHash ∷ !(SafeHash c AnchorData)

data DRepState c

Constructors

DRepState

Fields

drepExpiry ∷ !EpochNo

drepAnchor ∷ !(StrictMaybe (Anchor c))

drepDeposit ∷ !Coin

drepDelegs ∷ !(Set (Credential 'Staking c))

data DRep c

Constructors

DRepKeyHash !(KeyHash 'DRepRole c)

DRepScriptHash !(ScriptHash c)

DRepAlwaysAbstain

DRepAlwaysNoConfidence

Bundled Patterns

pattern DRepCredential ∷ Credential 'DRepRole c → DRep c

newtype CommitteeState era

Constructors

CommitteeState

Fields

csCommitteeCreds ∷ Map (Credential 'ColdCommitteeRole (EraCrypto era)) (CommitteeAuthorization (EraCrypto era))

authorizedHotCommitteeCredentials ∷ CommitteeState era → Set (Credential 'HotCommitteeRole (EraCrypto era))

Extract all unique hot credential authorizations for the current committee. Note that there is no unique mapping from Hot to Cold credential, therefore we produce a Set, instead of a Map.

data AnchorData

lookupDepositDState ∷ DState era → StakeCredential (EraCrypto era) → Maybe Coin

Function that looks up the deposit for currently delegated staking credential

lookupRewardDState ∷ DState era → StakeCredential (EraCrypto era) → Maybe Coin

Function that looks up curret reward for the delegated staking credential.

rewards ∷ DState era → UView (EraCrypto era) (Credential 'Staking (EraCrypto era)) RDPair

delegations ∷ DState era → UView (EraCrypto era) (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era))

ptrsMap ∷ DState era → Map Ptr (Credential 'Staking (EraCrypto era))

get the actual ptrs map, we don't need a view

payPoolDeposit ∷ EraPParams era ⇒ KeyHash 'StakePool (EraCrypto era) → PParams era → PState era → PState era

One only pays a deposit on the initial pool registration. So return the the Deposits unchanged if the keyhash already exists. There are legal situations where a pool may be registered multiple times.

refundPoolDeposit ∷ KeyHash 'StakePool (EraCrypto era) → PState era → (Coin, PState era)

obligationCertState ∷ CertState era → Obligations

Calculate total possible refunds in the system that are related to certificates

There is an invariant that the sum of all the fields should be the same as the utxosDeposited field of the UTxOState. Note that this does not depend upon the current values of the Key and Pool deposits of the PParams.

data Obligations

A composite of all the Deposits the system is obligated to eventually pay back.

Constructors

Obligations

Fields

oblStake ∷ !Coin

oblPool ∷ !Coin

oblDRep ∷ !Coin

oblProposal ∷ !Coin

sumObligation ∷ Obligations → Coin

certsTotalDepositsTxBody ∷ EraTxBody era ⇒ PParams era → CertState era → TxBody era → Coin

Compute the total deposits from the Certs of a TxBody.

This is the contribution of a TxBody towards the deposit pot (utxosDeposit field of the UTxOState) of the system

certsTotalRefundsTxBody ∷ EraTxBody era ⇒ PParams era → CertState era → TxBody era → Coin

Compute the total refunds from the Certs of a TxBody.

This is the contribution of a TxBody towards the total Obligations of the system See Obligations and obligationCertState for more information.

certDStateL ∷ Lens' (CertState era) (DState era)

certPStateL ∷ Lens' (CertState era) (PState era)

certVStateL ∷ Lens' (CertState era) (VState era)

dsUnifiedL ∷ Lens' (DState era) (UMap (EraCrypto era))

dsGenDelegsL ∷ Lens' (DState era) (GenDelegs (EraCrypto era))

dsIRewardsL ∷ Lens' (DState era) (InstantaneousRewards (EraCrypto era))

dsFutureGenDelegsL ∷ Lens' (DState era) (Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

psStakePoolParamsL ∷ Lens' (PState era) (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

psFutureStakePoolParamsL ∷ Lens' (PState era) (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

psRetiringL ∷ Lens' (PState era) (Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

psDepositsL ∷ Lens' (PState era) (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

vsDRepsL ∷ Lens' (VState era) (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

vsCommitteeStateL ∷ Lens' (VState era) (CommitteeState era)

vsNumDormantEpochsL ∷ Lens' (VState era) EpochNo

vsActualDRepExpiry ∷ Credential 'DRepRole (EraCrypto era) → VState era → Maybe EpochNo

csCommitteeCredsL ∷ Lens' (CommitteeState era) (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (CommitteeAuthorization (EraCrypto era)))

lookupDepositVState ∷ VState era → Credential 'DRepRole (EraCrypto era) → Maybe Coin

Function that looks up the deposit for currently registered DRep