Cardano.Ledger.CertState

Documentation

data CertState era

The state associated with the DELPL rule, which combines the DELEG rule and the POOL rule.

Constructors

CertState

Fields

certVState ∷ !(VState era)

certPState ∷ !(PState era)

certDState ∷ !(DState era)

data DState era

The state used by the DELEG rule, which roughly tracks stake delegation and some governance features.

Constructors

DState

Fields

dsUnified ∷ !(UMap (EraCrypto era))

Unified Reward Maps. This contains the reward map (which is the source of truth regarding the registered stake credentials, the deposit map, the delegation map, and the stake credential pointer map.

dsFutureGenDelegs ∷ !(Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

Future genesis key delegations

dsGenDelegs ∷ !(GenDelegs (EraCrypto era))

Genesis key delegations

dsIRewards ∷ !(InstantaneousRewards (EraCrypto era))

Instantaneous Rewards

data PState era

The state used by the POOL rule, which tracks stake pool information.

Constructors

PState

Fields

psStakePoolParams ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

The stake pool parameters.

psFutureStakePoolParams ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

The future stake pool parameters. Changes to existing stake pool parameters are staged in order to give delegators time to react to changes. See section 11.2, "Example Illustration of the Reward Cycle", of the Shelley Ledger Specification for a sequence diagram.

psRetiring ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

A map of retiring stake pools to the epoch when they retire.

psDeposits ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) Coin)

A map of the deposits for each pool

data VState era

Constructors

VState

Fields

vsDReps ∷ !(Set (Credential 'Voting (EraCrypto era)))

vsCCHotKeys ∷ !(Map (KeyHash 'Voting (EraCrypto era)) (KeyHash 'Voting (EraCrypto era)))

data InstantaneousRewards c

InstantaneousRewards captures the pending changes to the ledger state caused by MIR certificates. It consists of two mappings, the rewards which will be paid out from the reserves and the rewards which will be paid out from the treasury. It also consists of two coin values which represent the transfer of coins from one pot to the other pot. NOTE that the following property should always hold: deltaReserves + deltaTreasury = 0

Constructors

InstantaneousRewards

Fields

iRReserves ∷ !(Map (Credential 'Staking c) Coin)

iRTreasury ∷ !(Map (Credential 'Staking c) Coin)

deltaReserves ∷ !DeltaCoin

deltaTreasury ∷ !DeltaCoin

data FutureGenDeleg c

Constructors

FutureGenDeleg

Fields

fGenDelegSlot ∷ !SlotNo

fGenDelegGenKeyHash ∷ !(KeyHash 'Genesis c)

lookupDepositDState ∷ DState era → StakeCredential (EraCrypto era) → Maybe Coin

Function that looks up the deposit for currently delegated staking credential

lookupRewardDState ∷ DState era → StakeCredential (EraCrypto era) → Maybe Coin

Function that looks up curret reward for the delegated staking credential.

rewards ∷ DState era → View (EraCrypto era) (Credential 'Staking (EraCrypto era)) RDPair

delegations ∷ DState era → View (EraCrypto era) (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era))

ptrsMap ∷ DState era → Map Ptr (Credential 'Staking (EraCrypto era))

get the actual ptrs map, we don't need a view

payPoolDeposit ∷ EraPParams era ⇒ KeyHash 'StakePool (EraCrypto era) → PParams era → PState era → PState era

One only pays a deposit on the initial pool registration. So return the the Deposits unchanged if the keyhash already exists. There are legal situations where a pool may be registered multiple times.

refundPoolDeposit ∷ KeyHash 'StakePool (EraCrypto era) → PState era → (Coin, PState era)

obligationCertState ∷ CertState era → Coin

Calculate total possible refunds in the system. There is an invariant that this should be the same as the utxosDeposited field of the UTxOState. Note that this does not depend upon the current values of the Key and Pool deposits of the PParams.