Cardano.Ledger.CompactAddress

Description

Deprecated: Use Address instead

Documentation

data Addr c

An address for UTxO.

Contents of Addr data type are intentionally left as lazy, otherwise operating on compact form of an address will result in redundant work.

Constructors

Addr Network (PaymentCredential c) (StakeReference c)

AddrBootstrap (BootstrapAddress c)

data RewardAcnt c

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork :: !Network

getRwdCred :: !(Credential 'Staking c)

newtype BootstrapAddress c

Constructors

BootstrapAddress

Fields

unBootstrapAddress :: Address

fromBoostrapCompactAddress :: CompactAddress -> CompactAddr c

Convert Byron's comapct address into CompactAddr. This is just an efficient type cast.

compactAddr :: Addr c -> CompactAddr c

decompactAddr :: forall c. (HasCallStack, Crypto c) => CompactAddr c -> Addr c

data CompactAddr c

unCompactAddr :: CompactAddr c -> ShortByteString

Unwrap the compact address and get to the address' binary representation.

isPayCredScriptCompactAddr :: CompactAddr c -> Bool

Efficiently check whether compacted adddress is an address with a credential that is a payment script.

isBootstrapCompactAddr :: CompactAddr c -> Bool

Efficiently check whether compated adddress is a Byron address.

decodeAddr :: forall c m. (Crypto c, MonadFail m) => ByteString -> m (Addr c)

decodeAddrShort :: forall c m. (Crypto c, MonadFail m) => ShortByteString -> m (Addr c)

decodeAddrEither :: forall c. Crypto c => ByteString -> Either String (Addr c)

decodeAddrShortEither :: forall c. Crypto c => ShortByteString -> Either String (Addr c)

fromCborAddr :: forall c s. Crypto c => Decoder s (Addr c)

Decoder for an Addr. Works in all eras

fromCborBothAddr :: forall c s. Crypto c => Decoder s (Addr c, CompactAddr c)

This is the decoder for an address that returns both the actual Addr and the bytes, that it was encoded as.

fromCborCompactAddr :: forall c s. Crypto c => Decoder s (CompactAddr c)

Returns the actual bytes that represent an addres, while ensuring that they can be decoded in any era as an Addr when need be.

fromCborBackwardsBothAddr :: forall c s. Crypto c => Decoder s (Addr c, CompactAddr c)

Prior to Babbage era we did not check if a binary blob representing an address was fully consumed, so unfortunately we must preserve this behavior. However, we do not need to preserve the unconsumed bytes in memory, therefore we can to drop the garbage after we successfully decoded the malformed address. We also need to allow bogus pointer address to be deserializeable prior to Babbage era.

decodeRewardAcnt :: forall c b m. (Crypto c, AddressBuffer b, MonadFail m) => b -> m (RewardAcnt c)

fromCborRewardAcnt :: forall c s. Crypto c => Decoder s (RewardAcnt c)