Cardano.Ledger.Core

Description

This module defines core type families which we know to vary from era to era.

Families in this module should be indexed on era.

It is intended for qualified import: > import qualified Cardano.Ledger.Core as Core

Era-changing types

class (EraTxBody era, EraTxWits era, EraTxAuxData era, EraPParams era, NoThunks (Tx era), DecCBOR (Annotator (Tx era)), EncCBOR (Tx era), ToCBOR (Tx era), Show (Tx era), Eq (Tx era), EqRaw (Tx era)) => EraTx era where

A transaction.

Associated Types

type Tx era = (r :: Type) | r -> era

type TxUpgradeError era :: Type

type TxUpgradeError era = Void

Methods

mkBasicTx :: TxBody era -> Tx era

bodyTxL :: Lens' (Tx era) (TxBody era)

witsTxL :: Lens' (Tx era) (TxWits era)

auxDataTxL :: Lens' (Tx era) (StrictMaybe (AuxiliaryData era))

sizeTxF :: SimpleGetter (Tx era) Integer

validateNativeScript :: Tx era -> NativeScript era -> Bool

Using information from the transaction validate the supplied native script.

getMinFeeTx :: PParams era -> Tx era -> Coin

upgradeTx :: EraTx (PreviousEra era) => Tx (PreviousEra era) -> Either (TxUpgradeError era) (Tx era)

txIdTx :: EraTx era => Tx era -> TxId (EraCrypto era)

class (Val (Value era), ToJSON (TxOut era), DecCBOR (Value era), DecCBOR (CompactForm (Value era)), EncCBOR (Value era), ToCBOR (TxOut era), FromCBOR (TxOut era), EncCBOR (TxOut era), DecCBOR (TxOut era), DecShareCBOR (TxOut era), Share (TxOut era) ~ Interns (Credential 'Staking (EraCrypto era)), NoThunks (TxOut era), NFData (TxOut era), Show (TxOut era), Eq (TxOut era), EraPParams era) => EraTxOut era where

Abstract interface into specific fields of a TxOut

Minimal complete definition

mkBasicTxOut, upgradeTxOut, valueEitherTxOutL, addrEitherTxOutL, (getMinCoinSizedTxOut | getMinCoinTxOut)

Associated Types

type TxOut era = (r :: Type) | r -> era

The output of a UTxO for a particular era

Methods

mkBasicTxOut :: Addr (EraCrypto era) -> Value era -> TxOut era

upgradeTxOut :: EraTxOut (PreviousEra era) => TxOut (PreviousEra era) -> TxOut era

Every era, except Shelley, must be able to upgrade a TxOut from a previous era.

valueTxOutL :: Lens' (TxOut era) (Value era)

compactValueTxOutL :: HasCallStack => Lens' (TxOut era) (CompactForm (Value era))

valueEitherTxOutL :: Lens' (TxOut era) (Either (Value era) (CompactForm (Value era)))

Lens for getting and setting in TxOut either an address or its compact version by doing the least amount of work.

addrTxOutL :: Lens' (TxOut era) (Addr (EraCrypto era))

compactAddrTxOutL :: Lens' (TxOut era) (CompactAddr (EraCrypto era))

addrEitherTxOutL :: Lens' (TxOut era) (Either (Addr (EraCrypto era)) (CompactAddr (EraCrypto era)))

Lens for getting and setting in TxOut either an address or its compact version by doing the least amount of work.

The utility of this function comes from the fact that TxOut usually stores the address in either one of two forms: compacted or unpacked. In order to avoid extroneous conversions in getTxOutAddr and getTxOutCompactAddr we can define just this functionality. Also sometimes it is crucial to know at the callsite which form of address we have readily available without any conversions (eg. searching millions of TxOuts for a particular address)

getMinCoinSizedTxOut :: PParams era -> Sized (TxOut era) -> Coin

Produce the minimum lovelace that a given transaction output must contain. Information about the size of the TxOut is required in some eras. Use getMinCoinTxOut if you don't have the size readily available to you.

getMinCoinTxOut :: PParams era -> TxOut era -> Coin

Same as getMinCoinSizedTxOut, except information about the size of TxOut will be computed by serializing the TxOut. If the size turns out to be not needed, then serialization will have no overhead, since it is computed lazily.

bootAddrTxOutF :: EraTxOut era => SimpleGetter (TxOut era) (Maybe (BootstrapAddress (EraCrypto era)))

coinTxOutL :: (HasCallStack, EraTxOut era) => Lens' (TxOut era) Coin

compactCoinTxOutL :: (HasCallStack, EraTxOut era) => Lens' (TxOut era) (CompactForm Coin)

isAdaOnlyTxOutF :: EraTxOut era => SimpleGetter (TxOut era) Bool

This is a getter that implements an efficient way to check whether TxOut contains ADA only.

class (EraTxOut era, EraTxCert era, EraPParams era, HashAnnotated (TxBody era) EraIndependentTxBody (EraCrypto era), DecCBOR (Annotator (TxBody era)), EncCBOR (TxBody era), ToCBOR (TxBody era), NoThunks (TxBody era), NFData (TxBody era), Show (TxBody era), Eq (TxBody era), EqRaw (TxBody era)) => EraTxBody era where

Minimal complete definition

mkBasicTxBody, inputsTxBodyL, outputsTxBodyL, feeTxBodyL, withdrawalsTxBodyL, auxDataHashTxBodyL, spendableInputsTxBodyF, allInputsTxBodyF, certsTxBodyL, upgradeTxBody

Associated Types

type TxBody era = (r :: Type) | r -> era

The body of a transaction.

type TxBodyUpgradeError era :: Type

type TxBodyUpgradeError era = Void

Methods

mkBasicTxBody :: TxBody era

inputsTxBodyL :: Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

outputsTxBodyL :: Lens' (TxBody era) (StrictSeq (TxOut era))

feeTxBodyL :: Lens' (TxBody era) Coin

withdrawalsTxBodyL :: Lens' (TxBody era) (Withdrawals (EraCrypto era))

auxDataHashTxBodyL :: Lens' (TxBody era) (StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

spendableInputsTxBodyF :: SimpleGetter (TxBody era) (Set (TxIn (EraCrypto era)))

This getter will produce all inputs from the UTxO map that this transaction might spend, which ones will depend on the validity of the transaction itself. Starting in Alonzo this will include collateral inputs.

allInputsTxBodyF :: SimpleGetter (TxBody era) (Set (TxIn (EraCrypto era)))

This getter will produce all inputs from the UTxO map that this transaction is referencing, even if some of them cannot be spent by the transaction. For example starting with Babbage era it will also include reference inputs.

certsTxBodyL :: Lens' (TxBody era) (StrictSeq (TxCert era))

getTotalDepositsTxBody

Arguments

:: PParams era

-> (KeyHash 'StakePool (EraCrypto era) -> Bool)
Check whether stake pool is registered or not

-> TxBody era

-> Coin

Compute the total deposits from the certificates in a TxBody.

This is the contribution of a TxBody towards the consumed amount by the transaction

getTotalRefundsTxBody

Arguments

:: PParams era

-> (Credential 'Staking (EraCrypto era) -> Maybe Coin)
Lookup current deposit for Staking credential if one is registered

-> (Credential 'DRepRole (EraCrypto era) -> Maybe Coin)
Lookup current deposit for DRep credential if one is registered

-> TxBody era

-> Coin

Compute the total refunds from the Certs of a TxBody.

This is the contribution of a TxBody towards produced amount by the transaction

getGenesisKeyHashCountTxBody :: TxBody era -> Int

This function is not used in the ledger rules. It is only used by the downstream tooling to figure out how many witnesses should be supplied for Genesis keys.

upgradeTxBody :: EraTxBody (PreviousEra era) => TxBody (PreviousEra era) -> Either (TxBodyUpgradeError era) (TxBody era)

Upgrade the transaction body from the previous era.

This can fail where elements of the transaction body are deprecated. Compare this to translateEraThroughCBOR: - upgradeTxBody will use the Haskell representation, but will not preserve the serialised form. However, it will be suitable for iterated translation through eras. - translateEraThroughCBOR will preserve the binary representation, but is not guaranteed to work through multiple eras - that is, the serialised representation from era n is guaranteed valid in era n + 1, but not necessarily in era n + 2.

txIdTxBody :: EraTxBody era => TxBody era -> TxId (EraCrypto era)

class (Era era, Eq (TxAuxData era), EqRaw (TxAuxData era), Show (TxAuxData era), NoThunks (TxAuxData era), ToCBOR (TxAuxData era), EncCBOR (TxAuxData era), DecCBOR (Annotator (TxAuxData era)), HashAnnotated (TxAuxData era) EraIndependentTxAuxData (EraCrypto era)) => EraTxAuxData era where

TxAuxData which may be attached to a transaction

Associated Types

type TxAuxData era = (r :: Type) | r -> era

Methods

upgradeTxAuxData :: EraTxAuxData (PreviousEra era) => TxAuxData (PreviousEra era) -> TxAuxData era

Every era, except Shelley, must be able to upgrade a TxAuxData from a previous era.

Warning - Important to note that any memoized binary representation will not be preserved. If you need to retain underlying bytes you can use translateEraThroughCBOR

hashTxAuxData :: TxAuxData era -> AuxiliaryDataHash (EraCrypto era)

validateTxAuxData :: ProtVer -> TxAuxData era -> Bool

class (EraScript era, Eq (TxWits era), EqRaw (TxWits era), Show (TxWits era), Monoid (TxWits era), NoThunks (TxWits era), ToCBOR (TxWits era), EncCBOR (TxWits era), DecCBOR (Annotator (TxWits era))) => EraTxWits era where

A collection of witnesses in a Tx

Minimal complete definition

addrTxWitsL, bootAddrTxWitsL, scriptTxWitsL, upgradeTxWits

Associated Types

type TxWits era = (r :: Type) | r -> era

Methods

mkBasicTxWits :: TxWits era

addrTxWitsL :: Lens' (TxWits era) (Set (WitVKey 'Witness (EraCrypto era)))

bootAddrTxWitsL :: Lens' (TxWits era) (Set (BootstrapWitness (EraCrypto era)))

scriptTxWitsL :: Lens' (TxWits era) (Map (ScriptHash (EraCrypto era)) (Script era))

upgradeTxWits :: EraTxWits (PreviousEra era) => TxWits (PreviousEra era) -> TxWits era

class (Era era, Show (Script era), Eq (Script era), EqRaw (Script era), ToCBOR (Script era), EncCBOR (Script era), DecCBOR (Annotator (Script era)), NoThunks (Script era), SafeToHash (Script era), Eq (NativeScript era), Show (NativeScript era), NFData (NativeScript era), NoThunks (NativeScript era), EncCBOR (NativeScript era), DecCBOR (Annotator (NativeScript era))) => EraScript era where

Typeclass for script data types. Allows for script validation and hashing. You must understand the role of SafeToHash and scriptPrefixTag to make new instances. scriptPrefixTag is a magic number representing the tag of the script language. For each new script language defined, a new tag is chosen and the tag is included in the script hash for a script. The safeToHash constraint ensures that Scripts are never reserialised.

Minimal complete definition

upgradeScript, scriptPrefixTag, getNativeScript, fromNativeScript

Associated Types

type Script era = (r :: Type) | r -> era

Scripts which may lock transaction outputs in this era

type NativeScript era = (r :: Type) | r -> era

Methods

upgradeScript :: EraScript (PreviousEra era) => Script (PreviousEra era) -> Script era

Every era, except Shelley, must be able to upgrade a Script from a previous era.

Warning - Important to note that any memoized binary representation will not be preserved, you need to retain underlying bytes you can use translateEraThroughCBOR

scriptPrefixTag :: Script era -> ByteString

hashScript :: Script era -> ScriptHash (EraCrypto era)

getNativeScript :: Script era -> Maybe (NativeScript era)

fromNativeScript :: NativeScript era -> Script era

isNativeScript :: EraScript era => Script era -> Bool

hashScriptTxWitsL :: EraTxWits era => Lens (TxWits era) (TxWits era) (Map (ScriptHash (EraCrypto era)) (Script era)) [Script era]

This is a helper lens that will hash the scripts when adding as witnesses.

type family Value era :: Type

A value is something which quantifies a transaction output.

class (Era era, Eq (PParamsHKD Identity era), Ord (PParamsHKD Identity era), Show (PParamsHKD Identity era), NFData (PParamsHKD Identity era), EncCBOR (PParamsHKD Identity era), DecCBOR (PParamsHKD Identity era), ToCBOR (PParamsHKD Identity era), FromCBOR (PParamsHKD Identity era), NoThunks (PParamsHKD Identity era), ToJSON (PParamsHKD Identity era), FromJSON (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Ord (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), NFData (PParamsHKD StrictMaybe era), EncCBOR (PParamsHKD StrictMaybe era), DecCBOR (PParamsHKD StrictMaybe era), ToCBOR (PParamsHKD StrictMaybe era), FromCBOR (PParamsHKD StrictMaybe era), NoThunks (PParamsHKD StrictMaybe era), ToJSON (PParamsHKD StrictMaybe era)) => EraPParams era where

Minimal complete definition

emptyPParamsIdentity, emptyPParamsStrictMaybe, upgradePParamsHKD, downgradePParamsHKD, hkdMinFeeAL, hkdMinFeeBL, hkdMaxBBSizeL, hkdMaxTxSizeL, hkdMaxBHSizeL, hkdKeyDepositL, hkdPoolDepositL, hkdEMaxL, hkdNOptL, hkdA0L, hkdRhoL, hkdTauL, hkdDL, hkdExtraEntropyL, hkdProtocolVersionL, hkdMinUTxOValueL, hkdMinPoolCostL

Associated Types

type PParamsHKD (f :: Type -> Type) era = (r :: Type) | r -> era

Protocol parameters where the fields are represented with a HKD

type UpgradePParams (f :: Type -> Type) era :: Type

type DowngradePParams (f :: Type -> Type) era :: Type

Methods

applyPPUpdates :: PParams era -> PParamsUpdate era -> PParams era

Applies a protocol parameters update

default applyPPUpdates :: forall a u. (Generic (PParamsHKD Identity era), Generic (PParamsHKD StrictMaybe era), Updatable (Rep (PParamsHKD Identity era) a) (Rep (PParamsHKD StrictMaybe era) u)) => PParams era -> PParamsUpdate era -> PParams era

emptyPParamsIdentity :: PParamsHKD Identity era

emptyPParamsStrictMaybe :: PParamsHKD StrictMaybe era

upgradePParamsHKD :: (HKDFunctor f, EraPParams (PreviousEra era)) => UpgradePParams f era -> PParamsHKD f (PreviousEra era) -> PParamsHKD f era

Upgrade PParams from previous era to the current one

downgradePParamsHKD :: (HKDFunctor f, EraPParams (PreviousEra era)) => DowngradePParams f era -> PParamsHKD f era -> PParamsHKD f (PreviousEra era)

Downgrade PParams from the current era to the previous one

hkdMinFeeAL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The linear factor for the minimum fee calculation

hkdMinFeeBL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The constant factor for the minimum fee calculation

hkdMaxBBSizeL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Word32)

Maximal block body size

hkdMaxTxSizeL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Word32)

Maximal transaction size

hkdMaxBHSizeL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Word16)

Maximal block header size

hkdKeyDepositL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a key registration deposit

hkdPoolDepositL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a pool registration deposit

hkdEMaxL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f EpochInterval)

epoch bound on pool retirement

hkdNOptL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Natural)

Desired number of pools

hkdA0L :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f NonNegativeInterval)

Pool influence

hkdRhoL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f UnitInterval)

Monetary expansion

hkdTauL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f UnitInterval)

Treasury expansion

hkdDL :: (HKDFunctor f, ProtVerAtMost era 6) => Lens' (PParamsHKD f era) (HKD f UnitInterval)

Decentralization parameter

ppDG :: SimpleGetter (PParams era) UnitInterval

Decentralization parameter getter

default ppDG :: ProtVerAtMost era 6 => SimpleGetter (PParams era) UnitInterval

hkdExtraEntropyL :: (HKDFunctor f, ProtVerAtMost era 6) => Lens' (PParamsHKD f era) (HKD f Nonce)

Extra entropy

hkdProtocolVersionL :: (HKDFunctor f, ProtVerAtMost era 8) => Lens' (PParamsHKD f era) (HKD f ProtVer)

Protocol version

ppProtocolVersionL :: Lens' (PParams era) ProtVer

default ppProtocolVersionL :: ProtVerAtMost era 8 => Lens' (PParams era) ProtVer

ppuProtocolVersionL :: ProtVerAtMost era 8 => Lens' (PParamsUpdate era) (StrictMaybe ProtVer)

PParamsUpdate Protocol version

hkdMinUTxOValueL :: HKDFunctor f => ProtVerAtMost era 4 => Lens' (PParamsHKD f era) (HKD f Coin)

Minimum UTxO value

hkdMinPoolCostL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

Minimum Stake Pool Cost

Era STS

type family EraRule (k :: Symbol) era :: Type

Era STS map

class (Crypto (EraCrypto era), Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT, ProtVerLow era <= MaxVersion, ProtVerHigh era <= MaxVersion) => Era era where

Associated Types

type EraCrypto era :: Type

type PreviousEra era = (r :: Type) | r -> era

Map an era to its predecessor.

For example:

type instance PreviousEra (AllegraEra c) = ShelleyEra c

type ProtVerLow era :: Nat

Lowest major protocol version for this era

type ProtVerHigh era :: Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

Methods

eraName :: String

Textual name of the current era.

Designed to be used with TypeApplications:

>>> eraName @(ByronEra StandardCrypto)
Byron

Segregated Witness

The idea of segregated witnessing is to alter the encoding of transactions in a block such that the witnesses (the information needed to verify the validity of the transactions) can be stored separately from the body (the information needed to update the ledger state). In this way, a node which only cares about replaying transactions need not even decode the witness information.

In order to do this, we introduce two concepts: - A TxSeq, which represents the decoded structure of a sequence of transactions as represented in the encoded block; that is, with witnessing, metadata and other non-body parts split separately.

class (EraTx era, Eq (TxSeq era), Show (TxSeq era), EncCBORGroup (TxSeq era), DecCBOR (Annotator (TxSeq era))) => EraSegWits era where

Indicates that an era supports segregated witnessing.

This class embodies an isomorphism between 'TxSeq era' and 'StrictSeq (Tx era)', witnessed by fromTxSeq and toTxSeq.

Associated Types

type TxSeq era = (r :: Type) | r -> era

Methods

fromTxSeq :: TxSeq era -> StrictSeq (Tx era)

toTxSeq :: StrictSeq (Tx era) -> TxSeq era

hashTxSeq :: TxSeq era -> Hash (HASH (EraCrypto era)) EraIndependentBlockBody

Get the block body hash from the TxSeq. Note that this is not a regular "hash the stored bytes" function since the block body hash forms a small Merkle tree.

numSegComponents :: Word64

The number of segregated components

bBodySize :: forall era. EraSegWits era => ProtVer -> TxSeq era -> Int

Rewards

data RewardType

The staking rewards in Cardano are all either:

member rewards - rewards given to a registered stake credential which has delegated to a stake pool, or

leader rewards - rewards given to a registered stake pool (in particular, given to the stake credential in the stake pool registration certificate).

See Figure 47, "Functions used in the Reward Splitting", of the formal specification for more details.

Constructors

MemberReward

LeaderReward

data Reward c

The Reward type captures:

if the reward is a member or leader reward

the stake pool ID associated with the reward

the number of Lovelace in the reward

Constructors

Reward

Fields

rewardType :: !RewardType

rewardPool :: !(KeyHash 'StakePool c)

rewardAmount :: !Coin

Re-exports

module Cardano.Ledger.Hashes

Era

class (Crypto (EraCrypto era), Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT, ProtVerLow era <= MaxVersion, ProtVerHigh era <= MaxVersion) => Era era where

Associated Types

type EraCrypto era :: Type

type PreviousEra era = (r :: Type) | r -> era

Map an era to its predecessor.

For example:

type instance PreviousEra (AllegraEra c) = ShelleyEra c

type ProtVerLow era :: Nat

Lowest major protocol version for this era

type ProtVerHigh era :: Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

Methods

eraName :: String

Textual name of the current era.

Designed to be used with TypeApplications:

>>> eraName @(ByronEra StandardCrypto)
Byron

data ByronEra c

This is the era that preceded Shelley era. It cannot have any other class instances, except for Era type class.

Protocol Version

type AtMostEra (eraName :: Type -> Type) era = ProtVerAtMost era (ProtVerHigh (eraName (EraCrypto era)))

Restrict the era to equal to eraName or come before it.

type AtLeastEra (eraName :: Type -> Type) era = ProtVerAtLeast era (ProtVerLow (eraName (EraCrypto era)))

Restrict the era to equal to eraName or come after it

type ExactEra (inEra :: Type -> Type) era = ProtVerInBounds era (ProtVerLow (inEra (EraCrypto era))) (ProtVerHigh (inEra (EraCrypto era)))

Restrict an era to the specific era through the protocol version. This is equivalent to (inEra (Crypto era) ~ era)

type family ProtVerAtMost era (h :: Nat) :: Constraint where ...

Requirement for the era's lowest protocol version to be lower or equal to the supplied value

Equations

ProtVerAtMost era h = ProtVerIsInBounds "at most" era h (ProtVerLow era <=? h)

type family ProtVerAtLeast era (l :: Nat) :: Constraint where ...

Requirement for the era's highest protocol version to be higher or equal to the supplied value

Equations

ProtVerAtLeast era l = ProtVerIsInBounds "at least" era l (l <=? ProtVerHigh era)

type ProtVerInBounds era l h = (ProtVerAtLeast era l, ProtVerAtMost era h)

Restrict a lower and upper bounds of the protocol version for the particular era

atLeastEra :: AtLeastEra eraName era => ()

Enforce era to be at least the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

>>> atLeastEra @BabbageEra @(ConwayEra StandardCrypto)
>>> atLeastEra @BabbageEra @(BabbageEra StandardCrypto)

However this will result in a type error

>>> atLeastEra @BabbageEra @(AlonzoEra StandardCrypto)

atMostEra :: AtMostEra eraName era => ()

Enforce era to be at most the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

>>> atMostEra @BabbageEra @(ShelleyEra StandardCrypto)
>>> atMostEra @AlonzoEra @(MaryEra StandardCrypto)

However this will result in a type error

>>> atMostEra @BabbageEra @(ConwayEra StandardCrypto)

notSupportedInThisEra :: HasCallStack => a

notSupportedInThisEraL :: HasCallStack => Lens' a b

eraProtVerLow :: forall era. Era era => Version

Get the value level Version of the lowest major protocol version for the supplied era.

eraProtVerHigh :: forall era. Era era => Version

Get the value level Version of the highest major protocol version for the supplied era.

toEraCBOR :: forall era t. (Era era, EncCBOR t) => t -> Encoding

Convert a type that implements EncCBOR to plain Encoding using the lowest protocol version for the supplied era

fromEraCBOR :: forall era t s. (Era era, DecCBOR t) => Decoder s t

Convert a type that implements DecCBOR to plain Decoder using the lowest protocol version for the supplied era

fromEraShareCBOR :: forall era t s. (Era era, DecShareCBOR t) => Decoder s t

Convert a type that implements DecShareCBOR to plain Decoder using the lowest protocol version for the supplied era

eraDecoder :: forall era t s. Era era => Decoder s t -> Decoder s t

Convert a versioned Decoder to plain a Decoder using the lowest protocol version for the supplied era

class (Era era, ToJSON (TxCert era), DecCBOR (TxCert era), EncCBOR (TxCert era), ToCBOR (TxCert era), FromCBOR (TxCert era), NoThunks (TxCert era), NFData (TxCert era), Show (TxCert era), Eq (TxCert era)) => EraTxCert era where

Associated Types

type TxCert era = (r :: Type) | r -> era

type TxCertUpgradeError era :: Type

type TxCertUpgradeError era = Void

Methods

upgradeTxCert :: EraTxCert (PreviousEra era) => TxCert (PreviousEra era) -> Either (TxCertUpgradeError era) (TxCert era)

Every era, except Shelley, must be able to upgrade a TxCert from a previous era. However, not all certificates can be upgraded, because some eras lose some of the certificates, thus return type is an Either. Eg. from Babbage to Conway: MIR and Genesis certificates were removed.

getVKeyWitnessTxCert :: TxCert era -> Maybe (KeyHash 'Witness (EraCrypto era))

Return a witness key whenever a certificate requires one

getScriptWitnessTxCert :: TxCert era -> Maybe (ScriptHash (EraCrypto era))

Return a ScriptHash for certificate types that require a witness

mkRegPoolTxCert :: PoolParams (EraCrypto era) -> TxCert era

getRegPoolTxCert :: TxCert era -> Maybe (PoolParams (EraCrypto era))

mkRetirePoolTxCert :: KeyHash 'StakePool (EraCrypto era) -> EpochNo -> TxCert era

getRetirePoolTxCert :: TxCert era -> Maybe (KeyHash 'StakePool (EraCrypto era), EpochNo)

lookupRegStakeTxCert :: TxCert era -> Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can register such credential

lookupUnRegStakeTxCert :: TxCert era -> Maybe (Credential 'Staking (EraCrypto era))

Extract staking credential from any certificate that can unregister such credential

getTotalDepositsTxCerts

Arguments

:: Foldable f

=> PParams era

-> (KeyHash 'StakePool (EraCrypto era) -> Bool)
Check whether stake pool is registered or not

-> f (TxCert era)

-> Coin

Compute the total deposits from a list of certificates.

getTotalRefundsTxCerts

Arguments

:: Foldable f

=> PParams era

-> (Credential 'Staking (EraCrypto era) -> Maybe Coin)
Lookup current deposit for Staking credential if one is registered

-> (Credential 'DRepRole (EraCrypto era) -> Maybe Coin)
Lookup current deposit for DRep credential if one is registered

-> f (TxCert era)

-> Coin

Compute the total refunds from a list of certificates.

pattern RegPoolTxCert :: EraTxCert era => PoolParams (EraCrypto era) -> TxCert era

pattern RetirePoolTxCert :: EraTxCert era => KeyHash 'StakePool (EraCrypto era) -> EpochNo -> TxCert era

data Delegation c

Deprecated: No longer used

The delegation of one stake key to another.

Constructors

Delegation

Deprecated: No longer used

Fields

dDelegator :: !(StakeCredential c)

dDelegatee :: !(KeyHash 'StakePool c)

data PoolCert c

Constructors

RegPool !(PoolParams c)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool c) !EpochNo
A stake pool retirement certificate.

getPoolCertTxCert :: EraTxCert era => TxCert era -> Maybe (PoolCert (EraCrypto era))

poolCWitness :: PoolCert c -> Credential 'StakePool c

Deprecated: As useless. PoolIds have nothing to do with credentials

poolCertKeyHashWitness :: PoolCert c -> KeyHash 'Witness c

isRegStakeTxCert :: EraTxCert era => TxCert era -> Bool

Check if supplied TxCert is a stake registering certificate

isUnRegStakeTxCert :: EraTxCert era => TxCert era -> Bool

Check if supplied TxCert is a stake un-registering certificate

class (Era era, Eq (PParamsHKD Identity era), Ord (PParamsHKD Identity era), Show (PParamsHKD Identity era), NFData (PParamsHKD Identity era), EncCBOR (PParamsHKD Identity era), DecCBOR (PParamsHKD Identity era), ToCBOR (PParamsHKD Identity era), FromCBOR (PParamsHKD Identity era), NoThunks (PParamsHKD Identity era), ToJSON (PParamsHKD Identity era), FromJSON (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Ord (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), NFData (PParamsHKD StrictMaybe era), EncCBOR (PParamsHKD StrictMaybe era), DecCBOR (PParamsHKD StrictMaybe era), ToCBOR (PParamsHKD StrictMaybe era), FromCBOR (PParamsHKD StrictMaybe era), NoThunks (PParamsHKD StrictMaybe era), ToJSON (PParamsHKD StrictMaybe era)) => EraPParams era where

Minimal complete definition

emptyPParamsIdentity, emptyPParamsStrictMaybe, upgradePParamsHKD, downgradePParamsHKD, hkdMinFeeAL, hkdMinFeeBL, hkdMaxBBSizeL, hkdMaxTxSizeL, hkdMaxBHSizeL, hkdKeyDepositL, hkdPoolDepositL, hkdEMaxL, hkdNOptL, hkdA0L, hkdRhoL, hkdTauL, hkdDL, hkdExtraEntropyL, hkdProtocolVersionL, hkdMinUTxOValueL, hkdMinPoolCostL

Associated Types

type PParamsHKD (f :: Type -> Type) era = (r :: Type) | r -> era

Protocol parameters where the fields are represented with a HKD

type UpgradePParams (f :: Type -> Type) era :: Type

type DowngradePParams (f :: Type -> Type) era :: Type

Methods

applyPPUpdates :: PParams era -> PParamsUpdate era -> PParams era

Applies a protocol parameters update

default applyPPUpdates :: forall a u. (Generic (PParamsHKD Identity era), Generic (PParamsHKD StrictMaybe era), Updatable (Rep (PParamsHKD Identity era) a) (Rep (PParamsHKD StrictMaybe era) u)) => PParams era -> PParamsUpdate era -> PParams era

emptyPParamsIdentity :: PParamsHKD Identity era

emptyPParamsStrictMaybe :: PParamsHKD StrictMaybe era

upgradePParamsHKD :: (HKDFunctor f, EraPParams (PreviousEra era)) => UpgradePParams f era -> PParamsHKD f (PreviousEra era) -> PParamsHKD f era

Upgrade PParams from previous era to the current one

downgradePParamsHKD :: (HKDFunctor f, EraPParams (PreviousEra era)) => DowngradePParams f era -> PParamsHKD f era -> PParamsHKD f (PreviousEra era)

Downgrade PParams from the current era to the previous one

hkdMinFeeAL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The linear factor for the minimum fee calculation

hkdMinFeeBL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The constant factor for the minimum fee calculation

hkdMaxBBSizeL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Word32)

Maximal block body size

hkdMaxTxSizeL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Word32)

Maximal transaction size

hkdMaxBHSizeL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Word16)

Maximal block header size

hkdKeyDepositL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a key registration deposit

hkdPoolDepositL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a pool registration deposit

hkdEMaxL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f EpochInterval)

epoch bound on pool retirement

hkdNOptL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Natural)

Desired number of pools

hkdA0L :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f NonNegativeInterval)

Pool influence

hkdRhoL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f UnitInterval)

Monetary expansion

hkdTauL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f UnitInterval)

Treasury expansion

hkdDL :: (HKDFunctor f, ProtVerAtMost era 6) => Lens' (PParamsHKD f era) (HKD f UnitInterval)

Decentralization parameter

ppDG :: SimpleGetter (PParams era) UnitInterval

Decentralization parameter getter

default ppDG :: ProtVerAtMost era 6 => SimpleGetter (PParams era) UnitInterval

hkdExtraEntropyL :: (HKDFunctor f, ProtVerAtMost era 6) => Lens' (PParamsHKD f era) (HKD f Nonce)

Extra entropy

hkdProtocolVersionL :: (HKDFunctor f, ProtVerAtMost era 8) => Lens' (PParamsHKD f era) (HKD f ProtVer)

Protocol version

ppProtocolVersionL :: Lens' (PParams era) ProtVer

default ppProtocolVersionL :: ProtVerAtMost era 8 => Lens' (PParams era) ProtVer

ppuProtocolVersionL :: ProtVerAtMost era 8 => Lens' (PParamsUpdate era) (StrictMaybe ProtVer)

PParamsUpdate Protocol version

hkdMinUTxOValueL :: HKDFunctor f => ProtVerAtMost era 4 => Lens' (PParamsHKD f era) (HKD f Coin)

Minimum UTxO value

hkdMinPoolCostL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)

Minimum Stake Pool Cost

newtype PParams era

Protocol parameters

Constructors

PParams (PParamsHKD Identity era)

emptyPParams :: EraPParams era => PParams era

newtype PParamsUpdate era

The type of updates to Protocol parameters

Constructors

PParamsUpdate (PParamsHKD StrictMaybe era)

emptyPParamsUpdate :: EraPParams era => PParamsUpdate era

genericApplyPPUpdates :: forall era a u. (Generic (PParamsHKD Identity era), Generic (PParamsHKD StrictMaybe era), Updatable (Rep (PParamsHKD Identity era) a) (Rep (PParamsHKD StrictMaybe era) u)) => PParams era -> PParamsUpdate era -> PParams era

PParams lens

ppMinFeeAL :: forall era. EraPParams era => Lens' (PParams era) Coin

The linear factor for the minimum fee calculation

ppMinFeeBL :: forall era. EraPParams era => Lens' (PParams era) Coin

The constant factor for the minimum fee calculation

ppMaxBBSizeL :: forall era. EraPParams era => Lens' (PParams era) Word32

Maximal block body size

ppMaxTxSizeL :: forall era. EraPParams era => Lens' (PParams era) Word32

Maximal transaction size

ppMaxBHSizeL :: forall era. EraPParams era => Lens' (PParams era) Word16

Maximal block header size

ppKeyDepositL :: forall era. EraPParams era => Lens' (PParams era) Coin

The amount of a key registration deposit

ppPoolDepositL :: forall era. EraPParams era => Lens' (PParams era) Coin

The amount of a pool registration deposit

ppEMaxL :: forall era. EraPParams era => Lens' (PParams era) EpochInterval

epoch bound on pool retirement

ppNOptL :: forall era. EraPParams era => Lens' (PParams era) Natural

Desired number of pools

ppA0L :: forall era. EraPParams era => Lens' (PParams era) NonNegativeInterval

Pool influence

ppRhoL :: forall era. EraPParams era => Lens' (PParams era) UnitInterval

Monetary expansion

ppTauL :: forall era. EraPParams era => Lens' (PParams era) UnitInterval

Treasury expansion

ppDL :: forall era. (EraPParams era, ProtVerAtMost era 6) => Lens' (PParams era) UnitInterval

Decentralization parameter

ppExtraEntropyL :: forall era. (EraPParams era, ProtVerAtMost era 6) => Lens' (PParams era) Nonce

Extra entropy

ppMinUTxOValueL :: forall era. (EraPParams era, ProtVerAtMost era 4) => Lens' (PParams era) Coin

Minimum UTxO value

ppMinPoolCostL :: forall era. EraPParams era => Lens' (PParams era) Coin

Minimum Stake Pool Cost

PParamsUpdate lens

ppuMinFeeAL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Coin)

The linear factor for the minimum fee calculation

ppuMinFeeBL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Coin)

The constant factor for the minimum fee calculation

ppuMaxBBSizeL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Word32)

Maximal block body size

ppuMaxTxSizeL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Word32)

Maximal transaction size

ppuMaxBHSizeL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Word16)

Maximal block header size

ppuKeyDepositL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Coin)

The amount of a key registration deposit

ppuPoolDepositL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Coin)

The amount of a pool registration deposit

ppuEMaxL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

epoch bound on pool retirement

ppuNOptL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Natural)

Desired number of pools

ppuA0L :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe NonNegativeInterval)

Pool influence

ppuRhoL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Monetary expansion

ppuTauL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Treasury expansion

ppuDL :: forall era. (EraPParams era, ProtVerAtMost era 6) => Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Decentralization parameter

ppuExtraEntropyL :: forall era. (EraPParams era, ProtVerAtMost era 6) => Lens' (PParamsUpdate era) (StrictMaybe Nonce)

Extra entropy

ppuMinUTxOValueL :: forall era. (EraPParams era, ProtVerAtMost era 4) => Lens' (PParamsUpdate era) (StrictMaybe Coin)

Minimum UTxO value

ppuMinPoolCostL :: forall era. EraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Coin)

Minimum Stake Pool Cost

Utility

ppLens :: Lens' (PParams era) (PParamsHKD Identity era)

ppuLens :: Lens' (PParamsUpdate era) (PParamsHKD StrictMaybe era)

mapPParams :: (PParamsHKD Identity era1 -> PParamsHKD Identity era2) -> PParams era1 -> PParams era2

mapPParamsUpdate :: (PParamsHKD StrictMaybe era1 -> PParamsHKD StrictMaybe era2) -> PParamsUpdate era1 -> PParamsUpdate era2

upgradePParams :: (EraPParams era, EraPParams (PreviousEra era)) => UpgradePParams Identity era -> PParams (PreviousEra era) -> PParams era

downgradePParams :: (EraPParams era, EraPParams (PreviousEra era)) => DowngradePParams Identity era -> PParams era -> PParams (PreviousEra era)

upgradePParamsUpdate :: (EraPParams era, EraPParams (PreviousEra era)) => UpgradePParams StrictMaybe era -> PParamsUpdate (PreviousEra era) -> PParamsUpdate era

downgradePParamsUpdate :: (EraPParams era, EraPParams (PreviousEra era)) => DowngradePParams StrictMaybe era -> PParamsUpdate era -> PParamsUpdate (PreviousEra era)

PParamsUpdate to Data

data PParam era where

Pair the tag, and exisitenially hide the type of the lens for the field with that Lens'

Constructors

PParam :: ToPlutusData t => Word -> Lens' (PParamsUpdate era) (StrictMaybe t) -> PParam era

makePParamMap :: [PParam era] -> Map Word (PParam era)

Turn a list into a Map, this assures we have no duplicates.

type family PreviousEra era = (r :: Type) | r -> era

Map an era to its predecessor.

For example:

type instance PreviousEra (AllegraEra c) = ShelleyEra c

type family TranslationContext era :: Type

Per-era context used for TranslateEra.

This context will be passed to the translation instances of all types of that particular era. In practice, most instances won't need the context, but this approach makes the translation composable (as opposed to having a separate context per type).

type family TranslationError era f :: Type

Most translations should be infallible (default instance), but we leave the door open for partial translations.

For a partial translation, override the default type to be () or a concrete error type.

class (Era era, Era (PreviousEra era)) => TranslateEra era f

Translation of types between eras, e.g., from Shelley to Allegra.

When era is just a phantom type parameter, an empty standalone deriving can be used:

newtype Foo era = Foo Int

instance TranslateEra (Allegra c) Foo

Note that one could use DerivingAnyClass (deriving (TranslateEra (Allegra c))), but this would introduce an undesired coupling between the era-parametric type and (a) particular era(s). The intention is to have a module with orphan instances per era.

In most cases, the era parameter won't be phantom, and a manual instance will have to be written:

newtype Bar era = Bar (TxBody era)

instance CC.Crypto c => TranslateEra (Allegra c) Bar where
    translateEra ctxt = Bar <$> translateEra ctxt

-- With the following instance being in scope:
instance CC.Crypto c => TranslatEra (Allegra c) TxBody

Note: we use PreviousEra instead of NextEra as an era definitely knows its predecessor, but not necessarily its successor. Moreover, one could argue that it makes more sense to define the translation from era A to era B where era B is defined, than where era A is defined.

translateEra :: TranslateEra era f => TranslationContext era -> f (PreviousEra era) -> Except (TranslationError era f) (f era)

Translate a type f parameterised by the era from an era to the era after it.

The translation is a given the translation context of era.

A default instance is provided for when the two types are Coercible.

translateEraMaybe :: (TranslateEra era f, TranslationError era f ~ ()) => TranslationContext era -> f (PreviousEra era) -> Maybe (f era)

Variant of translateEra for when TranslationError is (), converting the result to a Maybe.

translateEra' :: (TranslateEra era f, TranslationError era f ~ Void) => TranslationContext era -> f (PreviousEra era) -> f era

Variant of translateEra for when TranslationError is Void and the translation thus cannot fail.

translateEraThroughCBOR

Arguments

:: forall era ti to. (Era era, ToCBOR (ti (PreviousEra era)), DecCBOR (Annotator (to era)))

=> Text
Label for error reporting

-> ti (PreviousEra era)

-> Except DecoderError (to era)

Translate a type through its binary representation from previous era to the current one.

Deprecations

hashAuxiliaryData :: EraTxAuxData era => TxAuxData era -> AuxiliaryDataHash (EraCrypto era)

Deprecated: Use hashTxAuxData instead

validateAuxiliaryData :: EraTxAuxData era => ProtVer -> TxAuxData era -> Bool

Deprecated: Use validateTxAuxData instead