Cardano.Ledger.Core

Description

This module defines core type families which we know to vary from era to era.

Families in this module should be indexed on era.

It is intended for qualified import: > import qualified Cardano.Ledger.Core as Core

Era-changing types

class (EraTxBody era, EraTxWits era, EraTxAuxData era, EraPParams era, NoThunks (Tx era), DecCBOR (Annotator (Tx era)), EncCBOR (Tx era), ToCBOR (Tx era), Show (Tx era), Eq (Tx era)) ⇒ EraTx era where

A transaction.

Associated Types

type Tx era = (r ∷ Type) | r → era

Methods

mkBasicTx ∷ TxBody era → Tx era

bodyTxL ∷ Lens' (Tx era) (TxBody era)

witsTxL ∷ Lens' (Tx era) (TxWits era)

auxDataTxL ∷ Lens' (Tx era) (StrictMaybe (AuxiliaryData era))

sizeTxF ∷ SimpleGetter (Tx era) Integer

validateScript ∷ PhasedScript 'PhaseOne era → Tx era → Bool

getMinFeeTx ∷ PParams era → Tx era → Coin

class (Val (Value era), ToJSON (TxOut era), DecCBOR (Value era), DecCBOR (CompactForm (Value era)), EncCBOR (Value era), ToCBOR (TxOut era), FromCBOR (TxOut era), EncCBOR (TxOut era), DecCBOR (TxOut era), DecShareCBOR (TxOut era), Share (TxOut era) ~ Interns (Credential 'Staking (EraCrypto era)), NoThunks (TxOut era), NFData (TxOut era), Show (TxOut era), Eq (TxOut era), EraPParams era) ⇒ EraTxOut era where

Abstract interface into specific fields of a TxOut

Minimal complete definition

mkBasicTxOut, valueEitherTxOutL, addrEitherTxOutL, (getMinCoinSizedTxOut | getMinCoinTxOut)

Associated Types

type TxOut era = (r ∷ Type) | r → era

The output of a UTxO for a particular era

Methods

mkBasicTxOut ∷ Addr (EraCrypto era) → Value era → TxOut era

valueTxOutL ∷ Lens' (TxOut era) (Value era)

compactValueTxOutL ∷ Lens' (TxOut era) (CompactForm (Value era))

valueEitherTxOutL ∷ Lens' (TxOut era) (Either (Value era) (CompactForm (Value era)))

Lens for getting and setting in TxOut either an address or its compact version by doing the least amount of work.

addrTxOutL ∷ Lens' (TxOut era) (Addr (EraCrypto era))

compactAddrTxOutL ∷ Lens' (TxOut era) (CompactAddr (EraCrypto era))

addrEitherTxOutL ∷ Lens' (TxOut era) (Either (Addr (EraCrypto era)) (CompactAddr (EraCrypto era)))

Lens for getting and setting in TxOut either an address or its compact version by doing the least amount of work.

The utility of this function comes from the fact that TxOut usually stores the address in either one of two forms: compacted or unpacked. In order to avoid extroneous conversions in getTxOutAddr and getTxOutCompactAddr we can define just this functionality. Also sometimes it is crucial to know at the callsite which form of address we have readily available without any conversions (eg. searching millions of TxOuts for a particular address)

getMinCoinSizedTxOut ∷ PParams era → Sized (TxOut era) → Coin

Produce the minimum lovelace that a given transaction output must contain. Information about the size of the TxOut is required in some eras. Use getMinCoinTxOut if you don't have the size readily available to you.

getMinCoinTxOut ∷ PParams era → TxOut era → Coin

Same as getMinCoinSizedTxOut, except information about the size of TxOut will be computed by serializing the TxOut. If the size turns out to be not needed, then serialization will have no overhead, since it is computed lazily.

bootAddrTxOutF ∷ EraTxOut era ⇒ SimpleGetter (TxOut era) (Maybe (BootstrapAddress (EraCrypto era)))

coinTxOutL ∷ EraTxOut era ⇒ Lens' (TxOut era) Coin

compactCoinTxOutL ∷ EraTxOut era ⇒ Lens' (TxOut era) (CompactForm Coin)

isAdaOnlyTxOutF ∷ EraTxOut era ⇒ SimpleGetter (TxOut era) Bool

This is a getter that implements an efficient way to check whether TxOut contains ADA only.

class (EraTxOut era, EraPParams era, HashAnnotated (TxBody era) EraIndependentTxBody (EraCrypto era), DecCBOR (Annotator (TxBody era)), EncCBOR (TxBody era), ToCBOR (TxBody era), NoThunks (TxBody era), NFData (TxBody era), Show (TxBody era), Eq (TxBody era)) ⇒ EraTxBody era where

Associated Types

type TxBody era = (r ∷ Type) | r → era

The body of a transaction.

Methods

mkBasicTxBody ∷ TxBody era

inputsTxBodyL ∷ Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

outputsTxBodyL ∷ Lens' (TxBody era) (StrictSeq (TxOut era))

feeTxBodyL ∷ Lens' (TxBody era) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody era) (Withdrawals (EraCrypto era))

auxDataHashTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

allInputsTxBodyF ∷ SimpleGetter (TxBody era) (Set (TxIn (EraCrypto era)))

class (Era era, Eq (TxAuxData era), Show (TxAuxData era), NoThunks (TxAuxData era), ToCBOR (TxAuxData era), EncCBOR (TxAuxData era), DecCBOR (Annotator (TxAuxData era)), HashAnnotated (TxAuxData era) EraIndependentTxAuxData (EraCrypto era)) ⇒ EraTxAuxData era where

TxAuxData which may be attached to a transaction

Associated Types

type TxAuxData era = (r ∷ Type) | r → era

Methods

hashTxAuxData ∷ TxAuxData era → AuxiliaryDataHash (EraCrypto era)

validateTxAuxData ∷ ProtVer → TxAuxData era → Bool

class (EraScript era, Eq (TxWits era), Show (TxWits era), Monoid (TxWits era), NoThunks (TxWits era), ToCBOR (TxWits era), EncCBOR (TxWits era), DecCBOR (Annotator (TxWits era))) ⇒ EraTxWits era where

A collection of witnesses in a Tx

Minimal complete definition

addrTxWitsL, bootAddrTxWitsL, scriptTxWitsL

Associated Types

type TxWits era = (r ∷ Type) | r → era

Methods

mkBasicTxWits ∷ TxWits era

addrTxWitsL ∷ Lens' (TxWits era) (Set (WitVKey 'Witness (EraCrypto era)))

bootAddrTxWitsL ∷ Lens' (TxWits era) (Set (BootstrapWitness (EraCrypto era)))

scriptTxWitsL ∷ Lens' (TxWits era) (Map (ScriptHash (EraCrypto era)) (Script era))

class (Era era, Show (Script era), Eq (Script era), ToCBOR (Script era), EncCBOR (Script era), DecCBOR (Annotator (Script era)), NoThunks (Script era), SafeToHash (Script era)) ⇒ EraScript era where

Typeclass for script data types. Allows for script validation and hashing. You must understand the role of SafeToHash and scriptPrefixTag to make new instances. scriptPrefixTag is a magic number representing the tag of the script language. For each new script language defined, a new tag is chosen and the tag is included in the script hash for a script. The safeToHash constraint ensures that Scripts are never reserialised.

Minimal complete definition

scriptPrefixTag, phaseScript

Associated Types

type Script era = (r ∷ Type) | r → era

Scripts which may lock transaction outputs in this era

Methods

scriptPrefixTag ∷ Script era → ByteString

hashScript ∷ Script era → ScriptHash (EraCrypto era)

phaseScript ∷ PhaseRep phase → Script era → Maybe (PhasedScript phase era)

isNativeScript ∷ Script era → Bool

hashScriptTxWitsL ∷ EraTxWits era ⇒ Lens (TxWits era) (TxWits era) (Map (ScriptHash (EraCrypto era)) (Script era)) [Script era]

This is a helper lens that will hash the scripts when adding as witnesses.

type family Value era ∷ Type

A value is something which quantifies a transaction output.

class (Era era, Eq (PParamsHKD Identity era), Ord (PParamsHKD Identity era), Show (PParamsHKD Identity era), NFData (PParamsHKD Identity era), EncCBOR (PParamsHKD Identity era), DecCBOR (PParamsHKD Identity era), ToCBOR (PParamsHKD Identity era), FromCBOR (PParamsHKD Identity era), NoThunks (PParamsHKD Identity era), ToJSON (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Ord (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), NFData (PParamsHKD StrictMaybe era), EncCBOR (PParamsHKD StrictMaybe era), DecCBOR (PParamsHKD StrictMaybe era), ToCBOR (PParamsHKD StrictMaybe era), FromCBOR (PParamsHKD StrictMaybe era), NoThunks (PParamsHKD StrictMaybe era), ToJSON (PParamsHKD StrictMaybe era)) ⇒ EraPParams era where

Minimal complete definition

emptyPParamsIdentity, emptyPParamsStrictMaybe, upgradePParamsHKD, downgradePParamsHKD, hkdMinFeeAL, hkdMinFeeBL, hkdMaxBBSizeL, hkdMaxTxSizeL, hkdMaxBHSizeL, hkdKeyDepositL, hkdPoolDepositL, hkdEMaxL, hkdNOptL, hkdA0L, hkdRhoL, hkdTauL, hkdDL, hkdExtraEntropyL, hkdProtocolVersionL, hkdMinUTxOValueL, hkdMinPoolCostL

Associated Types

type PParamsHKD (f ∷ Type → Type) era = (r ∷ Type) | r → era

Protocol parameters where the fields are represented with a HKD

type UpgradePParams (f ∷ Type → Type) era ∷ Type

type DowngradePParams (f ∷ Type → Type) era ∷ Type

Methods

applyPPUpdates ∷ PParams era → PParamsUpdate era → PParams era

Applies a protocol parameters update

default applyPPUpdates ∷ ∀ a u. (Generic (PParamsHKD Identity era), Generic (PParamsHKD StrictMaybe era), Updatable (Rep (PParamsHKD Identity era) a) (Rep (PParamsHKD StrictMaybe era) u)) ⇒ PParams era → PParamsUpdate era → PParams era

emptyPParamsIdentity ∷ PParamsHKD Identity era

emptyPParamsStrictMaybe ∷ PParamsHKD StrictMaybe era

upgradePParamsHKD ∷ (HKDFunctor f, EraPParams (PreviousEra era)) ⇒ UpgradePParams f era → PParamsHKD f (PreviousEra era) → PParamsHKD f era

Upgrade PParams from previous era to the current one

downgradePParamsHKD ∷ (HKDFunctor f, EraPParams (PreviousEra era)) ⇒ DowngradePParams f era → PParamsHKD f era → PParamsHKD f (PreviousEra era)

Downgrade PParams from the current era to the previous one

hkdMinFeeAL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The linear factor for the minimum fee calculation

hkdMinFeeBL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The constant factor for the minimum fee calculation

hkdMaxBBSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Natural)

Maximal block body size

hkdMaxTxSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Natural)

Maximal transaction size

hkdMaxBHSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Natural)

Maximal block header size

hkdKeyDepositL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a key registration deposit

hkdPoolDepositL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

The amount of a pool registration deposit

hkdEMaxL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f EpochNo)

epoch bound on pool retirement

hkdNOptL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Natural)

Desired number of pools

hkdA0L ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f NonNegativeInterval)

Pool influence

hkdRhoL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f UnitInterval)

Monetary expansion

hkdTauL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f UnitInterval)

Treasury expansion

hkdDL ∷ (HKDFunctor f, ProtVerAtMost era 6) ⇒ Lens' (PParamsHKD f era) (HKD f UnitInterval)

Decentralization parameter

ppDG ∷ SimpleGetter (PParams era) UnitInterval

Decentralization parameter getter

default ppDG ∷ ProtVerAtMost era 6 ⇒ SimpleGetter (PParams era) UnitInterval

hkdExtraEntropyL ∷ (HKDFunctor f, ProtVerAtMost era 6) ⇒ Lens' (PParamsHKD f era) (HKD f Nonce)

Extra entropy

hkdProtocolVersionL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f ProtVer)

Protocol version

hkdMinUTxOValueL ∷ HKDFunctor f ⇒ ProtVerAtMost era 4 ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

Minimum UTxO value

hkdMinPoolCostL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

Minimum Stake Pool Cost

Era STS

type family EraRule (k ∷ Symbol) era ∷ Type

Era STS map

class (Crypto (EraCrypto era), Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT, ProtVerLow era <= MaxVersion, ProtVerHigh era <= MaxVersion) ⇒ Era era

Associated Types

type EraCrypto era ∷ Type

type PreviousEra era = (r ∷ Type) | r → era

Map an era to its predecessor.

For example:

type instance PreviousEra (AllegraEra c) = ShelleyEra c

type ProtVerLow era ∷ Nat

Lowest major protocol version for this era

type ProtVerHigh era ∷ Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

Segregated Witness

The idea of segregated witnessing is to alter the encoding of transactions in a block such that the witnesses (the information needed to verify the validity of the transactions) can be stored separately from the body (the information needed to update the ledger state). In this way, a node which only cares about replaying transactions need not even decode the witness information.

In order to do this, we introduce two concepts: - A TxSeq, which represents the decoded structure of a sequence of transactions as represented in the encoded block; that is, with witnessing, metadata and other non-body parts split separately.

class (EraTx era, Eq (TxSeq era), Show (TxSeq era), EncCBORGroup (TxSeq era), DecCBOR (Annotator (TxSeq era))) ⇒ EraSegWits era where

Indicates that an era supports segregated witnessing.

This class embodies an isomorphism between 'TxSeq era' and 'StrictSeq (Tx era)', witnessed by fromTxSeq and toTxSeq.

Associated Types

type TxSeq era = (r ∷ Type) | r → era

Methods

fromTxSeq ∷ TxSeq era → StrictSeq (Tx era)

toTxSeq ∷ StrictSeq (Tx era) → TxSeq era

hashTxSeq ∷ TxSeq era → Hash (HASH (EraCrypto era)) EraIndependentBlockBody

Get the block body hash from the TxSeq. Note that this is not a regular "hash the stored bytes" function since the block body hash forms a small Merkle tree.

numSegComponents ∷ Word64

The number of segregated components

bBodySize ∷ ∀ era. EraSegWits era ⇒ ProtVer → TxSeq era → Int

Phases

data Phase

There are only two Phases

Constructors

PhaseOne

PhaseTwo

data PhaseRep t where

A GADT that witnesses the two Phases

Constructors

PhaseOneRep ∷ PhaseRep 'PhaseOne

PhaseTwoRep ∷ PhaseRep 'PhaseTwo

type family SomeScript (phase ∷ Phase) (era ∷ Type) ∷ Type

Type family that defines a script type given a Phase and an Era.

data PhasedScript phase era where

A concrete type that witnesses the Phase of SomeScript

Constructors

Phase1Script ∷ SomeScript 'PhaseOne era → PhasedScript 'PhaseOne era

Phase2Script ∷ Language → ShortByteString → PhasedScript 'PhaseTwo era

getPhase1 ∷ EraScript era ⇒ Map k (Script era) → Map k (Script era, PhasedScript 'PhaseOne era)

getPhase2 ∷ EraScript era ⇒ Map k (Script era) → Map k (PhasedScript 'PhaseTwo era)

Rewards

data RewardType

The staking rewards in Cardano are all either:

member rewards - rewards given to a registered stake credential which has delegated to a stake pool, or

leader rewards - rewards given to a registered stake pool (in particular, given to the stake credential in the stake pool registration certificate).

See Figure 47, "Functions used in the Reward Splitting", of the formal specification for more details.

Constructors

MemberReward

LeaderReward

data Reward c

The Reward type captures:

if the reward is a member or leader reward

the stake pool ID associated with the reward

the number of Lovelace in the reward

Constructors

Reward

Fields

rewardType ∷ !RewardType

rewardPool ∷ !(KeyHash 'StakePool c)

rewardAmount ∷ !Coin

Re-exports

module Cardano.Ledger.Hashes

module Cardano.Ledger.Core.Era

module Cardano.Ledger.Core.PParams

module Cardano.Ledger.Core.Translation

Deprecations

hashAuxiliaryData ∷ EraTxAuxData era ⇒ TxAuxData era → AuxiliaryDataHash (EraCrypto era)

Deprecated: Use hashTxAuxData instead

validateAuxiliaryData ∷ EraTxAuxData era ⇒ ProtVer → TxAuxData era → Bool

Deprecated: Use validateTxAuxData instead