Cardano.Ledger.Core.Era

Era

class (Crypto (EraCrypto era), Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT, ProtVerLow era <= MaxVersion, ProtVerHigh era <= MaxVersion) ⇒ Era era

Associated Types

type EraCrypto era ∷ Type

type PreviousEra era = (r ∷ Type) | r → era

Map an era to its predecessor.

For example:

type instance PreviousEra (AllegraEra c) = ShelleyEra c

type ProtVerLow era ∷ Nat

Lowest major protocol version for this era

type ProtVerHigh era ∷ Nat

Highest major protocol version for this era. By default se to ProtVerLow

type ProtVerHigh era = ProtVerLow era

data ByronEra c

This is the era that preceded Shelley era. It cannot have any other class instances, except for Era type class.

Protocol Version

type AtMostEra (eraName ∷ Type → Type) era = ProtVerAtMost era (ProtVerHigh (eraName (EraCrypto era)))

Restrict the era to equal to eraName or come before it.

type AtLeastEra (eraName ∷ Type → Type) era = ProtVerAtLeast era (ProtVerLow (eraName (EraCrypto era)))

Restrict the era to equal to eraName or come after it

type ExactEra (inEra ∷ Type → Type) era = ProtVerInBounds era (ProtVerLow (inEra (EraCrypto era))) (ProtVerHigh (inEra (EraCrypto era)))

Restrict an era to the specific era through the protocol version. This is equivalent to (inEra (Crypto era) ~ era)

type family ProtVerAtMost era (h ∷ Nat) ∷ Constraint where ...

Requirement for the era's lowest protocol version to be lower or equal to the supplied value

Equations

ProtVerAtMost era h = ProtVerIsInBounds "at most" era h (ProtVerLow era <=? h)

type family ProtVerAtLeast era (l ∷ Nat) ∷ Constraint where ...

Requirement for the era's highest protocol version to be higher or equal to the supplied value

Equations

ProtVerAtLeast era l = ProtVerIsInBounds "at least" era l (l <=? ProtVerHigh era)

type ProtVerInBounds era l h = (ProtVerAtLeast era l, ProtVerAtMost era h)

Restrict a lower and upper bounds of the protocol version for the particular era

atLeastEra ∷ AtLeastEra eraName era ⇒ ()

Enforce era to be at least the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

>>> atLeastEra @BabbageEra @(ConwayEra StandardCrypto)
>>> atLeastEra @BabbageEra @(BabbageEra StandardCrypto)

However this will result in a type error

>>> atLeastEra @BabbageEra @(AlonzoEra StandardCrypto)

atMostEra ∷ AtMostEra eraName era ⇒ ()

Enforce era to be at most the specified era at the type level. In other words compiler will produce type error when applied to eras prior to the specified era. This function should be used in order to avoid redundant constraints warning.

For example these will type check

>>> atMostEra @BabbageEra @(ShelleyEra StandardCrypto)
>>> atMostEra @AlonzoEra @(MaryEra StandardCrypto)

However this will result in a type error

>>> atMostEra @BabbageEra @(ConwayEra StandardCrypto)

notSupportedInThisEra ∷ HasCallStack ⇒ a

notSupportedInThisEraL ∷ HasCallStack ⇒ Lens' a b

eraProtVerLow ∷ ∀ era. Era era ⇒ Version

Get the value level Version of the lowest major protocol version for the supplied era.

eraProtVerHigh ∷ ∀ era. Era era ⇒ Version

Get the value level Version of the highest major protocol version for the supplied era.

toEraCBOR ∷ ∀ era t. (Era era, EncCBOR t) ⇒ t → Encoding

Convert a type that implements EncCBOR to plain Encoding using the lowest protocol version for the supplied era

fromEraCBOR ∷ ∀ era t s. (Era era, DecCBOR t) ⇒ Decoder s t

Convert a type that implements DecCBOR to plain Decoder using the lowest protocol version for the supplied era

eraDecoder ∷ ∀ era t s. (Era era, DecCBOR t) ⇒ Decoder s t → Decoder s t

Convert a versioned Decoder to plain a Decoder using the lowest protocol version for the supplied era