Cardano.Ledger.Credential

Documentation

data Credential (kr :: KeyRole) crypto

Script hash or key hash for a payment or a staking object.

Note that credentials (unlike raw key hashes) do appear to vary from era to era, since they reference the hash of a script, which can change. This parameter is a phantom, however, so in actuality the instances will remain the same.

Constructors

ScriptHashObj !(ScriptHash crypto)

KeyHashObj !(KeyHash kr crypto)

newtype GenesisCredential crypto

Constructors

GenesisCredential

Fields

unGenesisCredential :: KeyHash 'Genesis crypto

type PaymentCredential crypto = Credential 'Payment crypto

data Ptr

Pointer to a slot number, transaction index and an index in certificate list.

Constructors

Ptr !SlotNo !TxIx !CertIx

ptrSlotNo :: Ptr -> SlotNo

ptrTxIx :: Ptr -> TxIx

ptrCertIx :: Ptr -> CertIx

type StakeCredential crypto = Credential 'Staking crypto

data StakeReference crypto

Constructors

StakeRefBase !(StakeCredential crypto)

StakeRefPtr !Ptr

StakeRefNull