Cardano.Ledger.Credential

Documentation

data Credential (kr :: KeyRole) c

Script hash or key hash for a payment or a staking object.

Note that credentials (unlike raw key hashes) do appear to vary from era to era, since they reference the hash of a script, which can change. This parameter is a phantom, however, so in actuality the instances will remain the same.

Constructors

ScriptHashObj !(ScriptHash c)

KeyHashObj !(KeyHash kr c)

newtype GenesisCredential c

Constructors

GenesisCredential

Fields

unGenesisCredential :: KeyHash 'Genesis c

type PaymentCredential c = Credential 'Payment c

credKeyHashWitness :: Credential r c -> Maybe (KeyHash 'Witness c)

Convert a KeyHash into a Witness KeyHash. Does nothing for Script credentials.

credScriptHash :: Credential kr c -> Maybe (ScriptHash c)

Extract ScriptHash from a Credential. Returns Nothing for KeyHashes

credToText :: Credential kr c -> Text

parseCredential :: (MonadFail m, Crypto c) => Text -> m (Credential kr c)

data Ptr

Pointer to a slot number, transaction index and an index in certificate list.

Constructors

Ptr !SlotNo !TxIx !CertIx

ptrSlotNo :: Ptr -> SlotNo

ptrTxIx :: Ptr -> TxIx

ptrCertIx :: Ptr -> CertIx

type StakeCredential c = Credential 'Staking c

data StakeReference c

Constructors

StakeRefBase !(StakeCredential c)

StakeRefPtr !Ptr

StakeRefNull

normalizePtr :: Ptr -> Ptr

Convert any invalid Ptr to a Ptr that contains all zeros for its fields. Any pointer that contains a SlotNo, TxIx or CertIx that is too large to fit into Word32, Word16 and Word16 respectively is considered to be an invalid Ptr. Valid Ptrs will be returned unmodified.

Note - This is in no way related to dangling pointers, with an exception that any invalid Ptr is guarateed to be a dangling Ptr.