Cardano.Ledger.EpochBoundary

Description

This modules implements the necessary functions for the changes that can happen at epoch boundaries.

Documentation

newtype Stake c

Type of stake as map from hash key to coins associated.

Constructors

Stake

Fields

unStake ∷ VMap VB VP (Credential 'Staking c) (CompactForm Coin)

sumAllStake ∷ Stake c → Coin

sumAllStakeCompact ∷ Stake c → CompactForm Coin

sumStakePerPool ∷ VMap VB VB (Credential 'Staking c) (KeyHash 'StakePool c) → Stake c → Map (KeyHash 'StakePool c) Coin

data SnapShot c

Snapshot of the stake distribution.

Constructors

SnapShot

Fields

ssStake ∷ !(Stake c)

ssDelegations ∷ !(VMap VB VB (Credential 'Staking c) (KeyHash 'StakePool c))

ssPoolParams ∷ !(VMap VB VB (KeyHash 'StakePool c) (PoolParams c))

data SnapShots c

Snapshots of the stake distribution.

Note that ssStakeMark and ssStakeMarkPoolDistr are lazy on purpose since we only want to force the thunk after one stability window when we know that they are stable (so that we do not compute them if we do not have to). See more info in the Optimize TICKF ADR

Constructors

SnapShots

Fields

ssStakeMark ∷ SnapShot c

ssStakeMarkPoolDistr ∷ PoolDistr c

ssStakeSet ∷ !(SnapShot c)

ssStakeGo ∷ !(SnapShot c)

ssFee ∷ !Coin

emptySnapShot ∷ SnapShot c

emptySnapShots ∷ SnapShots c

poolStake ∷ KeyHash 'StakePool c → VMap VB VB (Credential 'Staking c) (KeyHash 'StakePool c) → Stake c → Stake c

Get stake of one pool

maxPool ∷ EraPParams era ⇒ PParams era → Coin → Rational → Rational → Coin

Version of maxPool` that extracts ppA0L and ppNOptL from a PParams

maxPool' ∷ NonNegativeInterval → Natural → Coin → Rational → Rational → Coin

Calculate maximal pool reward

calculatePoolDistr ∷ SnapShot c → PoolDistr c

calculatePoolDistr' ∷ ∀ c. (KeyHash 'StakePool c → Bool) → SnapShot c → PoolDistr c

calculatePoolStake ∷ (KeyHash 'StakePool c → Bool) → VMap VB VB (Credential 'Staking c) (KeyHash 'StakePool c) → Stake c → Map (KeyHash 'StakePool c) Word64

Sum up the Coin (as CompactForm Coin = Word64) for each StakePool

ssStakeMarkL ∷ Lens' (SnapShots c) (SnapShot c)

ssStakeMarkPoolDistrL ∷ Lens' (SnapShots c) (PoolDistr c)

ssStakeSetL ∷ Lens' (SnapShots c) (SnapShot c)

ssStakeGoL ∷ Lens' (SnapShots c) (SnapShot c)

ssFeeL ∷ Lens' (SnapShots c) Coin

ssStakeL ∷ Lens' (SnapShot c) (Stake c)

ssStakeDistrL ∷ Lens' (SnapShot c) (VMap VB VP (Credential 'Staking c) (CompactForm Coin))

ssDelegationsL ∷ Lens' (SnapShot c) (VMap VB VB (Credential 'Staking c) (KeyHash 'StakePool c))

ssPoolParamsL ∷ Lens' (SnapShot c) (VMap VB VB (KeyHash 'StakePool c) (PoolParams c))