Cardano.Ledger.Hashes

Hashing algorithms

data Hash h a

class (KnownNat (SizeHash h), Typeable h) ⇒ HashAlgorithm h

Minimal complete definition

hashAlgorithmName, digest

type HASH = Blake2b_256

Hashing algorithm used for hashing everything, except addresses, for which ADDRHASH is used.

type ADDRHASH = Blake2b_224

Hashing algorithm used for hashing cryptographic keys and scripts. As the type synonym name alludes, this is the hashing algorithm used for addresses.

Era-independent hash type identifiers.

Various identifiers in the ledger are hashes of particular structures. While the structures may change from era to era, the hash will remain the same, and we can refer to the hash of, say, a transaction, without knowing the actual transaction type. As such, we define a number of these hashes here.

data EraIndependentTxBody

data EraIndependentBlockHeader

data EraIndependentBlockBody

data EraIndependentMetadata

data EraIndependentScript

data EraIndependentData

data EraIndependentScriptData

data EraIndependentTxAuxData

data EraIndependentPParamView

data EraIndependentScriptIntegrity

Hashes

`DSIGN Verification Key Hashes

newtype KeyHash (r ∷ KeyRole)

Discriminated hash of public Key

Constructors

KeyHash

Fields

unKeyHash ∷ Hash ADDRHASH (VerKeyDSIGN DSIGN)

data KeyRole

The role of a key.

All key roles are fixed and unique, except for the Witness role. In particular, keys can be cast to a Witness role with the help of asWitness, because same witness can be valid for many roles.

In fact, it is perfectly allowable for a key to be used in many roles by the end user; there is nothing prohibiting somebody using the same underlying key or a script as their payment and staking credential, as well as the key for their stake pool. However, in the ledger code mixing up keys with different roles could be catastrophic, that is why we have this separation.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

DRepRole

HotCommitteeRole

ColdCommitteeRole

Guard

hashKey ∷ VKey kd → KeyHash kd

Hash a given public key

hashTxBodySignature ∷ SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody) → Hash HASH (SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody))

Hash a given signature

Script Hashes

newtype ScriptHash

Constructors

ScriptHash (Hash ADDRHASH EraIndependentScript)

type DataHash = SafeHash EraIndependentData

AuxiliaryData

newtype TxAuxDataHash

Constructors

TxAuxDataHash

Fields

unTxAuxDataHash ∷ SafeHash EraIndependentTxAuxData

VRF Verification Key Hashes

data KeyRoleVRF

Constructors

StakePoolVRF

GenDelegVRF

BlockIssuerVRF

newtype VRFVerKeyHash (r ∷ KeyRoleVRF)

Discriminated hash of VRF Verification Key

Constructors

VRFVerKeyHash

Fields

unVRFVerKeyHash ∷ Hash HASH KeyRoleVRF

toVRFVerKeyHash ∷ Hash HASH (VerKeyVRF v) → VRFVerKeyHash (r ∷ KeyRoleVRF)

fromVRFVerKeyHash ∷ VRFVerKeyHash (r ∷ KeyRoleVRF) → Hash HASH (VerKeyVRF v)

Genesis DSIGN and VRF Verification Key Hashes

data GenDelegPair

Constructors

GenDelegPair

Fields

genDelegKeyHash ∷ !(KeyHash 'GenesisDelegate)

genDelegVrfHash ∷ !(VRFVerKeyHash 'GenDelegVRF)

newtype GenDelegs

Constructors

GenDelegs

Fields

unGenDelegs ∷ Map (KeyHash 'Genesis) GenDelegPair

SafeHash

In cardano-ledger, hashing a type X is based upon the serialization of X. Serialization is based upon the EncCBOR and DecCBOR type classes, and the serialization that can be handled by DecCBOR instances for a particular type, are not necessarily unique. For this reason, when an X object comes over the network in serialized form, we must preserve the original bytes that arrived over the network, otherwise when the system hashes that object, the hash in the ledger, and the hash of that object from the other side of the network may not agree. In otherwords reserialization for the purpose of hash calculation is not an option. The SafeToHash type class ensures that types with a (SafeToHash X) instance store the original bytes that arrived over the network for the value of X. The recommended way to store the original bytes is to use the type MemoBytes, although there are a few types that store their original bytes in other ways. In order to encourage the use of newtype over MemoBytes newtype defined as a MemoBytes, which would get the functionality of retaining bytes and deriving of SafeToHash instance for free.

data SafeHash i

A SafeHash is a hash of something that is safe to hash. Such types store their own serialisation bytes. The prime example is (MemoBytes t), but other examples are things that consist of only ByteStrings (i.e. they are their own serialization) or for some other reason store their original bytes.

We do NOT export the constructor SafeHash, but instead export other functions such as hashAnnotated and extractHash which have constraints that limit their application to types which preserve their original serialization bytes.

class SafeToHash t where

Only Types that preserve their serialisation bytes are members of the class SafeToHash. There are only a limited number of primitive direct instances of SafeToHash, all but two of them are present in this file. Instead of making explicit instances, we almost always use a newtype (around a type S) where their is already an instance (SafeToHash S). In that case the newtype has its SafeToHash instance derived using newtype deriving. The prime example of s is MemoBytes. The only exceptions are the legacy Shelley types: Metadata and ShelleyTx, that preserve their serialization bytes using a different mechanism than the use of MemoBytes. SafeToHash is a superclass requirement of the classes HashAnnotated which provide more convenient ways to construct SafeHashes than using makeHashWithExplicitProxys.

Minimal complete definition

originalBytes

Methods

originalBytes ∷ t → ByteString

Extract the original bytes from t

originalBytesSize ∷ t → Int

makeHashWithExplicitProxys ∷ Proxy i → t → SafeHash i

Header Hash

newtype HashHeader

Constructors

HashHeader

Fields

unHashHeader ∷ Hash HASH EraIndependentBlockHeader

Creating SafeHash

class SafeToHash x ⇒ HashAnnotated x i | x → i

Types that are SafeToHash AND have the type uniquely determines the index type tag of SafeHash index

The SafeToHash and the HashAnnotated classes are designed so that their instances can be easily derived (because their methods have default methods when the type is a newtype around a type that is SafeToHash). For example,

hashAnnotated ∷ HashAnnotated x i ⇒ x → SafeHash i

Create a (SafeHash i), given (HashAnnotated x i) instance.

unsafeMakeSafeHash ∷ Hash HASH i → SafeHash i

Don't use this except in Testing to make Arbitrary instances, etc. or in cases when it can be guaranteed that original bytes were used for computing the hash.

Other operations

castSafeHash ∷ ∀ i j. SafeHash i → SafeHash j

To change the index parameter of SafeHash (which is a phantom type) use castSafeHash

extractHash ∷ SafeHash i → Hash HASH i

Extract the hash out of a SafeHash

standardHashSize ∷ Int

standardAddrHashSize ∷ Int