Cardano.Ledger.Hashes

Description

Various identifiers in the ledger are hashes of particular structures. While the structures may change from era to era, the hash will remain the same, and we can refer to the hash of, say, a transaction, without knowing the actual transaction type. As such, we define a number of these hashes here.

Era-independent hash type identifiers.

data EraIndependentTxBody

data EraIndependentBlockHeader

data EraIndependentBlockBody

data EraIndependentMetadata

data EraIndependentScript

data EraIndependentData

data EraIndependentScriptData

data EraIndependentAuxiliaryData

data EraIndependentPParamView

data EraIndependentScriptIntegrity

Script hashes

newtype ScriptHash crypto

Constructors

ScriptHash (Hash (ADDRHASH crypto) EraIndependentScript)

type DataHash crypto = SafeHash crypto EraIndependentData