Cardano.Ledger.Keys

VKey

newtype VKey (kd ∷ KeyRole)

Discriminated verification key

We wrap the basic VerKeyDSIGN in order to add the key role.

Constructors

VKey

Fields

unVKey ∷ VerKeyDSIGN DSIGN

Key Roles

data KeyRole

The role of a key.

All key roles are fixed and unique, except for the Witness role. In particular, keys can be cast to a Witness role with the help of asWitness, because same witness can be valid for many roles.

In fact, it is perfectly allowable for a key to be used in many roles by the end user; there is nothing prohibiting somebody using the same underlying key or a script as their payment and staking credential, as well as the key for their stake pool. However, in the ledger code mixing up keys with different roles could be catastrophic, that is why we have this separation.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

DRepRole

HotCommitteeRole

ColdCommitteeRole

Guard

class HasKeyRole (a ∷ KeyRole → Type) where

Minimal complete definition

Nothing

Methods

coerceKeyRole ∷ a r → a r'

General coercion of key roles.

The presence of this function is mostly to help the user realise where they are converting key roles.

default coerceKeyRole ∷ Coercible (a r) (a r') ⇒ a r → a r'

asWitness ∷ HasKeyRole a ⇒ a r → a 'Witness

Use a key as a witness.

This is the most common coercion between key roles, because most keys can be used as witnesses to some types of transaction. As such, we provide an explicit coercion for it.

Key Hashes

newtype KeyHash (r ∷ KeyRole)

Discriminated hash of public Key

Constructors

KeyHash

Fields

unKeyHash ∷ Hash ADDRHASH (VerKeyDSIGN DSIGN)

hashKey ∷ VKey kd → KeyHash kd

Hash a given public key

Signature

type DSIGN = Ed25519DSIGN

Cryptographic signing algorithm used on Cardano blockchain.

type DSignable = Signable DSIGN

signedDSIGN ∷ Signable DSIGN a ⇒ SignKeyDSIGN DSIGN → a → SignedDSIGN DSIGN a

Produce a digital signature

verifySignedDSIGN ∷ Signable DSIGN a ⇒ VKey kd → a → SignedDSIGN DSIGN a → Bool

Verify a digital signature

hashSignature ∷ SignedDSIGN (Hash h) → Hash (SignedDSIGN (Hash h))

Deprecated: In favor of hashTxBodySignature. Fallback on hashWith if you need more general hashing functionality.

Hash a given signature

VRF Key Hashes

data KeyRoleVRF

Constructors

StakePoolVRF

GenDelegVRF

BlockIssuerVRF

newtype VRFVerKeyHash (r ∷ KeyRoleVRF)

Discriminated hash of VRF Verification Key

Constructors

VRFVerKeyHash

Fields

unVRFVerKeyHash ∷ Hash HASH KeyRoleVRF

toVRFVerKeyHash ∷ Hash HASH (VerKeyVRF v) → VRFVerKeyHash (r ∷ KeyRoleVRF)

fromVRFVerKeyHash ∷ VRFVerKeyHash (r ∷ KeyRoleVRF) → Hash HASH (VerKeyVRF v)

Genesis delegations

data GenDelegPair

Constructors

GenDelegPair

Fields

genDelegKeyHash ∷ !(KeyHash 'GenesisDelegate)

genDelegVrfHash ∷ !(VRFVerKeyHash 'GenDelegVRF)

newtype GenDelegs

Constructors

GenDelegs

Fields

unGenDelegs ∷ Map (KeyHash 'Genesis) GenDelegPair

module Cardano.Ledger.Keys.WitVKey

module Cardano.Ledger.Keys.Bootstrap

To be removed

Re-exports from cardano-crypto-class

decodeSignedDSIGN ∷ DSIGNAlgorithm v ⇒ Decoder s (SignedDSIGN v a)

encodeSignedDSIGN ∷ DSIGNAlgorithm v ⇒ SignedDSIGN v a → Encoding

hashWithSerialiser ∷ HashAlgorithm h ⇒ (a → Encoding) → a → Hash h a

A variation on hashWith, but specially for CBOR encodings.

Deprecated

type Hash = Hash HASH

Deprecated: In favor of Hash HASH

DSIGN

type SignedDSIGN = SignedDSIGN DSIGN

Deprecated: In favor of SignedDSIGN DSIGN

type SignKeyDSIGN = SignKeyDSIGN DSIGN

Deprecated: In favor of SignKeyDSIGN DSIGN

KES

type KESignable c = Signable (KES c)

Deprecated: In favor of Signable (KES c)

type SignedKES c = SignedKES (KES c)

Deprecated: In favor of SignedKES (KES c)

type SignKeyKES c = SignKeyKES (KES c)

Deprecated: In favor of SignKeyKES (KES c)

type VerKeyKES c = VerKeyKES (KES c)

Deprecated: In favor of VerKeyKES (KES c)

VRF

type VRFSignable c = Signable (VRF c)

Deprecated: In favor of Signable (VRF c)

type CertifiedVRF c = CertifiedVRF (VRF c)

Deprecated: In favor of CertifiedVRF (VRF c)

type SignKeyVRF c = SignKeyVRF (VRF c)

Deprecated: In favor of SignKeyVRF (VRF c)

type VerKeyVRF c = VerKeyVRF (VRF c)

Deprecated: In favor of VerKeyVRF (VRF c)

hashVerKeyVRF ∷ Crypto c ⇒ VerKeyVRF (VRF c) → VRFVerKeyHash (r ∷ KeyRoleVRF)

Deprecated: Use hashVerKeyVRF instead