Cardano.Ledger.Keys

Documentation

data KeyRole

The role of a key.

Note that a role is not _fixed_, nor is it unique. In particular, keys may variously be used as witnesses, and so in many case we will change the role of a key to the Witness role.

It is also perfectly allowable for a key to be used in many roles; there is nothing prohibiting somebody using the same underlying key as their payment and staking key, as well as the key for their stake pool. So these roles are more intended for two purposes:

To make explicit how we are using a key in the specifications

To provide a guide to downstream implementors, for whom the profusion of keys may be confusing.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

DRepRole

HotCommitteeRole

ColdCommitteeRole

class HasKeyRole (a :: KeyRole -> Type -> Type) where

Minimal complete definition

Nothing

Methods

coerceKeyRole :: a r c -> a r' c

General coercion of key roles.

The presence of this function is mostly to help the user realise where they are converting key roles.

default coerceKeyRole :: Coercible (a r c) (a r' c) => a r c -> a r' c

asWitness :: HasKeyRole a => a r c -> a 'Witness c

Use a key as a witness.

This is the most common coercion between key roles, because most keys can be used as witnesses to some types of transaction. As such, we provide an explicit coercion for it.

DSIGN

type DSignable c = Signable (DSIGN c)

newtype VKey (kd :: KeyRole) c

Discriminated verification key

We wrap the basic VerKeyDSIGN in order to add the key role.

Constructors

VKey

Fields

unVKey :: VerKeyDSIGN (DSIGN c)

data KeyPair (kd :: KeyRole) c

Deprecated: Use `Test.Cardano.Ledger.Core.KeyPair (KeyPair)` instead

Constructors

KeyPair

Deprecated: Use `Test.Cardano.Ledger.Core.KeyPair (KeyPair)` instead

Fields

vKey :: !(VKey kd c)

sKey :: !(SignKeyDSIGN (DSIGN c))

signedDSIGN :: (Crypto c, Signable (DSIGN c) a) => SignKeyDSIGN (DSIGN c) -> a -> SignedDSIGN c a

Produce a digital signature

verifySignedDSIGN :: (Crypto c, Signable (DSIGN c) a) => VKey kd c -> a -> SignedDSIGN c a -> Bool

Verify a digital signature

hashSignature :: Crypto c => SignedDSIGN c (Hash c h) -> Hash c (SignedDSIGN c (Hash c h))

Hash a given signature

Key hashes

newtype KeyHash (discriminator :: KeyRole) c

Discriminated hash of public Key

Constructors

KeyHash (Hash (ADDRHASH c) (VerKeyDSIGN (DSIGN c)))

hashKey :: Crypto c => VKey kd c -> KeyHash kd c

Hash a given public key

Genesis delegations

data GenDelegPair c

Constructors

GenDelegPair

Fields

genDelegKeyHash :: !(KeyHash 'GenesisDelegate c)

genDelegVrfHash :: !(Hash c (VerKeyVRF c))

newtype GenDelegs c

Constructors

GenDelegs

Fields

unGenDelegs :: Map (KeyHash 'Genesis c) (GenDelegPair c)

newtype GKeys c

Constructors

GKeys

Fields

unGKeys :: Set (VKey 'Genesis c)

KES

type KESignable c = Signable (KES c)

VRF

type VRFSignable c = Signable (VRF c)

Re-exports from cardano-crypto-class

decodeSignedDSIGN :: DSIGNAlgorithm v => Decoder s (SignedDSIGN v a)

encodeSignedDSIGN :: DSIGNAlgorithm v => SignedDSIGN v a -> Encoding

hashWithSerialiser :: HashAlgorithm h => (a -> Encoding) -> a -> Hash h a

decodeSignedKES :: KESAlgorithm v => Decoder s (SignedKES v a)

decodeVerKeyKES :: KESAlgorithm v => Decoder s (VerKeyKES v)

encodeSignedKES :: KESAlgorithm v => SignedKES v a -> Encoding

encodeVerKeyKES :: KESAlgorithm v => VerKeyKES v -> Encoding

signedKES :: (KESAlgorithm v, Signable v a) => ContextKES v -> Period -> a -> SignKeyKES v -> SignedKES v a

updateKES :: KESAlgorithm v => ContextKES v -> SignKeyKES v -> Period -> Maybe (SignKeyKES v)

verifyKES :: (KESAlgorithm v, Signable v a, HasCallStack) => ContextKES v -> VerKeyKES v -> Period -> a -> SigKES v -> Either String ()

verifySignedKES :: (KESAlgorithm v, Signable v a) => ContextKES v -> VerKeyKES v -> Period -> a -> SignedKES v a -> Either String ()

decodeVerKeyVRF :: VRFAlgorithm v => Decoder s (VerKeyVRF v)

encodeVerKeyVRF :: VRFAlgorithm v => VerKeyVRF v -> Encoding

hashVerKeyVRF :: (VRFAlgorithm v, HashAlgorithm h) => VerKeyVRF v -> Hash h (VerKeyVRF v)

verifyVRF :: (VRFAlgorithm v, HasCallStack, Signable v a) => ContextVRF v -> VerKeyVRF v -> a -> CertVRF v -> Maybe (OutputVRF v)

Re-parametrised types over crypto

type CertifiedVRF c = CertifiedVRF (VRF c)

type Hash c = Hash (HASH c)

type SignedDSIGN c = SignedDSIGN (DSIGN c)

type SignKeyDSIGN c = SignKeyDSIGN (DSIGN c)

type SignedKES c = SignedKES (KES c)

type SignKeyKES c = SignKeyKES (KES c)

type SignKeyVRF c = SignKeyVRF (VRF c)

type VerKeyKES c = VerKeyKES (KES c)

type VerKeyVRF c = VerKeyVRF (VRF c)

module Cardano.Ledger.Keys.WitVKey

module Cardano.Ledger.Keys.Bootstrap