Cardano.Ledger.Language

Description

This module provides data structures and operations for talking about Non-native Script languages. It is expected that new languages (or new versions of old languages) will be added here.

Documentation

data Language

Non-Native Script language. This is an Enumerated type. This is expected to be an open type. We will add new Constuctors to this type as additional Non-Native scripting language as are added. We use an enumerated type for two reasons. 1) We can write total functions by case analysis over the constructors 2) We will use DataKinds to make some datatypes indexed by Language For now, the only Non-Native Scriting language is Plutus We might add new languages in the futures.

Note that the the serialization of Language depends on the ordering.

Constructors

PlutusV1

PlutusV2

mkLanguageEnum ∷ Int → Maybe Language

Make a language form its Enum index.

languageToText ∷ Language → Text

languageFromText ∷ MonadFail m ⇒ Text → m Language

nonNativeLanguages ∷ [Language]

data SLanguage (l ∷ Language) where

Singleton for 'Language'

Constructors

SPlutusV1 ∷ SLanguage 'PlutusV1

SPlutusV2 ∷ SLanguage 'PlutusV2

fromSLanguage ∷ SLanguage l → Language

Reflection for 'SLanguage'

class Typeable l ⇒ IsLanguage (l ∷ Language) where

For implicit reflection on 'SLanguage' See Cardano.Ledger.Alonzo.TxInfo for example usage

Methods

isLanguage ∷ SLanguage l

toSLanguage ∷ ∀ l m. (IsLanguage l, MonadFail m) ⇒ Language → m (SLanguage l)