Cardano.Ledger.MemoBytes

Description

MemoBytes is an abstraction for a data type that encodes its own serialization. The idea is to use a newtype around a MemoBytes applied to a non-memoizing type. For example: newtype Foo = Foo (MemoBytes NonMemoizingFoo) This way all the instances for Foo (Eq, Show, EncCBOR, DecCBOR, NoThunks, Generic`) can be derived for free. MemoBytes plays an important role in the SafeToHash class introduced in the module SafeHash

Documentation

data MemoBytes t era where

Pair together a type t and its serialization. Used to encode a type that is serialized over the network, and to remember the original bytes that were used to transmit it. Important since hashes are computed from the serialization of a type, and EncCBOR instances do not have unique serializations.

Bundled Patterns

pattern Memo :: Era era => t era -> ShortByteString -> MemoBytes t era

type family MemoHashIndex (t :: Type -> Type) :: Type

type Mem t era = Annotator (MemoBytes t era)

Useful when deriving DecCBOR(Annotator T) deriving via (Mem T) instance (Era era) => DecCBOR (Annotator T)

mkMemoBytes :: forall era t. Era era => t era -> ByteString -> MemoBytes t era

Smart constructor

getMemoBytesType :: MemoBytes t era -> t era

Extract the inner type of the MemoBytes

getMemoBytesHash :: MemoBytes t era -> SafeHash (EraCrypto era) (MemoHashIndex t)

Extract the hash value of the binary representation of the MemoBytes

memoBytes :: forall era w t. Era era => Encode w (t era) -> MemoBytes t era

Create MemoBytes from its CBOR encoding

shorten :: ByteString -> ShortByteString

Turn a lazy bytestring into a short bytestring.

showMemo :: Show (t era) => MemoBytes t era -> String

Turn a MemoBytes into a string, Showing both its internal structure and its original bytes. Useful since the Show instance of MemoBytes does not display the original bytes.

printMemo :: Show (t era) => MemoBytes t era -> IO ()

shortToLazy :: ShortByteString -> ByteString

Helper function. Converts a short bytestring to a lazy bytestring.

contentsEq :: Eq (t era) => MemoBytes t era -> MemoBytes t era -> Bool

Returns true if the contents of the MemoBytes are equal

Memoized

class Memoized t

Class that relates the actual type with its raw and byte representations

Associated Types

type RawType t = (r :: Type -> Type) | r -> t

mkMemoized :: forall era t. (Era era, EncCBOR (RawType t era), Memoized t) => RawType t era -> t era

Construct memoized type from the raw type using its EncCBOR instance

getMemoSafeHash :: Memoized t => t era -> SafeHash (EraCrypto era) (MemoHashIndex (RawType t))

Extract memoized SafeHash

getMemoRawType :: Memoized t => t era -> RawType t era

Extract the raw type from the memoized version

zipMemoRawType :: (Memoized t1, Memoized t2) => (RawType t1 era -> RawType t2 era -> a) -> t1 era -> t2 era -> a

This is a helper function that operates on raw types of two memoized types.

eqRawType :: forall t era. (Memoized t, Eq (RawType t era)) => t era -> t era -> Bool

getMemoRawBytes :: Memoized t => t era -> ShortByteString

Extract the raw bytes from the memoized version

lensMemoRawType :: (Era era, EncCBOR (RawType t era), Memoized t) => (RawType t era -> a) -> (RawType t era -> b -> RawType t era) -> Lens (t era) (t era) a b

This is a helper Lens creator for any Memoized type.

getterMemoRawType :: Memoized t => (RawType t era -> a) -> SimpleGetter (t era) a

This is a helper SimpleGetter creator for any Memoized type

Raw equality

class EqRaw a where

Type class that implements equality on the Haskell type, ignoring any of the potentially memoized binary representation of the type.

Minimal complete definition

Nothing

Methods

eqRaw :: a -> a -> Bool

default eqRaw :: (a ~ t era, Memoized t, Eq (RawType t era)) => a -> a -> Bool