Cardano.Ledger.Plutus.CostModels

Cost Model

data CostModel

A language dependent cost model for the Plutus evaluator. Note that the EvaluationContext is entirely dependent on the cost model parameters (ie the Decode Text Integer) and that this type uses the smart constructor mkCostModel to hide the evaluation context.

newtype CostModelError

See CostModels for an explanation of how CostModelError is used.

Constructors

CostModelError CostModelApplyError

data CostModelApplyError

Constructors

CMUnknownParamError !Text

CMInternalReadError

CMInternalWriteError !String

CMTooFewParamsError

Fields

cmTooFewExpected :: !Int

cmTooFewActual :: !Int

mkCostModel :: Language -> [Integer] -> Either CostModelApplyError CostModel

Convert cost model parameters to a cost model, making use of the conversion function mkEvaluationContext from the Plutus API.

mkCostModelsLenient :: Map Word8 [Integer] -> CostModels

This function attempts to convert a Map with potential cost models to into validated CostModels. If it is a valid cost model for a known version of Plutus, it is added to costModelsValid. If it is an invalid cost model for a known version of Plutus, the error is stored in costModelsErrors and the cost model is stored in costModelsUnknown. Lastly, if the Plutus version is unknown, the cost model is also stored in costModelsUnknown.

encodeCostModel :: CostModel -> Encoding

Encoding for the CostModel. Important to note that it differs from Encoding used by getLanguageView

getCostModelLanguage :: CostModel -> Language

getCostModelParams :: CostModel -> [Integer]

getCostModelEvaluationContext :: CostModel -> EvaluationContext

getEvaluationContext :: CostModel -> EvaluationContext

costModelParamNames :: Language -> [Text]

costModelToMap :: CostModel -> Map Text Integer

costModelFromMap :: MonadFail m => Language -> Map Text Integer -> m CostModel

costModelParamsCount :: Language -> Int

Number of parameters in a CostModel for a specific language

decodeCostModelFailHard :: Language -> Decoder s CostModel

Cost Models

data CostModels

For a known version of Plutus, attempting to construct a cost model with too few parameters (depending on the version) will result in an error. CostModelApplyError exists to collect these errors in the CostModels type. The CostModels type itself needs to be flexible enough to accept any map of Word8 to '[Integer]', so that cost models can be placed in the protocol parameters ahead of changes to the Plutus evaluation context. In this way, serializing a cost model, updating software, and deserializing can result in errors going away.

Additionally, CostModels needs to be able to store cost models for future version of Plutus, which we cannot yet even validate. These are stored in costModelsUnknown.

mkCostModels :: Map Language CostModel -> CostModels

Construct an all valid CostModels

emptyCostModels :: CostModels

updateCostModels

Arguments

:: CostModels
Old CostModels that will be overwritten

-> CostModels
New CostModels that will overwrite

-> CostModels

Updates the first CostModels with the second one, so that only the cost models that are present in the second one get updated while all the others stay unchanged. Language specific errors and unknown cost models are removed, whenever a valid CostModel for the language is supplied in the update.

decodeValidAndUnknownCostModels :: Decoder s CostModels

costModelsValid :: CostModels -> Map Language CostModel

costModelsErrors :: CostModels -> Map Language CostModelError

costModelsUnknown :: CostModels -> Map Word8 [Integer]

flattenCostModels :: CostModels -> Map Word8 [Integer]

Turn a CostModels into a mapping of potential language versions and cost model values, with no distinction between valid and invalid cost models. This is used for serialization, so that judgements about validity can be made upon deserialization.