Cardano.Ledger.Plutus.Evaluate

Documentation

data PlutusWithContext where

This type contains all that is necessary from Ledger to evaluate a plutus script.

Constructors

PlutusWithContext

Fields

∷ PlutusLanguage l

⇒ { pwcProtocolVersion ∷ !Version

Major protocol version that is necessary for [de]serialization

, pwcScript ∷ !(Either (Plutus l) (PlutusRunnable l))

Actual plutus script that will be evaluated. Script is allowed to be in two forms: serialized and deserialized. This is necesary for implementing the opptimization that preserves deserialized PlutusRunnable after verifying wellformedness of plutus scripts during transaction validation (yet to be implemented).

, pwcScriptHash ∷ !ScriptHash

Hash of the above script as it would appear on-chain. In other words it is not just a hash of the script contents. (See hashScript for more info)

, pwcArgs ∷ !(PlutusArgs l)

All of the arguments to the Plutus scripts, including the redeemer and the Plutus context that was obtained from the transaction translation

, pwcExUnits ∷ !ExUnits

Limit on the execution units

, pwcCostModel ∷ !CostModel

CostModel to be used during script evaluation. It must match the language version in the pwcScript

} → PlutusWithContext

data ScriptFailure

Constructors

ScriptFailure

Fields

scriptFailureMessage ∷ Text

scriptFailurePlutus ∷ PlutusWithContext

data ScriptResult

Constructors

Passes [PlutusWithContext]

Fails [PlutusWithContext] (NonEmpty ScriptFailure)

scriptPass ∷ PlutusWithContext → ScriptResult

scriptFail ∷ ScriptFailure → ScriptResult

data PlutusDebugInfo

Constructors

DebugBadHex String

DebugCannotDecode String

DebugSuccess

Fields

[Text]

Execution logs from the plutus interpreter

ExBudget

Execution budget that was consumed. It will always be less or equal to what was supplied during execution.

DebugFailure

Fields

[Text]

Execution logs from the plutus interpreter

EvaluationError

Evaluation error from Plutus interpreter

PlutusWithContext

Everything that is needed in order to run the script

(Maybe ExBudget)

Expected execution budget. This value is Nothing when the supplied script can't be executed within 5 second limit or there is a problem with decoding plutus script itself.

DebugTimedOut
Script did not terminate within the imposed limit

Fields

Int

Wall clock limit in microseconds that was imposed on the script execution.

data PlutusDebugOverrides

Various overrides that can be supplied to plutusDebug and plutusDebugUnbouded

Constructors

PlutusDebugOverrides

Fields

pdoScript ∷ !(Maybe ByteString)

Hex encoded version of the script

pdoProtocolVersion ∷ !(Maybe Version)

Protocol version to be used for decoding and exection

pdoLanguage ∷ !(Maybe Language)

Plutus ledger language version

pdoCostModelValues ∷ !(Maybe [Int64])

Cost model to be used for deciding execution units

pdoExUnitsMem ∷ !(Maybe Natural)

Memory execution units to be used for execution

pdoExUnitsSteps ∷ !(Maybe Natural)

CPU execution units to be used for execution

pdoExUnitsEnforced ∷ !Bool

Setting this flag to True will disable reporting expected execution units upon a failure, which would protect against a potentially unbounded script execution.

defaultPlutusDebugOverrides ∷ PlutusDebugOverrides

debugPlutus ∷ HasCallStack ⇒ String → Int → PlutusDebugOverrides → IO PlutusDebugInfo

Execute a hex encoded script with the context that was produced within the ledger predicate failure. Using PlutusDebugOverrides it is possible to override any part of the execution.

debugPlutusUnbounded ∷ HasCallStack ⇒ String → PlutusDebugOverrides → PlutusDebugInfo

This is just like debugPlutus, except it is pure and if a supplied script contains an infinite loop or a very expensive computation, it might not terminate within a reasonable timeframe.

runPlutusScript ∷ PlutusWithContext → ScriptResult

runPlutusScriptWithLogs ∷ PlutusWithContext → ([Text], ScriptResult)

evaluatePlutusWithContext ∷ VerboseMode → PlutusWithContext → ([Text], Either EvaluationError ExBudget)

explainPlutusEvaluationError ∷ PlutusWithContext → EvaluationError → ScriptResult

Explain why a script might fail. Scripts come in three flavors:

with 3 arguments [data,redeemer,context] for PlutusV1 and PlustuV2

with 2 arguments [redeemer,context] for PlutusV1 and PlustuV2

with 1 argument context for PlutusV3 onwards