Cardano.Ledger.Plutus.Evaluate

Documentation

data PlutusWithContext c where

This type contains all that is necessary from Ledger to evaluate a plutus script.

Constructors

PlutusWithContext

Fields

∷ PlutusLanguage l

⇒ { pwcProtocolVersion ∷ !Version

Mayjor protocol version that is necessary for [de]serialization

, pwcScript ∷ !(Either (Plutus l) (PlutusRunnable l))

Actual plutus script that will be evaluated. Script is allowed to be in two forms: serialized and deserialized. This is necesary for implementing the opptimization that preserves deserialized PlutusRunnable after verifying wellformedness of plutus scripts during transaction validation (yet to be implemented).

, pwcScriptHash ∷ !(ScriptHash c)

Hash of the above script as it would appear on-chain. In other words it is not just a hash of the script contents. (See hashScript for more info)

, pwcArgs ∷ !(PlutusArgs l)

All of the arguments to the Plutus scripts, including the redeemer and the Plutus context that was obtained from the transaction translation

, pwcExUnits ∷ !ExUnits

Limit on the execution units

, pwcCostModel ∷ !CostModel

CostModel to be used during script evaluation. It must match the language version in the pwcScript

} → PlutusWithContext c

data ScriptFailure c

Constructors

ScriptFailure

Fields

scriptFailureMessage ∷ Text

scriptFailurePlutus ∷ PlutusWithContext c

data ScriptResult c

Constructors

Passes [PlutusWithContext c]

Fails [PlutusWithContext c] (NonEmpty (ScriptFailure c))

scriptPass ∷ PlutusWithContext c → ScriptResult c

scriptFail ∷ ScriptFailure c → ScriptResult c

data PlutusDebugInfo c

Constructors

DebugBadHex String

DebugCannotDecode String

DebugSuccess

Fields

[Text]

Execution logs from the plutus interpreter

ExBudget

Execution budget that was consumed. It will always be less or equal to what was supplied during execution.

DebugFailure

Fields

[Text]

Execution logs from the plutus interpreter

EvaluationError

Evaluation error from Plutus interpreter

(PlutusWithContext c)

Everything that is needed in order to run the script

(Maybe ExBudget)

Expected execution budget. This value is Nothing when the supplied script can't be executed within 5 second limit or there is a problem with decoding plutus script itself.

debugPlutus ∷ Crypto c ⇒ String → IO (PlutusDebugInfo c)

runPlutusScript ∷ PlutusWithContext c → ScriptResult c

runPlutusScriptWithLogs ∷ PlutusWithContext c → ([Text], ScriptResult c)

evaluatePlutusWithContext ∷ VerboseMode → PlutusWithContext c → ([Text], Either EvaluationError ExBudget)

explainPlutusEvaluationError ∷ PlutusWithContext c → EvaluationError → ScriptResult c

Explain why a script might fail. Scripts come in three flavors:

with 3 arguments [data,redeemer,context] for PlutusV1 and PlustuV2

with 2 arguments [redeemer,context] for PlutusV1 and PlustuV2

with 1 argument context for PlutusV3 onwards