Cardano.Ledger.Plutus.Language

Description

This module provides data structures and operations for talking about Non-native Script languages. It is expected that new languages (or new versions of old languages) will be added here.

Plutus Script

newtype Plutus (l ∷ Language)

Serialized representation of a Plutus script that distinguishes the language version at the type level. When encoded in CBOR language version is also encoded.

Constructors

Plutus

Fields

plutusBinary ∷ PlutusBinary

isValidPlutus ∷ PlutusLanguage l ⇒ Version → Plutus l → Bool

Verify that the binary version of the Plutus script is deserializable.

newtype PlutusBinary

Binary representation of a Plutus script.

Constructors

PlutusBinary

Fields

unPlutusBinary ∷ ShortByteString

newtype PlutusRunnable (l ∷ Language)

This is a deserialized version of the Plutus type that can be used directly with evaluation functions that rely on evaluatePlutusRunnable.

The only way to obtain this type is by the means of deserializing Plutus with decodePlutusRunnable

Constructors

PlutusRunnable

Fields

plutusRunnable ∷ ScriptForEvaluation

plutusFromRunnable ∷ PlutusRunnable l → Plutus l

Serialize the runnable version of the plutus script

decodePlutusRunnable majVer (plutusFromRunnable pr) == Right pr

decodeWithPlutus ∷ (∀ si l. PlutusLanguage l ⇒ Plutus l → Decoder si a) → Decoder so a

Plutus script decoder that will enforce the type level language equals the one included in the serialization

hashPlutusScript ∷ ∀ c l. (Crypto c, PlutusLanguage l) ⇒ Plutus l → ScriptHash c

Compute a ScriptHash of a Plutus script. This function is equivalent to hashScript, except it is restricted to Plutus scripts

Value level Plutus Language version

data Language

Non-Native Plutus Script language. This is expected to be an open type. We will add new Constuctors to this type as additional Plutus language versions as are added. We use an enumerated type for two reasons.

We can write total functions by case analysis over the constructors

We use DataKinds to make some datatypes indexed by Language.

Note that the the serialization of Language depends on the ordering.

Constructors

PlutusV1

PlutusV2

PlutusV3

mkLanguageEnum ∷ Int → Maybe Language

Make a language from its Enum index.

languageToText ∷ Language → Text

languageFromText ∷ MonadFail m ⇒ Text → m Language

nonNativeLanguages ∷ [Language]

guardPlutus ∷ Language → Decoder s ()

Prevent decoding a version of Plutus until the appropriate protocol version.

Type level Plutus Language version

data SLanguage (l ∷ Language) where

Singleton for 'Language'

Constructors

SPlutusV1 ∷ SLanguage 'PlutusV1

SPlutusV2 ∷ SLanguage 'PlutusV2

SPlutusV3 ∷ SLanguage 'PlutusV3

class (Typeable l, NFData (PlutusArgs l), EncCBOR (PlutusArgs l), DecCBOR (PlutusArgs l), Pretty (PlutusArgs l), Show (PlutusArgs l), Eq (PlutusArgs l)) ⇒ PlutusLanguage (l ∷ Language) where

For implicit reflection on 'SLanguage' See Cardano.Ledger.Alonzo.Plutus.TxInfo for example usage

Associated Types

data PlutusArgs l ∷ Type

Methods

isLanguage ∷ SLanguage l

plutusLanguageTag ∷ Plutus l → Word8

Tag that will be used as a prefix to compute the ScriptHash

decodePlutusRunnable

Arguments

∷ Version
Which major protocol version to use for deserialization and further execution

→ Plutus l
Binary version of the script that will be deserialized

→ Either ScriptDecodeError (PlutusRunnable l)

evaluatePlutusRunnable

Arguments

∷ Version
Which major protocol version to use for script execution

→ VerboseMode
Whether to produce log output

→ EvaluationContext
Includes the cost model to use for tallying up the execution costs

→ ExBudget
The resource budget which must not be exceeded during evaluation

→ PlutusRunnable l
The script to evaluate

→ PlutusArgs l
The arguments to the script

→ (LogOutput, Either EvaluationError ExBudget)

evaluatePlutusRunnableBudget

Arguments

∷ Version
Which major protocol version to use for script execution

→ VerboseMode
Whether to produce log output

→ EvaluationContext
Includes the cost model to use for tallying up the execution costs

→ PlutusRunnable l
The script to evaluate

→ PlutusArgs l
The arguments to the script

→ (LogOutput, Either EvaluationError ExBudget)

Similar to evaluatePlutusRunnable, except does not require ExBudget to be provided as input and instead computes it as output. This function is meant to be used for testing.

mkTermToEvaluate

Arguments

∷ Version
Which major protocol version to use for script execution

→ PlutusRunnable l
The script to evaluate

→ PlutusArgs l
The arguments to the script

→ Either EvaluationError (Term NamedDeBruijn DefaultUni DefaultFun ())

plutusLanguage ∷ ∀ l proxy. PlutusLanguage l ⇒ proxy l → Language

Construct value level laguage version from the type level

plutusSLanguage ∷ PlutusLanguage l ⇒ proxy l → SLanguage l

fromSLanguage ∷ PlutusLanguage l ⇒ SLanguage l → Language

Deprecated: In favor of plutusLanguage

Reflection for 'SLanguage'

toSLanguage ∷ ∀ l m. (PlutusLanguage l, MonadFail m) ⇒ Language → m (SLanguage l)

withSLanguage ∷ Language → (∀ l. PlutusLanguage l ⇒ SLanguage l → a) → a

asSLanguage ∷ SLanguage l → proxy l → proxy l

withSamePlutusLanguage ∷ ∀ f1 f2 l1 l2 a. (PlutusLanguage l1, PlutusLanguage l2) ⇒ f1 l1 → f2 l2 → (∀ l. PlutusLanguage l ⇒ f1 l → f2 l → a) → Maybe a

Plutus Script Context

data LegacyPlutusArgs l

Constructors

LegacyPlutusArgs2
Scripts that require 2 arguments.

Fields

!Data

Redeemer

!(PlutusScriptContext l)

PlutusScriptContext

LegacyPlutusArgs3
Scripts that require 3 arguments. Which is only PlutusV1/V2 spending scripts

Fields

!Data

Mandatory Spending Datum

!Data

Redeemer

!(PlutusScriptContext l)

PlutusScriptContext

type family PlutusScriptContext (l ∷ Language) = r | r → l where ...

Equations

PlutusScriptContext 'PlutusV1 = ScriptContext

PlutusScriptContext 'PlutusV2 = ScriptContext

PlutusScriptContext 'PlutusV3 = ScriptContext

Orphan instances

NFData ScriptContext

Instance details

Methods

rnf ∷ ScriptContext → ()

NFData ScriptContext

Instance details

Methods

rnf ∷ ScriptContext → ()

NFData ScriptContext

Instance details

Methods

rnf ∷ ScriptContext → ()