Cardano.Ledger.Plutus.TxInfo

Documentation

data TxOutSource c

A transaction output can be translated because it is a newly created output, or because it is the output which is connected to a transaction input being spent.

Constructors

TxOutFromInput !(TxIn c)

TxOutFromOutput !TxIx

txOutSourceToText ∷ TxOutSource c → Text

transAddr ∷ Addr c → Maybe Address

Translate an address. NetworkId is discarded and Byron Addresses will result in Nothing.

transRewardAccount ∷ RewardAccount c → Credential

Translate reward account by discarding NetowrkId and only translating the staking credential.

Note - This function is the right one to use starting with PlutusV3, prior to that an extra StakingHash wrapper is needed.

transDataHash ∷ DataHash c → DatumHash

transKeyHash ∷ KeyHash d c → PubKeyHash

transSafeHash ∷ SafeHash c i → BuiltinByteString

transScriptHash ∷ ScriptHash c → ScriptHash

transTxId ∷ TxId c → TxId

transStakeReference ∷ StakeReference c → Maybe StakingCredential

transCred ∷ Credential kr c → Credential

slotToPOSIXTime ∷ EpochInfo (Either Text) → SystemStart → SlotNo → Either Text POSIXTime

transTxIn ∷ TxIn c → TxOutRef

transCoinToValue ∷ Coin → Value

transCoinToLovelace ∷ Coin → Lovelace

transDataPair ∷ (DataHash c, Data era) → (DatumHash, Datum)

transExUnits ∷ ExUnits → ExBudget

exBudgetToExUnits ∷ ExBudget → Maybe ExUnits

transBoundedRational ∷ BoundedRational r ⇒ r → Rational

transEpochNo ∷ EpochNo → Integer

transEpochInterval ∷ EpochInterval → Integer

transDatum ∷ Data era → Datum