Cardano.Ledger.PoolDistr

Description

Deprecated: Use State instead

Documentation

data IndividualPoolStake

The IndividualPoolStake contains all the stake controlled by a single stake pool (the combination of owners and delegates) for a given epoch, and also the hash of the stake pool's registered VRF key.

When a stake pool produces a block, the header contains the full VRF verification key and VRF value for leader election. We check the VRF key against the value in IndividualPoolStake and we check the VRF value using the epoch nonce and the relative stake of the pool as given in IndividualPoolStake. The stake is relative to the total amount of active stake in the network. Stake is active if it is both registered and delegated to a registered stake pool.

Constructors

IndividualPoolStake

Fields

individualPoolStake ∷ !Rational

Pool stake distribution. This is a ratio of individualTotalPoolStake/pdTotalActiveStake

individualTotalPoolStake ∷ !(CompactForm Coin)

Total stake delegated to this pool. In addition to all the stake that is part of individualPoolStake we also add proposal-deposits to this field.

individualPoolStakeVrf ∷ !(VRFVerKeyHash 'StakePoolVRF)

data PoolDistr

A map of stake pool IDs (the hash of the stake pool operator's verification key) to IndividualPoolStake. Also holds absolute values necessary for the calculations in the computeDRepDistr.

Constructors

PoolDistr

Fields

unPoolDistr ∷ !(Map (KeyHash 'StakePool) IndividualPoolStake)

pdTotalActiveStake ∷ !(CompactForm Coin)

Total stake delegated to registered stake pools. In addition to the stake considered for the individualPoolStake Rational, we add proposal-deposits to this field.

poolDistrDistrL ∷ Lens' PoolDistr (Map (KeyHash 'StakePool) IndividualPoolStake)

poolDistrTotalL ∷ Lens' PoolDistr (CompactForm Coin)

individualTotalPoolStakeL ∷ Lens' IndividualPoolStake (CompactForm Coin)