Cardano.Ledger.Rules.ValidationMode

Description

Describes modes under which we might validate certain rules in the ledger.

What does this mean? Sometimes, we will want to check only certain conditions specified in the rules. For example, when replaying a previously validated chain, we do not care about rerunning _any_ checks, only making the relevant changes to the ledger state.

Documentation

Static checks are used to indicate that a particular predicate depends only on the signal to the transition, rather than the state or environment. This is particularly relevant where the signal is something such as a transaction, which is fixed, whereas the state and environment depend upon the chain tip upon which we are trying to build a block.

lblStatic ∷ Label

Indicates that this check depends only upon the signal to the transition, not the state or environment.

(?!#) ∷ Bool → PredicateFailure sts → Rule sts ctx () infix 1

Construct a static predicate check.

The choice of # as a postfix here is made because often these are crypto checks.

(?!#:) ∷ Either e () → (e → PredicateFailure sts) → Rule sts ctx () infix 1

Construct a static predicate check with an explanation.

The choice of # as a postfix here is made because often these are crypto checks.

failBecauseS ∷ PredicateFailure sts → Rule sts ctx ()

Fail, if static checks are enabled.

Interface for independent Tests

class Inject t s where

Methods

inject ∷ t → s

type Test failure = Validation (NonEmpty failure) ()

runTest ∷ InjectRuleFailure rule f era ⇒ Test (f era) → Rule (EraRule rule era) ctx ()

runTestOnSignal ∷ InjectRuleFailure rule f era ⇒ Test (f era) → Rule (EraRule rule era) ctx ()