Cardano.Ledger.SafeHash

Description

Deprecated: Use Hashes instead

SafeHash and SafeToHash

data SafeHash i

A SafeHash is a hash of something that is safe to hash. Such types store their own serialisation bytes. The prime example is (MemoBytes t), but other examples are things that consist of only ByteStrings (i.e. they are their own serialization) or for some other reason store their original bytes.

We do NOT export the constructor SafeHash, but instead export other functions such as hashAnnotated and extractHash which have constraints that limit their application to types which preserve their original serialization bytes.

class SafeToHash t where

Only Types that preserve their serialisation bytes are members of the class SafeToHash. There are only a limited number of primitive direct instances of SafeToHash, all but two of them are present in this file. Instead of making explicit instances, we almost always use a newtype (around a type S) where their is already an instance (SafeToHash S). In that case the newtype has its SafeToHash instance derived using newtype deriving. The prime example of s is MemoBytes. The only exceptions are the legacy Shelley types: Metadata and ShelleyTx, that preserve their serialization bytes using a different mechanism than the use of MemoBytes. SafeToHash is a superclass requirement of the classes HashAnnotated which provide more convenient ways to construct SafeHashes than using makeHashWithExplicitProxys.

Minimal complete definition

originalBytes

Methods

originalBytes ∷ t → ByteString

Extract the original bytes from t

originalBytesSize ∷ t → Int

makeHashWithExplicitProxys ∷ Proxy i → t → SafeHash i

Creating SafeHash

class SafeToHash x ⇒ HashAnnotated x i | x → i

Types that are SafeToHash AND have the type uniquely determines the index type tag of SafeHash index

The SafeToHash and the HashAnnotated classes are designed so that their instances can be easily derived (because their methods have default methods when the type is a newtype around a type that is SafeToHash). For example,

hashAnnotated ∷ HashAnnotated x i ⇒ x → SafeHash i

Create a (SafeHash i), given (HashAnnotated x i) instance.

unsafeMakeSafeHash ∷ Hash HASH i → SafeHash i

Don't use this except in Testing to make Arbitrary instances, etc. or in cases when it can be guaranteed that original bytes were used for computing the hash.

Other operations

castSafeHash ∷ ∀ i j. SafeHash i → SafeHash j

To change the index parameter of SafeHash (which is a phantom type) use castSafeHash

extractHash ∷ SafeHash i → Hash HASH i

Extract the hash out of a SafeHash

indexProxy ∷ ∀ x i. HashAnnotated x i ⇒ x → Proxy i