Cardano.Ledger.Slot

Documentation

newtype SlotNo

The 0-based index for the Ourboros time slot.

Constructors

SlotNo

Fields

unSlotNo ∷ Word64

getTheSlotOfNoReturn ∷ HasCallStack ⇒ SlotNo → ShelleyBase (EpochNo, SlotNo, EpochNo)

Figure out a slot number that is two stability windows before the end of the next epoch. Together with the slot number we also return the current epoch number and the next epoch number.

The reason why it is called the point of no return, is because that is the point when HardForkCombinator (HFC) initiates a controlled hard fork, if there is a major protocol version update that forks into a new era.

newtype Duration

Constructors

Duration

Fields

unDuration ∷ Word64

(-*) ∷ SlotNo → SlotNo → Duration

(+*) ∷ SlotNo → Duration → SlotNo

(*-) ∷ SlotNo → Duration → SlotNo

Subtract a duration from a slot

newtype EpochNo

An epoch, i.e. the number of the epoch.

Constructors

EpochNo

Fields

unEpochNo ∷ Word64

newtype EpochSize

Constructors

EpochSize

Fields

unEpochSize ∷ Word64

data EpochInfo (m ∷ TYPE LiftedRep → Type)

Information about epochs

Different epochs may have different sizes and different slot lengths. This information is encapsulated by EpochInfo. It is parameterized over a monad m because the information about how long each epoch is may depend on information derived from the blockchain itself. It ultimately requires acess to state, and so either uses the monad for that or uses the monad to reify failure due to cached state information being too stale for the current query.

newtype BlockNo

The 0-based index of the block in the blockchain. BlockNo is <= SlotNo and is only equal at slot N if there is a block for every slot where N <= SlotNo.

Constructors

BlockNo

Fields

unBlockNo ∷ Word64

epochInfoEpoch ∷ HasCallStack ⇒ EpochInfo Identity → SlotNo → ShelleyBase EpochNo

epochInfoFirst ∷ HasCallStack ⇒ EpochInfo Identity → EpochNo → ShelleyBase SlotNo

epochInfoSize ∷ HasCallStack ⇒ EpochInfo Identity → EpochNo → ShelleyBase EpochSize