Cardano.Ledger.TxIn

Documentation

newtype TxId crypto

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_unTxId :: SafeHash crypto EraIndependentTxBody

data TxIn crypto

The input of a UTxO.

Constructors

TxIn !(TxId crypto) !TxIx

mkTxInPartial :: HasCallStack => TxId crypto -> Integer -> TxIn crypto

Construct TxIn while throwing an error for an out of range TxIx. Make sure to use it only for testing.

data TxIx

Transaction index.

txid :: forall era c. (HashAlgorithm (HASH c), HashAnnotated (TxBody era) EraIndependentTxBody c) => TxBody era -> TxId c

Compute the id of a transaction.