Cardano.Ledger.TxIn

Documentation

newtype TxId c

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

unTxId :: SafeHash c EraIndependentTxBody

_unTxId :: TxId c -> SafeHash c EraIndependentTxBody

Deprecated: In favor of unTxId

data TxIn c

The input of a UTxO.

Constructors

TxIn !(TxId c) !TxIx

mkTxInPartial :: HasCallStack => TxId c -> Integer -> TxIn c

Construct TxIn while throwing an error for an out of range TxIx. Make sure to use it only for testing.

txInToText :: TxIn c -> Text

data TxIx

Transaction index.