Cardano.Ledger.UMap

Description

A UMap (for Unified map) represents 3 Maps with the same domain in one direction, and a fourth one in the inverse direction. The advantage of using UMap is that UMap stores all the information compactly, by exploiting the the large amount of sharing in the 2 maps.

Constructing UMap

data Trip c where

a Trip compactly represents the range of 4 maps with the same domain as a single triple. The space compacting Trip datatype, and the pattern Triple are equivalent to:

data Trip c = Triple
  { coinT :: !(StrictMaybe RDPair),
    ptrT :: !(Set Ptr),
    poolidT :: !(StrictMaybe (KeyHash 'StakePool c))
  }
 deriving (Show, Eq, Generic, NoThunks, NFData)

To name the constructors of Trip we use the notation Txxx where each x is either F for full, i.e. the component is present, or E for empty, the component is not present. There are three components 1) the (CompactForm Coin) as a Word64, 2) the Ptr set, and 3) the pool id (KeyHash 'StakePool c) . So TEEE means none of the components are present, and TEEF means only the pool id is present. etc. The pattern Triple will correctly use the optimal constructor.

Bundled Patterns

pattern Triple ∷ StrictMaybe RDPair → Set Ptr → StrictMaybe (KeyHash 'StakePool c) → Trip c
A Triple can be extracted and injected into the TEEE ... TFFF constructors.

tripReward ∷ Trip c → Maybe RDPair

Extract the Reward-Deposit Pair if it is present. We can tell that the reward is present when Txxx has an F in the first position TFFF TFFE TFEF TFEE ^ ^ ^ ^ equivalent to the pattern (Triple (SJust c) _ _) -> Just c

tripRewardActiveDelegation ∷ Trip c → Maybe RDPair

Extract a delegated Reward-Deposit Pair if it is present. We can tell that the pair is present and active when Txxx has an F in the 1st position (present) and 3rd position (delegated). I.e. TFFF and TFEF ^ ^ ^ ^ This is equivalent to: pattern (Triple (SJust c) _ (SJust _)) -> Just c

tripDelegation ∷ Trip c → Maybe (KeyHash 'StakePool c)

Extract the Delegation PoolParams, if present. We can tell that the PoolParams are present when Txxx has an F in the third position TFFF TFEF TEFF TEEF ^ ^ ^ ^ equivalent to the pattern (Triple _ _ (SJust p)) -> Just p

data UMap c

A unified map represents 4 Maps with domain (Credential 'Staking c) for keys and one more in the inverse direction with Ptr for keys and (Credential 'Staking c) for values.

Constructors

UMap !(Map (Credential 'Staking c) (Trip c)) !(Map Ptr (Credential 'Staking c))

umInvariant ∷ Credential 'Staking c → Ptr → UMap c → Bool

It is worthwhile stating the invariant that holds on a Unified Map The ptrmap and the ptrT field of the tripmap are inverses.

View and its components

data View c k v where

A View lets one view a UMap in three different ways A view with type (View c key value) can be used like a (Map key value)

Constructors

RewardDeposits ∷ !(UMap c) → View c (Credential 'Staking c) RDPair

Delegations ∷ !(UMap c) → View c (Credential 'Staking c) (KeyHash 'StakePool c)

Ptrs ∷ !(UMap c) → View c Ptr (Credential 'Staking c)

unView ∷ View c k v → UMap c

Extract the underlying UMap from a View

unUnify ∷ View c k v → Map k v

Materialize a real Map from a View This is expensive, use it wisely (like maybe once per epoch boundary to make a SnapShot) See also domRestrictedView, which domain restricts before computing a view.

viewToVMap ∷ View c k v → VMap VB VB k v

Materialize a real Vector Map from a View This is expensive, use it wisely (like maybe once per epoch boundary to make a SnapShot)

rewView ∷ UMap c → Map (Credential 'Staking c) Coin

Materialize the RewardDeposits Map from a UMap

compactRewView ∷ UMap c → Map (Credential 'Staking c) (CompactForm Coin)

delView ∷ UMap c → Map (Credential 'Staking c) (KeyHash 'StakePool c)

Materialize the Delegation Map from a UMap

ptrView ∷ UMap c → Map Ptr (Credential 'Staking c)

Materialize the Ptr Map from a UMap

depositView ∷ UMap c → Map (Credential 'Staking c) Coin

Materialize the Deposit Map from a UMap

domRestrictedView ∷ Set k → View c k v → Map k v

Return the materialized View of a domain restricted Umap. if setk is small this should be efficient.

zero ∷ Trip c → Bool

Is there no information in a Triple? If so then we can delete it from the UnifedMap

zeroMaybe ∷ Trip c → Maybe (Trip c)

data family CompactForm a ∷ Type

toCompact ∷ Compactible a ⇒ a → Maybe (CompactForm a)

fromCompact ∷ Compactible a ⇒ CompactForm a → a

addCompact ∷ CompactForm Coin → CompactForm Coin → CompactForm Coin

sumCompactCoin ∷ Foldable t ⇒ t (CompactForm Coin) → CompactForm Coin

sumRewardsView ∷ View c k RDPair → CompactForm Coin

sumDepositView ∷ View c k RDPair → CompactForm Coin

compactCoinOrError ∷ HasCallStack ⇒ Coin → CompactForm Coin

Set and Map operations on Views

empty ∷ UMap c

delete ∷ k → View c k v → UMap c

delete' ∷ k → View c k v → View c k v

insertWith ∷ (v → v → v) → k → v → View c k v → UMap c

insertWith' ∷ (v → v → v) → k → v → View c k v → View c k v

Special insertion:

Keeps the value already in the ViewMap if the key k is already there:

insertWith' (\ old new -> old) k v xs

Replaces the value already in the ViewMap with v, if key k is already there:

insertWith' (\ old new -> new) k v xs

Replaces the value already in the ViewMap with the sum, if key k is already there:

insertWith' (\ old new -> old+new) k v xs

Ignores combine if the key k is NOT already in the ViewMap, and inserts v:

insertWith' combine k v xs

insert ∷ k → v → View c k v → UMap c

insert' ∷ k → v → View c k v → View c k v

adjust ∷ (RDPair → RDPair) → k → View c k RDPair → UMap c

lookup ∷ k → View c k v → Maybe v

isNull ∷ View c k v → Bool

domain ∷ View c k v → Set k

range ∷ View c k v → Set v

(∪) ∷ View c k v → (k, v) → UMap c

Union with left preference, so if k, already exists, do nothing, if it doesn't exist insert it.

(⨃) ∷ View c k v → Map k v → UMap c

Union with right preference, so if k, already exists, then old v is overwritten with the new v Special rules apply for the RewardDeposits view, where only the rdReward field of the RDPair is overwritten, and the old rdDeposit value persists. In this case it is an invariant that the domain mp is a subset of the domain of the RewardDeposits View. See the single case in module Cardano.Ledger.Shelley.Rules.Delegs, in the dealing with Withdrawals's where it is used at this type.

(∪+) ∷ View c k RDPair → Map k (CompactForm Coin) → UMap c

Add the aggegated Coin from aggRewMap to the RewardDeposits view of the UMap we assume the domain of aggRewMap is a subset of the domain of tripmap The other Views (other than RewardDeposits) are not reachable, since they do not have DPair as 3rd parameter

unionKeyDeposits ∷ View c k RDPair → Map k (CompactForm Coin) → UMap c

(⋪) ∷ Set k → View c k v → UMap c

(⋫) ∷ View c k v → Set v → UMap c

This is slow for Delegations and RewardDeposits Views, better hope the sets are small

(◁) ∷ View c k v → Map k u → Map k u

member ∷ k → View c k v → Bool

notMember ∷ k → View c k v → Bool

domRestrict ∷ View c k v → Map k u → Map k u

findWithDefault ∷ a → k → View c k a → a

Find the value associated with a key from a View, return the default if the key is not there.

size ∷ View c k a → Int

A View is a view, so the size of the view is NOT the same as the size of the underlying triple map.

unify ∷ Map (Credential 'Staking c) RDPair → Map (Credential 'Staking c) (KeyHash 'StakePool c) → Map Ptr (Credential 'Staking c) → UMap c

Create a UMap from 3 separate maps. For use in tests only.

data RDPair

Constructors

RDPair

Fields

rdReward ∷ !(CompactForm Coin)

rdDeposit ∷ !(CompactForm Coin)

rdPairView ∷ UMap c → Map (Credential 'Staking c) RDPair

Materialize the RDPairs Map from a UMap