Cardano.Ledger.UTxO

Primitives

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map (TxIn (EraCrypto era)) (TxOut era)

class EraTx era ⇒ EraUTxO era where

Associated Types

type ScriptsNeeded era = (r ∷ Type) | r → era

A customizable type on per era basis for the information required to find all scripts needed for the transaction.

Methods

getConsumedValue

Arguments

∷ PParams era

→ (Credential 'Staking (EraCrypto era) → Maybe Coin)
Function that can lookup current delegation deposits

→ (Credential 'DRepRole (EraCrypto era) → Maybe Coin)
Function that can lookup current drep deposits

→ UTxO era

→ TxBody era

→ Value era

Calculate all the value that is being consumed by the transaction.

getProducedValue

Arguments

∷ PParams era

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ TxBody era

→ Value era

getScriptsProvided

Arguments

∷ UTxO era
For some era it is necessary to look into the UTxO to find all of the available scripts for the transaction

→ Tx era

→ ScriptsProvided era

Initial eras will look into witness set to find all of the available scripts, but starting with Babbage we can look for available scripts in the UTxO using reference inputs.

getScriptsNeeded ∷ UTxO era → TxBody era → ScriptsNeeded era

Produce all the information required for figuring out which scripts are required for the transaction to be valid, once those scripts are evaluated

getScriptsHashesNeeded ∷ ScriptsNeeded era → Set (ScriptHash (EraCrypto era))

Extract the set of all script hashes that are needed for script validation.

getWitsVKeyNeeded ∷ CertState era → UTxO era → TxBody era → Set (KeyHash 'Witness (EraCrypto era))

Extract all of the KeyHash witnesses that are required for validating the transaction

getMinFeeTxUtxo ∷ PParams era → Tx era → UTxO era → Coin

Minimum fee computation, excluding witnesses and including ref scripts size

newtype ScriptsProvided era

The only reason it is a newtype instead of just a Map is becuase for later eras is expensive to compute the actual map, so we want to use the type safety guidance to avoid redundant work.

Constructors

ScriptsProvided

Fields

unScriptsProvided ∷ Map (ScriptHash (EraCrypto era)) (Script era)

Functions

txins ∷ EraTxBody era ⇒ TxBody era → Set (TxIn (EraCrypto era))

Compute the UTxO inputs of a transaction. txins has the same problems as txouts, see notes below.

txinLookup ∷ TxIn (EraCrypto era) → UTxO era → Maybe (TxOut era)

Lookup a txin for a given UTxO collection

txInsFilter

Arguments

∷ UTxO era
Source UTxO

→ Set (TxIn (EraCrypto era))
Which of the TxIns you would like to keep.

→ UTxO era

Filter out TxIn's from the UTxO map

txouts ∷ ∀ era. EraTxBody era ⇒ TxBody era → UTxO era

Compute the transaction outputs of a transaction.

balance ∷ EraTxOut era ⇒ UTxO era → Value era

Determine the total balance contained in the UTxO.

coinBalance ∷ EraTxOut era ⇒ UTxO era → Coin

Determine the total Ada only balance contained in the UTxO. This is equivalent to `coin . balance`, but it will be more efficient

sumAllValue ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Value era

Sum all the value in any Foldable with TxOuts

sumAllCoin ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Coin

Sum all the Coins in any Foldable with with TxOuts. Care should be taken since it is susceptible to integer overflow, therefore make sure this function is not applied to unvalidated TxOuts

areAllAdaOnly ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Bool

Check whether any of the supplied TxOuts contain any MultiAssets. Returns True if non of them do.

verifyWitVKey ∷ (Typeable kr, Crypto c, DSignable c (Hash c EraIndependentTxBody)) ⇒ Hash c EraIndependentTxBody → WitVKey kr c → Bool

Verify a transaction body witness

getScriptHash ∷ Addr c → Maybe (ScriptHash c)

Extract script hash from value address with script.