Cardano.Ledger.UTxO

Primitives

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map (TxIn (EraCrypto era)) (TxOut era)

class EraTxBody era ⇒ EraUTxO era where

Associated Types

type ScriptsNeeded era = (r ∷ Type) | r → era

A customizable type on per era basis for the information required to find all scripts needed for the transaction.

Methods

getConsumedValue

Arguments

∷ PParams era

→ (StakeCredential (EraCrypto era) → Maybe Coin)
Function that can lookup current delegation deposits

→ UTxO era

→ TxBody era

→ Value era

Calculate all the value that is being consumed by the transaction.

getScriptsNeeded ∷ UTxO era → TxBody era → ScriptsNeeded era

Produce all the information required for figuring out which scripts are required for the transaction to be valid, once those scripts are evaluated

getScriptsHashesNeeded ∷ ScriptsNeeded era → Set (ScriptHash (EraCrypto era))

Extract the set of all script hashes that are needed for script validation.

Functions

txins ∷ EraTxBody era ⇒ TxBody era → Set (TxIn (EraCrypto era))

Compute the UTxO inputs of a transaction. txins has the same problems as txouts, see notes below.

txinLookup ∷ TxIn (EraCrypto era) → UTxO era → Maybe (TxOut era)

Lookup a txin for a given UTxO collection

txouts ∷ ∀ era. EraTxBody era ⇒ TxBody era → UTxO era

Compute the transaction outputs of a transaction.

balance ∷ EraTxOut era ⇒ UTxO era → Value era

Determine the total balance contained in the UTxO.

coinBalance ∷ EraTxOut era ⇒ UTxO era → Coin

Determine the total Ada only balance contained in the UTxO. This is equivalent to `coin . balance`, but it will be more efficient

sumAllValue ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Value era

Sum all the value in any Foldable with TxOuts

sumAllCoin ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Coin

Sum all the Coins in any Foldable with with TxOuts. Care should be taken since it is susceptible to integer overflow, therefore make sure this function is not applied to unvalidated TxOuts

areAllAdaOnly ∷ (EraTxOut era, Foldable f) ⇒ f (TxOut era) → Bool

Check whether any of the supplied TxOuts contain any MultiAssets. Returns True if non of them do.

verifyWitVKey ∷ (Typeable kr, Crypto c, DSignable c (Hash c EraIndependentTxBody)) ⇒ Hash c EraIndependentTxBody → WitVKey kr c → Bool

Verify a transaction body witness

getScriptHash ∷ Addr c → Maybe (ScriptHash c)

Extract script hash from value address with script.