Cardano.Ledger.Val

Description

This module defines a generalised notion of a "value" - that is, something with which we may quantify a transaction output.

Documentation

class (Abelian t, Eq t) => Val t where

Minimal complete definition

(<×>), coin, inject, modifyCoin, size, pointwise, isAdaOnly, isAdaOnlyCompact, injectCompact

Methods

zero :: t

the value with nothing in it

(<+>) :: t -> t -> t infixl 6

add two value

(<×>) :: Integral i => i -> t -> t infixl 7

scale a value by an Integral constant

(<->) :: t -> t -> t infixl 6

subtract two values

isZero :: t -> Bool

Is the argument zero?

coin :: t -> Coin

Get the ADA present in the value (since ADA is our "blessed" currency)

inject :: Coin -> t

Create a value containing only this amount of ADA

modifyCoin :: (Coin -> Coin) -> t -> t

modify the blessed Coin part of t

size :: t -> Integer

pointwise :: (Integer -> Integer -> Bool) -> t -> t -> Bool

used to compare values pointwise. Rather than using: (v1 <= v2) use: pointwise (<=) v1 v2 | If a quantity is stored in only one of v1 or v2, we use 0 for the missing quantity.

isAdaOnly :: t -> Bool

Check if value contains only ADA. Must hold property:

inject (coin v) == v

isAdaOnlyCompact :: CompactForm t -> Bool

injectCompact :: CompactForm Coin -> CompactForm t

scale :: (Val t, Integral i) => i -> t -> t

invert :: Val t => t -> t

sumVal :: (Foldable t, Val v) => t v -> v

adaOnly :: Val v => v -> Bool

Deprecated: In favor of isAdaOnly

class DecodeNonNegative v where

Methods

decodeNonNegative :: Decoder s v

class DecodeMint v where

Methods

decodeMint :: Decoder s v

class EncodeMint v where

Methods

encodeMint :: v -> Encoding