Test.Cardano.Ledger.Core.Address

Description

This module contains previous implementations for Addr and CoompactAddr deserialization. This is used as an alternative implementation for testing and as a perforance reference for benchmarking.

Documentation

deserialiseAddrOld ∷ ∀ c m. (Crypto c, MonadFail m) ⇒ ByteString → m (Addr c)

Deserialise an address from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

deserialiseRewardAccountOld ∷ ∀ c m. (Crypto c, MonadFail m) ⇒ ByteString → m (RewardAccount c)

Deserialise an reward account from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

decompactAddrOld ∷ Crypto c ⇒ CompactAddr c → Addr c

This is an old decompacter that didn't guard against random junk at the end.

decompactAddrOldLazy ∷ ∀ c. Crypto c ⇒ CompactAddr c → Addr c

This lazy deserializer is kept around purely for benchmarking, so we can verify that new deserializer decodeAddrStateT is doing the work lazily.

Deprecations

deserialiseRewardAcntOld ∷ ∀ c m. (Crypto c, MonadFail m) ⇒ ByteString → m (RewardAccount c)

Deprecated: Use deserialiseRewardAccountOld instead