Test.Cardano.Ledger.Core.Arbitrary

Documentation

module Test.Cardano.Ledger.Binary.Arbitrary

genAddrBadPtr ∷ Gen Addr

Deprecated: Addresses with bad pointers are no longer possible

genCompactAddrBadPtr ∷ Gen CompactAddr

Deprecated: Addresses with bad pointers are no longer possible

genBadPtr ∷ Gen Ptr

Deprecated: Bad pointers are no longer possible

Generate a Ptr with full 64bit range for values. Not allowed starting in Babbage

genericShrinkMemo ∷ ∀ era a. (Era era, EncCBOR (RawType (a era)), Memoized (a era), Generic (RawType (a era)), RecursivelyShrink (Rep (RawType (a era))), GSubterms (Rep (RawType (a era))) (RawType (a era))) ⇒ a era → [a era]

Plutus

genValidAndUnknownCostModels ∷ Gen CostModels

genValidCostModel ∷ Language → Gen CostModel

genValidCostModels ∷ Set Language → Gen CostModels

Utils

Will need to find a better home in the future

uniformSubSet

Arguments

∷ (StatefulGen g m, Ord k)

⇒ Maybe Int
Size of the subset. If supplied will be clamped to [0, Set.size s] interval, otherwise will be generated randomly.

→ Set k

→ g

→ m (Set k)

uniformSubMap

Arguments

∷ (StatefulGen g m, Ord k)

⇒ Maybe Int
Size of the subMap. If supplied will be clamped to [0, Map.size s] interval, otherwise will be generated randomly.

→ Map k v

→ g

→ m (Map k v)

uniformSubMapElems

Arguments

∷ (StatefulGen g m, Monoid f)

⇒ (k → v → f → f)

→ Maybe Int
Size of the subMap. If supplied will be clamped to [0, Map.size s] interval, otherwise will be generated randomly.

→ Map k v

→ g

→ m f

Orphan instances

Arbitrary AddrAttributes

Instance details

Methods

arbitrary ∷ Gen AddrAttributes

shrink ∷ AddrAttributes → [AddrAttributes]

Arbitrary Address

Instance details

Methods

arbitrary ∷ Gen Address

shrink ∷ Address → [Address]

Arbitrary Addr

Instance details

Methods

arbitrary ∷ Gen Addr

shrink ∷ Addr → [Addr]

Arbitrary BootstrapAddress

Instance details

Methods

arbitrary ∷ Gen BootstrapAddress

shrink ∷ BootstrapAddress → [BootstrapAddress]

Arbitrary CompactAddr

Instance details

Methods

arbitrary ∷ Gen CompactAddr

shrink ∷ CompactAddr → [CompactAddr]

Arbitrary RewardAccount

Instance details

Methods

arbitrary ∷ Gen RewardAccount

shrink ∷ RewardAccount → [RewardAccount]

Arbitrary Withdrawals

Instance details

Methods

arbitrary ∷ Gen Withdrawals

shrink ∷ Withdrawals → [Withdrawals]

Arbitrary ActiveSlotCoeff

Instance details

Methods

arbitrary ∷ Gen ActiveSlotCoeff

shrink ∷ ActiveSlotCoeff → [ActiveSlotCoeff]

Arbitrary Anchor

Instance details

Methods

arbitrary ∷ Gen Anchor

shrink ∷ Anchor → [Anchor]

Arbitrary BlocksMade

Instance details

Methods

arbitrary ∷ Gen BlocksMade

shrink ∷ BlocksMade → [BlocksMade]

Arbitrary CertIx

Instance details

Methods

arbitrary ∷ Gen CertIx

shrink ∷ CertIx → [CertIx]

Arbitrary DnsName

Instance details

Methods

arbitrary ∷ Gen DnsName

shrink ∷ DnsName → [DnsName]

Arbitrary Network

Instance details

Methods

arbitrary ∷ Gen Network

shrink ∷ Network → [Network]

Arbitrary NonNegativeInterval
Decimal numbers only

Instance details

Methods

arbitrary ∷ Gen NonNegativeInterval

shrink ∷ NonNegativeInterval → [NonNegativeInterval]

Arbitrary Nonce

Instance details

Methods

arbitrary ∷ Gen Nonce

shrink ∷ Nonce → [Nonce]

Arbitrary Port

Instance details

Methods

arbitrary ∷ Gen Port

shrink ∷ Port → [Port]

Arbitrary PositiveInterval
Decimal numbers only

Instance details

Methods

arbitrary ∷ Gen PositiveInterval

shrink ∷ PositiveInterval → [PositiveInterval]

Arbitrary PositiveUnitInterval
Decimal numbers only

Instance details

Methods

arbitrary ∷ Gen PositiveUnitInterval

shrink ∷ PositiveUnitInterval → [PositiveUnitInterval]

Arbitrary ProtVer

Instance details

Methods

arbitrary ∷ Gen ProtVer

shrink ∷ ProtVer → [ProtVer]

Arbitrary TxIx

Instance details

Methods

arbitrary ∷ Gen TxIx

shrink ∷ TxIx → [TxIx]

Arbitrary UnitInterval
Decimal numbers only

Instance details

Methods

arbitrary ∷ Gen UnitInterval

shrink ∷ UnitInterval → [UnitInterval]

Arbitrary Url

Instance details

Methods

arbitrary ∷ Gen Url

shrink ∷ Url → [Url]

Arbitrary Coin

Instance details

Methods

arbitrary ∷ Gen Coin

shrink ∷ Coin → [Coin]

Arbitrary DeltaCoin

Instance details

Methods

arbitrary ∷ Gen DeltaCoin

shrink ∷ DeltaCoin → [DeltaCoin]

Arbitrary PoolCert

Instance details

Methods

arbitrary ∷ Gen PoolCert

shrink ∷ PoolCert → [PoolCert]

Arbitrary Ptr

Instance details

Methods

arbitrary ∷ Gen Ptr

shrink ∷ Ptr → [Ptr]

Arbitrary SlotNo32

Instance details

Methods

arbitrary ∷ Gen SlotNo32

shrink ∷ SlotNo32 → [SlotNo32]

Arbitrary StakeReference

Instance details

Methods

arbitrary ∷ Gen StakeReference

shrink ∷ StakeReference → [StakeReference]

Arbitrary DRep

Instance details

Methods

arbitrary ∷ Gen DRep

shrink ∷ DRep → [DRep]

Arbitrary DRepState

Instance details

Methods

arbitrary ∷ Gen DRepState

shrink ∷ DRepState → [DRepState]

Arbitrary GenDelegPair

Instance details

Methods

arbitrary ∷ Gen GenDelegPair

shrink ∷ GenDelegPair → [GenDelegPair]

Arbitrary GenDelegs

Instance details

Methods

arbitrary ∷ Gen GenDelegs

shrink ∷ GenDelegs → [GenDelegs]

Arbitrary ScriptHash

Instance details

Methods

arbitrary ∷ Gen ScriptHash

shrink ∷ ScriptHash → [ScriptHash]

Arbitrary TxAuxDataHash

Instance details

Methods

arbitrary ∷ Gen TxAuxDataHash

shrink ∷ TxAuxDataHash → [TxAuxDataHash]

Arbitrary BootstrapWitness

Instance details

Methods

arbitrary ∷ Gen BootstrapWitness

shrink ∷ BootstrapWitness → [BootstrapWitness]

Arbitrary ChainCode

Instance details

Methods

arbitrary ∷ Gen ChainCode

shrink ∷ ChainCode → [ChainCode]

Arbitrary CostModel

Instance details

Methods

arbitrary ∷ Gen CostModel

shrink ∷ CostModel → [CostModel]

Arbitrary CostModels
This Arbitrary instance assumes the inflexible deserialization scheme prior to version 9.

Instance details

Methods

arbitrary ∷ Gen CostModels

shrink ∷ CostModels → [CostModels]

Arbitrary ExUnits

Instance details

Methods

arbitrary ∷ Gen ExUnits

shrink ∷ ExUnits → [ExUnits]

Arbitrary Prices

Instance details

Methods

arbitrary ∷ Gen Prices

shrink ∷ Prices → [Prices]

Arbitrary Language

Instance details

Methods

arbitrary ∷ Gen Language

shrink ∷ Language → [Language]

Arbitrary Reward

Instance details

Methods

arbitrary ∷ Gen Reward

shrink ∷ Reward → [Reward]

Arbitrary RewardType

Instance details

Methods

arbitrary ∷ Gen RewardType

shrink ∷ RewardType → [RewardType]

Arbitrary CommitteeAuthorization

Instance details

Methods

arbitrary ∷ Gen CommitteeAuthorization

shrink ∷ CommitteeAuthorization → [CommitteeAuthorization]

Arbitrary FutureGenDeleg

Instance details

Methods

arbitrary ∷ Gen FutureGenDeleg

shrink ∷ FutureGenDeleg → [FutureGenDeleg]

Arbitrary InstantaneousRewards

Instance details

Methods

arbitrary ∷ Gen InstantaneousRewards

shrink ∷ InstantaneousRewards → [InstantaneousRewards]

Arbitrary ChainAccountState

Instance details

Methods

arbitrary ∷ Gen ChainAccountState

shrink ∷ ChainAccountState → [ChainAccountState]

Arbitrary IndividualPoolStake

Instance details

Methods

arbitrary ∷ Gen IndividualPoolStake

shrink ∷ IndividualPoolStake → [IndividualPoolStake]

Arbitrary PoolDistr

Instance details

Methods

arbitrary ∷ Gen PoolDistr

shrink ∷ PoolDistr → [PoolDistr]

Arbitrary SnapShot

Instance details

Methods

arbitrary ∷ Gen SnapShot

shrink ∷ SnapShot → [SnapShot]

Arbitrary SnapShots

Instance details

Methods

arbitrary ∷ Gen SnapShots

shrink ∷ SnapShots → [SnapShots]

Arbitrary Stake
In the system, Stake never contains more than the sum of all Ada (which is constant). This makes it safe to store individual Coins (in CompactForm) as Word64. But we must be careful that we never generate Stake where the sum of all the coins exceeds (maxBound :: Word64) There will never be a real Stake in the system with that many Ada, because total Ada is constant. So using a restricted Arbitrary Generator is OK.

Instance details

Methods

arbitrary ∷ Gen Stake

shrink ∷ Stake → [Stake]

Arbitrary PoolMetadata

Instance details

Methods

arbitrary ∷ Gen PoolMetadata

shrink ∷ PoolMetadata → [PoolMetadata]

Arbitrary PoolParams

Instance details

Methods

arbitrary ∷ Gen PoolParams

shrink ∷ PoolParams → [PoolParams]

Arbitrary SizeOfPoolOwners

Instance details

Methods

arbitrary ∷ Gen SizeOfPoolOwners

shrink ∷ SizeOfPoolOwners → [SizeOfPoolOwners]

Arbitrary SizeOfPoolRelays

Instance details

Methods

arbitrary ∷ Gen SizeOfPoolRelays

shrink ∷ SizeOfPoolRelays → [SizeOfPoolRelays]

Arbitrary StakePoolRelay

Instance details

Methods

arbitrary ∷ Gen StakePoolRelay

shrink ∷ StakePoolRelay → [StakePoolRelay]

Arbitrary StakePoolState

Instance details

Methods

arbitrary ∷ Gen StakePoolState

shrink ∷ StakePoolState → [StakePoolState]

Arbitrary TxId

Instance details

Methods

arbitrary ∷ Gen TxId

shrink ∷ TxId → [TxId]

Arbitrary TxIn

Instance details

Methods

arbitrary ∷ Gen TxIn

shrink ∷ TxIn → [TxIn]

Arbitrary EpochInterval

Instance details

Methods

arbitrary ∷ Gen EpochInterval

shrink ∷ EpochInterval → [EpochInterval]

Arbitrary Data

Instance details

Methods

arbitrary ∷ Gen Data

shrink ∷ Data → [Data]

Arbitrary ExBudget

Instance details

Methods

arbitrary ∷ Gen ExBudget

shrink ∷ ExBudget → [ExBudget]

Arbitrary ExCPU

Instance details

Methods

arbitrary ∷ Gen ExCPU

shrink ∷ ExCPU → [ExCPU]

Arbitrary ExMemory

Instance details

Methods

arbitrary ∷ Gen ExMemory

shrink ∷ ExMemory → [ExMemory]

GenValid ActiveSlotCoeff

Instance details

Methods

genValid ∷ Gen ActiveSlotCoeff

shrinkValid ∷ ActiveSlotCoeff → [ActiveSlotCoeff]

GenValid NonNegativeInterval

Instance details

Methods

genValid ∷ Gen NonNegativeInterval

shrinkValid ∷ NonNegativeInterval → [NonNegativeInterval]

GenValid PositiveInterval

Instance details

Methods

genValid ∷ Gen PositiveInterval

shrinkValid ∷ PositiveInterval → [PositiveInterval]

GenValid PositiveUnitInterval

Instance details

Methods

genValid ∷ Gen PositiveUnitInterval

shrinkValid ∷ PositiveUnitInterval → [PositiveUnitInterval]

GenValid UnitInterval

Instance details

Methods

genValid ∷ Gen UnitInterval

shrinkValid ∷ UnitInterval → [UnitInterval]

Validity ActiveSlotCoeff

Instance details

Methods

validate ∷ ActiveSlotCoeff → Validation

Validity NonNegativeInterval

Instance details

Methods

validate ∷ NonNegativeInterval → Validation

Validity PositiveInterval

Instance details

Methods

validate ∷ PositiveInterval → Validation

Validity PositiveUnitInterval

Instance details

Methods

validate ∷ PositiveUnitInterval → Validation

Validity UnitInterval

Instance details

Methods

validate ∷ UnitInterval → Validation

(Era era, EncCBOR (f era), Arbitrary (f era)) ⇒ Arbitrary (Sized (f era))

Instance details

Methods

arbitrary ∷ Gen (Sized (f era))

shrink ∷ Sized (f era) → [Sized (f era)]

Arbitrary (Attributes AddrAttributes)

Instance details

Methods

arbitrary ∷ Gen (Attributes AddrAttributes)

shrink ∷ Attributes AddrAttributes → [Attributes AddrAttributes]

(Arbitrary a, HasZero a) ⇒ Arbitrary (NonZero a)

Instance details

Methods

arbitrary ∷ Gen (NonZero a)

shrink ∷ NonZero a → [NonZero a]

Arbitrary (CompactForm Coin)

Instance details

Methods

arbitrary ∷ Gen (CompactForm Coin)

shrink ∷ CompactForm Coin → [CompactForm Coin]

(Era era, Arbitrary (PParamsHKD Identity era)) ⇒ Arbitrary (PParams era)

Instance details

Methods

arbitrary ∷ Gen (PParams era)

shrink ∷ PParams era → [PParams era]

(Era era, Arbitrary (PParamsHKD StrictMaybe era)) ⇒ Arbitrary (PParamsUpdate era)

Instance details

Methods

arbitrary ∷ Gen (PParamsUpdate era)

shrink ∷ PParamsUpdate era → [PParamsUpdate era]

Arbitrary (Credential r)

Instance details

Methods

arbitrary ∷ Gen (Credential r)

shrink ∷ Credential r → [Credential r]

Arbitrary (NoGenesis era)

Instance details

Methods

arbitrary ∷ Gen (NoGenesis era)

shrink ∷ NoGenesis era → [NoGenesis era]

Arbitrary (NoUpdate a)

Instance details

Methods

arbitrary ∷ Gen (NoUpdate a)

shrink ∷ NoUpdate a → [NoUpdate a]

Arbitrary (KeyHash r)

Instance details

Methods

arbitrary ∷ Gen (KeyHash r)

shrink ∷ KeyHash r → [KeyHash r]

Arbitrary (SafeHash i)

Instance details

Methods

arbitrary ∷ Gen (SafeHash i)

shrink ∷ SafeHash i → [SafeHash i]

Arbitrary (VRFVerKeyHash r)

Instance details

Methods

arbitrary ∷ Gen (VRFVerKeyHash r)

shrink ∷ VRFVerKeyHash r → [VRFVerKeyHash r]

Arbitrary (VKey kd)

Instance details

Methods

arbitrary ∷ Gen (VKey kd)

shrink ∷ VKey kd → [VKey kd]

Typeable kr ⇒ Arbitrary (WitVKey kr)

Instance details

Methods

arbitrary ∷ Gen (WitVKey kr)

shrink ∷ WitVKey kr → [WitVKey kr]

Era era ⇒ Arbitrary (BinaryData era)

Instance details

Methods

arbitrary ∷ Gen (BinaryData era)

shrink ∷ BinaryData era → [BinaryData era]

Era era ⇒ Arbitrary (Data era)

Instance details

Methods

arbitrary ∷ Gen (Data era)

shrink ∷ Data era → [Data era]

Era era ⇒ Arbitrary (Datum era)

Instance details

Methods

arbitrary ∷ Gen (Datum era)

shrink ∷ Datum era → [Datum era]

Arbitrary (CommitteeState era)

Instance details

Methods

arbitrary ∷ Gen (CommitteeState era)

shrink ∷ CommitteeState era → [CommitteeState era]

(Era era, Arbitrary (Accounts era)) ⇒ Arbitrary (DState era)

Instance details

Methods

arbitrary ∷ Gen (DState era)

shrink ∷ DState era → [DState era]

Arbitrary (PState era)

Instance details

Methods

arbitrary ∷ Gen (PState era)

shrink ∷ PState era → [PState era]

(EraTxOut era, Arbitrary (TxOut era)) ⇒ Arbitrary (UTxO era)

Instance details

Methods

arbitrary ∷ Gen (UTxO era)

shrink ∷ UTxO era → [UTxO era]

Arbitrary a ⇒ Arbitrary (Mismatch r a)

Instance details

Methods

arbitrary ∷ Gen (Mismatch r a)

shrink ∷ Mismatch r a → [Mismatch r a]