Test.Cardano.Ledger.Core.Arbitrary

Documentation

module Test.Cardano.Ledger.Binary.Arbitrary

genAddrBadPtr ∷ Crypto c ⇒ Gen (Addr c)

genCompactAddrBadPtr ∷ Crypto c ⇒ Gen (CompactAddr c)

genBadPtr ∷ Gen Ptr

Generate a Ptr with full 64bit range for values. Not allowed starting in Babbage

genValidUMap ∷ Gen (UMap StandardCrypto)

genValidUMapNonEmpty ∷ Gen (UMap StandardCrypto)

genValidTriples ∷ Gen (Map (Credential 'Staking StandardCrypto) RDPair, Map (Credential 'Staking StandardCrypto) (KeyHash 'StakePool StandardCrypto), Map Ptr (Credential 'Staking StandardCrypto))

Used for testing UMap operations

genValidTriplesNonEmpty ∷ Gen (Map (Credential 'Staking StandardCrypto) RDPair, Map (Credential 'Staking StandardCrypto) (KeyHash 'StakePool StandardCrypto), Map Ptr (Credential 'Staking StandardCrypto))

genInvariantNonEmpty ∷ Gen (Credential 'Staking StandardCrypto, Ptr, UMap StandardCrypto)

genRightPreferenceUMap ∷ Gen (UMap StandardCrypto, Map (Credential 'Staking StandardCrypto) RDPair)

genInsertDeleteRoundtripRDPair ∷ Gen (UMap StandardCrypto, Credential 'Staking StandardCrypto, RDPair)

genInsertDeleteRoundtripDelegation ∷ Gen (UMap StandardCrypto, Credential 'Staking StandardCrypto, KeyHash 'StakePool StandardCrypto)

genInsertDeleteRoundtripPtr ∷ Gen (UMap StandardCrypto, Ptr, Credential 'Staking StandardCrypto)

Orphan instances

Arbitrary ActiveSlotCoeff

Instance details

Methods

arbitrary ∷ Gen ActiveSlotCoeff

shrink ∷ ActiveSlotCoeff → [ActiveSlotCoeff]

Arbitrary CertIx

Instance details

Methods

arbitrary ∷ Gen CertIx

shrink ∷ CertIx → [CertIx]

Arbitrary DnsName

Instance details

Methods

arbitrary ∷ Gen DnsName

shrink ∷ DnsName → [DnsName]

Arbitrary Network

Instance details

Methods

arbitrary ∷ Gen Network

shrink ∷ Network → [Network]

Arbitrary NonNegativeInterval
Decimal numbers only

Instance details

Methods

arbitrary ∷ Gen NonNegativeInterval

shrink ∷ NonNegativeInterval → [NonNegativeInterval]

Arbitrary Nonce

Instance details

Methods

arbitrary ∷ Gen Nonce

shrink ∷ Nonce → [Nonce]

Arbitrary Port

Instance details

Methods

arbitrary ∷ Gen Port

shrink ∷ Port → [Port]

Arbitrary PositiveInterval
Decimal numbers only

Instance details

Methods

arbitrary ∷ Gen PositiveInterval

shrink ∷ PositiveInterval → [PositiveInterval]

Arbitrary PositiveUnitInterval
Decimal numbers only

Instance details

Methods

arbitrary ∷ Gen PositiveUnitInterval

shrink ∷ PositiveUnitInterval → [PositiveUnitInterval]

Arbitrary ProtVer

Instance details

Methods

arbitrary ∷ Gen ProtVer

shrink ∷ ProtVer → [ProtVer]

Arbitrary TxIx

Instance details

Methods

arbitrary ∷ Gen TxIx

shrink ∷ TxIx → [TxIx]

Arbitrary UnitInterval
Decimal numbers only

Instance details

Methods

arbitrary ∷ Gen UnitInterval

shrink ∷ UnitInterval → [UnitInterval]

Arbitrary Url

Instance details

Methods

arbitrary ∷ Gen Url

shrink ∷ Url → [Url]

Arbitrary Coin

Instance details

Methods

arbitrary ∷ Gen Coin

shrink ∷ Coin → [Coin]

Arbitrary DeltaCoin

Instance details

Methods

arbitrary ∷ Gen DeltaCoin

shrink ∷ DeltaCoin → [DeltaCoin]

Arbitrary Ptr

Instance details

Methods

arbitrary ∷ Gen Ptr

shrink ∷ Ptr → [Ptr]

Arbitrary ChainCode

Instance details

Methods

arbitrary ∷ Gen ChainCode

shrink ∷ ChainCode → [ChainCode]

Arbitrary PoolMetadata

Instance details

Methods

arbitrary ∷ Gen PoolMetadata

shrink ∷ PoolMetadata → [PoolMetadata]

Arbitrary SizeOfPoolOwners

Instance details

Methods

arbitrary ∷ Gen SizeOfPoolOwners

shrink ∷ SizeOfPoolOwners → [SizeOfPoolOwners]

Arbitrary SizeOfPoolRelays

Instance details

Methods

arbitrary ∷ Gen SizeOfPoolRelays

shrink ∷ SizeOfPoolRelays → [SizeOfPoolRelays]

Arbitrary StakePoolRelay

Instance details

Methods

arbitrary ∷ Gen StakePoolRelay

shrink ∷ StakePoolRelay → [StakePoolRelay]

Arbitrary RewardType

Instance details

Methods

arbitrary ∷ Gen RewardType

shrink ∷ RewardType → [RewardType]

Arbitrary RDPair

Instance details

Methods

arbitrary ∷ Gen RDPair

shrink ∷ RDPair → [RDPair]

GenValid ActiveSlotCoeff

Instance details

Methods

genValid ∷ Gen ActiveSlotCoeff

shrinkValid ∷ ActiveSlotCoeff → [ActiveSlotCoeff]

GenValid NonNegativeInterval

Instance details

Methods

genValid ∷ Gen NonNegativeInterval

shrinkValid ∷ NonNegativeInterval → [NonNegativeInterval]

GenValid PositiveInterval

Instance details

Methods

genValid ∷ Gen PositiveInterval

shrinkValid ∷ PositiveInterval → [PositiveInterval]

GenValid PositiveUnitInterval

Instance details

Methods

genValid ∷ Gen PositiveUnitInterval

shrinkValid ∷ PositiveUnitInterval → [PositiveUnitInterval]

GenValid UnitInterval

Instance details

Methods

genValid ∷ Gen UnitInterval

shrinkValid ∷ UnitInterval → [UnitInterval]

Validity ActiveSlotCoeff

Instance details

Methods

validate ∷ ActiveSlotCoeff → Validation

Validity NonNegativeInterval

Instance details

Methods

validate ∷ NonNegativeInterval → Validation

Validity PositiveInterval

Instance details

Methods

validate ∷ PositiveInterval → Validation

Validity PositiveUnitInterval

Instance details

Methods

validate ∷ PositiveUnitInterval → Validation

Validity UnitInterval

Instance details

Methods

validate ∷ UnitInterval → Validation

(Era era, EncCBOR (f era), Arbitrary (f era)) ⇒ Arbitrary (Sized (f era))

Instance details

Methods

arbitrary ∷ Gen (Sized (f era))

shrink ∷ Sized (f era) → [Sized (f era)]

Crypto c ⇒ Arbitrary (Addr c)

Instance details

Methods

arbitrary ∷ Gen (Addr c)

shrink ∷ Addr c → [Addr c]

Arbitrary (BootstrapAddress c)

Instance details

Methods

arbitrary ∷ Gen (BootstrapAddress c)

shrink ∷ BootstrapAddress c → [BootstrapAddress c]

Crypto c ⇒ Arbitrary (CompactAddr c)

Instance details

Methods

arbitrary ∷ Gen (CompactAddr c)

shrink ∷ CompactAddr c → [CompactAddr c]

Crypto c ⇒ Arbitrary (RewardAcnt c)

Instance details

Methods

arbitrary ∷ Gen (RewardAcnt c)

shrink ∷ RewardAcnt c → [RewardAcnt c]

Crypto c ⇒ Arbitrary (Withdrawals c)

Instance details

Methods

arbitrary ∷ Gen (Withdrawals c)

shrink ∷ Withdrawals c → [Withdrawals c]

Crypto c ⇒ Arbitrary (AuxiliaryDataHash c)

Instance details

Methods

arbitrary ∷ Gen (AuxiliaryDataHash c)

shrink ∷ AuxiliaryDataHash c → [AuxiliaryDataHash c]

Crypto c ⇒ Arbitrary (BlocksMade c)

Instance details

Methods

arbitrary ∷ Gen (BlocksMade c)

shrink ∷ BlocksMade c → [BlocksMade c]

Era era ⇒ Arbitrary (CertState era)

Instance details

Methods

arbitrary ∷ Gen (CertState era)

shrink ∷ CertState era → [CertState era]

Era era ⇒ Arbitrary (DState era)

Instance details

Methods

arbitrary ∷ Gen (DState era)

shrink ∷ DState era → [DState era]

Crypto c ⇒ Arbitrary (FutureGenDeleg c)

Instance details

Methods

arbitrary ∷ Gen (FutureGenDeleg c)

shrink ∷ FutureGenDeleg c → [FutureGenDeleg c]

Crypto c ⇒ Arbitrary (InstantaneousRewards c)

Instance details

Methods

arbitrary ∷ Gen (InstantaneousRewards c)

shrink ∷ InstantaneousRewards c → [InstantaneousRewards c]

Era era ⇒ Arbitrary (PState era)

Instance details

Methods

arbitrary ∷ Gen (PState era)

shrink ∷ PState era → [PState era]

Era era ⇒ Arbitrary (VState era)

Instance details

Methods

arbitrary ∷ Gen (VState era)

shrink ∷ VState era → [VState era]

Arbitrary (CompactForm Coin)

Instance details

Methods

arbitrary ∷ Gen (CompactForm Coin)

shrink ∷ CompactForm Coin → [CompactForm Coin]

(Era era, Arbitrary (PParamsHKD Identity era)) ⇒ Arbitrary (PParams era)

Instance details

Methods

arbitrary ∷ Gen (PParams era)

shrink ∷ PParams era → [PParams era]

(Era era, Arbitrary (PParamsHKD StrictMaybe era)) ⇒ Arbitrary (PParamsUpdate era)

Instance details

Methods

arbitrary ∷ Gen (PParamsUpdate era)

shrink ∷ PParamsUpdate era → [PParamsUpdate era]

Crypto c ⇒ Arbitrary (StakeReference c)

Instance details

Methods

arbitrary ∷ Gen (StakeReference c)

shrink ∷ StakeReference c → [StakeReference c]

Crypto c ⇒ Arbitrary (Stake c)
In the system, Stake never contains more than the sum of all Ada (which is constant). This makes it safe to store individual Coins (in CompactForm) as Word64. But we must be careful that we never generate Stake where the sum of all the coins exceeds (maxBound :: Word64) There will never be a real Stake in the system with that many Ada, because total Ada is constant. So using a restricted Arbitrary Generator is OK.

Instance details

Methods

arbitrary ∷ Gen (Stake c)

shrink ∷ Stake c → [Stake c]

Crypto c ⇒ Arbitrary (ScriptHash c)

Instance details

Methods

arbitrary ∷ Gen (ScriptHash c)

shrink ∷ ScriptHash c → [ScriptHash c]

Crypto c ⇒ Arbitrary (GenDelegPair c)

Instance details

Methods

arbitrary ∷ Gen (GenDelegPair c)

shrink ∷ GenDelegPair c → [GenDelegPair c]

Crypto c ⇒ Arbitrary (GenDelegs c)

Instance details

Methods

arbitrary ∷ Gen (GenDelegs c)

shrink ∷ GenDelegs c → [GenDelegs c]

Crypto c ⇒ Arbitrary (BootstrapWitness c)

Instance details

Methods

arbitrary ∷ Gen (BootstrapWitness c)

shrink ∷ BootstrapWitness c → [BootstrapWitness c]

Crypto c ⇒ Arbitrary (IndividualPoolStake c)

Instance details

Methods

arbitrary ∷ Gen (IndividualPoolStake c)

shrink ∷ IndividualPoolStake c → [IndividualPoolStake c]

Crypto c ⇒ Arbitrary (PoolDistr c)

Instance details

Methods

arbitrary ∷ Gen (PoolDistr c)

shrink ∷ PoolDistr c → [PoolDistr c]

Crypto c ⇒ Arbitrary (PoolParams c)

Instance details

Methods

arbitrary ∷ Gen (PoolParams c)

shrink ∷ PoolParams c → [PoolParams c]

Crypto c ⇒ Arbitrary (Reward c)

Instance details

Methods

arbitrary ∷ Gen (Reward c)

shrink ∷ Reward c → [Reward c]

Crypto c ⇒ Arbitrary (TxId c)

Instance details

Methods

arbitrary ∷ Gen (TxId c)

shrink ∷ TxId c → [TxId c]

Crypto c ⇒ Arbitrary (TxIn c)

Instance details

Methods

arbitrary ∷ Gen (TxIn c)

shrink ∷ TxIn c → [TxIn c]

Crypto c ⇒ Arbitrary (Trip c)

Instance details

Methods

arbitrary ∷ Gen (Trip c)

shrink ∷ Trip c → [Trip c]

Crypto c ⇒ Arbitrary (UMap c)

Instance details

Methods

arbitrary ∷ Gen (UMap c)

shrink ∷ UMap c → [UMap c]

(EraTxOut era, Arbitrary (TxOut era)) ⇒ Arbitrary (UTxO era)

Instance details

Methods

arbitrary ∷ Gen (UTxO era)

shrink ∷ UTxO era → [UTxO era]

Crypto c ⇒ Arbitrary (Credential r c)

Instance details

Methods

arbitrary ∷ Gen (Credential r c)

shrink ∷ Credential r c → [Credential r c]

Crypto c ⇒ Arbitrary (KeyHash a c)

Instance details

Methods

arbitrary ∷ Gen (KeyHash a c)

shrink ∷ KeyHash a c → [KeyHash a c]

DSIGNAlgorithm (DSIGN c) ⇒ Arbitrary (VKey kd c)

Instance details

Methods

arbitrary ∷ Gen (VKey kd c)

shrink ∷ VKey kd c → [VKey kd c]

(Typeable kr, Crypto c) ⇒ Arbitrary (WitVKey kr c)

Instance details

Methods

arbitrary ∷ Gen (WitVKey kr c)

shrink ∷ WitVKey kr c → [WitVKey kr c]

Crypto c ⇒ Arbitrary (SafeHash c i)

Instance details

Methods

arbitrary ∷ Gen (SafeHash c i)

shrink ∷ SafeHash c i → [SafeHash c i]

DSIGNAlgorithm (DSIGN c) ⇒ Arbitrary (KeyPair kd c)

Instance details

Methods

arbitrary ∷ Gen (KeyPair kd c)

shrink ∷ KeyPair kd c → [KeyPair kd c]