Test.Cardano.Ledger.Core.KeyPair

Documentation

mkAddr ∷ Crypto c ⇒ (KeyPair 'Payment c, KeyPair 'Staking c) → Addr c

mkScriptAddr ∷ Crypto c ⇒ ScriptHash c → KeyPair 'Staking c → Addr c

mkCred ∷ Crypto c ⇒ KeyPair kr c → Credential kr c

data KeyPair (kd ∷ KeyRole) c

Constructors

KeyPair

Fields

vKey ∷ !(VKey kd c)

sKey ∷ !(SignKeyDSIGN (DSIGN c))

type KeyPairs c = [(KeyPair 'Payment c, KeyPair 'Staking c)]

Representation of a list of pairs of key pairs, e.g., pay and stake keys

mkWitnessVKey ∷ ∀ c kr. (Crypto c, DSignable c (Hash (HASH c) EraIndependentTxBody)) ⇒ SafeHash c EraIndependentTxBody → KeyPair kr c → WitVKey 'Witness c

Create a witness for transaction

mkWitnessesVKey ∷ ∀ c kr. (Crypto c, DSignable c (Hash (HASH c) EraIndependentTxBody)) ⇒ SafeHash c EraIndependentTxBody → [KeyPair kr c] → Set (WitVKey 'Witness c)

Create witnesses for transaction

makeWitnessesFromScriptKeys ∷ (Crypto c, DSignable c (Hash c EraIndependentTxBody)) ⇒ SafeHash c EraIndependentTxBody → Map (KeyHash kr c) (KeyPair kr c) → Set (KeyHash kr c) → Set (WitVKey 'Witness c)

From a list of key pairs and a set of key hashes required for a multi-sig scripts, return the set of required keys.

mkKeyHashWitFunPair ∷ ∀ kr. Gen (KeyHash kr StandardCrypto, SafeHash StandardCrypto EraIndependentTxBody → WitVKey 'Witness StandardCrypto)

When wrting a test which needs a KeyHash, and you will also later need a witness for that function, use keyHashWitFunPair. Since one cannot make a witness until one has SafeHash of the TxBody that the KeyHash is embedded. The second part of the pair is a (SafeHash to WitVKey) function. Use it something like this do (key,witfun) <- keyHashWitFunPair txbody <- ... key ... let safehash = hashAnnotated txbody tx = ... txbody ... (witfun safehash) ...

mkVKeyRewardAccount ∷ Crypto c ⇒ Network → KeyPair 'Staking c → RewardAccount c

mkKeyPair ∷ Crypto c ⇒ Int → KeyPair r c

mkKeyPairWithSeed ∷ ∀ r c. Crypto c ⇒ ByteString → KeyPair r c

mkKeyHash ∷ Crypto c ⇒ Int → KeyHash kd c

data ByronKeyPair

Constructors

ByronKeyPair

Fields

bkpVerificationKey ∷ !VerificationKey

bkpSigningKey ∷ !SigningKey

mkBootKeyPairWithSeed ∷ ByteString → ByronKeyPair

genByronVKeyAddr ∷ Gen (VerificationKey, Address)

genByronAddrFromVKey ∷ VerificationKey → Gen Address

Deprecations

mkVKeyRwdAcnt ∷ Crypto c ⇒ Network → KeyPair 'Staking c → RewardAccount c

Deprecated: Use mkVKeyRewardAccount instead