Test.Cardano.Ledger.Core.KeyPair

Documentation

mkAddr ∷ Crypto c ⇒ (KeyPair 'Payment c, KeyPair 'Staking c) → Addr c

mkCred ∷ Crypto c ⇒ KeyPair kr c → Credential kr c

data KeyPair (kd ∷ KeyRole) c

Constructors

KeyPair

Fields

vKey ∷ !(VKey kd c)

sKey ∷ !(SignKeyDSIGN (DSIGN c))

type KeyPairs c = [(KeyPair 'Payment c, KeyPair 'Staking c)]

Representation of a list of pairs of key pairs, e.g., pay and stake keys

mkWitnessVKey ∷ ∀ c kr. (Crypto c, DSignable c (Hash (HASH c) EraIndependentTxBody)) ⇒ SafeHash c EraIndependentTxBody → KeyPair kr c → WitVKey 'Witness c

Create a witness for transaction

mkWitnessesVKey ∷ ∀ c kr. (Crypto c, DSignable c (Hash (HASH c) EraIndependentTxBody)) ⇒ SafeHash c EraIndependentTxBody → [KeyPair kr c] → Set (WitVKey 'Witness c)

Create witnesses for transaction

makeWitnessesFromScriptKeys ∷ (Crypto c, DSignable c (Hash c EraIndependentTxBody)) ⇒ SafeHash c EraIndependentTxBody → Map (KeyHash kr c) (KeyPair kr c) → Set (KeyHash kr c) → Set (WitVKey 'Witness c)

From a list of key pairs and a set of key hashes required for a multi-sig scripts, return the set of required keys.

mkVKeyRwdAcnt ∷ Crypto c ⇒ Network → KeyPair 'Staking c → RewardAcnt c